Example data set from Gorsuch (1997) for an example factor extension.

Gorsuch (1997) suggests an alternative to the classic Dwyer (1937) factor extension technique. This data set is taken from that article. Useful for comparing link{fa.extension} with and without the correct=TRUE option.

data(Gorsuch)

Details

Gorsuc (1997) suggested an alternative model for factor extension. His method is appropriate for the case of repeated variables. This is handled in link{fa.extension} with correct=FALSE

Source

Richard L. Gorsuch (1997) New Procedure for Extension Analysis in Exploratory Factor Analysis. Educational and Psychological Measurement, 57, 725-740.

References

Dwyer, Paul S. (1937), The determination of the factor loadings of a given test from the known factor loadings of other tests. Psychometrika, 3, 173-178

Examples

data(Gorsuch)

Ro <- Gorsuch[1:6,1:6]
Roe <- Gorsuch[1:6,7:10]
fo <- fa(Ro,2,rotate="none")
fa.extension(Roe,fo,correct=FALSE)

#> 
#> Call: fa.extension(Roe = Roe, fo = fo, correct = FALSE)
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>            MR1  MR2   h2   u2
#> info2     0.72 0.34 0.63 0.37
#> tension2 -0.40 0.54 0.45 0.55
#> v123      0.82 0.41 0.84 0.16
#> v564     -0.45 0.71 0.71 0.29
#> 
#>                        MR1  MR2
#> SS loadings           1.55 1.08
#> Proportion Var        0.39 0.27
#> Cumulative Var        0.39 0.66
#> Proportion Explained  0.59 0.41
#> Cumulative Proportion 0.59 1.00