Perform bifactor, promax or targeted rotations and return the inter factor angles.

The bifactor rotation implements the rotation introduced by Jennrich and Bentler (2011) by calling GPForth in the GPArotation package. promax is an oblique rotation function introduced by Hendrickson and White (1964) and implemented in the promax function in the stats package. Unfortunately, promax does not report the inter factor correlations. Promax does. TargetQ does a target rotation with elements that can be missing (NA), or numeric (e.g., 0, 1). It uses the GPArotation package. target.rot does general target rotations to an arbitrary target matrix. The default target rotation is for an independent cluster solution. equamax facilitates the call to GPArotation to do an equamax rotation. Equamax, although available as a specific option within GPArotation is easier to call by name if using equamax. The varimin rotation suggested by Ertl (2013) is implemented by appropriate calls to GPArotation.

bifactor(L, Tmat=diag(ncol(L)), normalize=FALSE, eps=1e-5, maxit=1000)
biquartimin(L, Tmat=diag(ncol(L)), normalize=FALSE, eps=1e-5, maxit=1000)
TargetQ(L, Tmat=diag(ncol(L)), normalize=FALSE, eps=1e-5, maxit=1000,Target=NULL)
Promax(x,m=4, normalize=FALSE, pro.m = 4)
target.rot(x,keys=NULL)
varimin(L, Tmat = diag(ncol(L)), normalize = FALSE, eps = 1e-05, maxit = 1000)
vgQ.bimin(L)   #called by bifactor
vgQ.targetQ(L,Target=NULL)  #called by TargetQ
vgQ.varimin(L)  #called by varimin
equamax(L, Tmat=diag(ncol(L)), eps=1e-5, maxit=1000)

Arguments

x
A loadings matrix

m
the power to which to raise the varimax loadings (for Promax)

pro.m
the power to which to raise the various loadings in Promax.

keys
An arbitrary target matrix, can be composed of any weights, but probably -1,0, 1 weights. If missing, the target is the independent cluster structure determined by assigning every item to it's highest loaded factor.

L
A loadings matrix

Target
A matrix of values (mainly 0s, some 1s, some NAs) to which the matrix is transformed.

Tmat
An initial rotation matrix

normalize
parameter passed to optimization routine (GPForth in the GPArotation package and Promax)

eps
parameter passed to optimization routine (GPForth in the GPArotation package)

maxit
parameter passed to optimization routine (GPForth in the GPArotation package)

Details

The two most useful of these six functions is probably biquartimin which implements the oblique bifactor rotation introduced by Jennrich and Bentler (2011). The second is TargetQ which allows for missing NA values in the target. Next best is the orthogonal case, bifactor. None of these seem to be implemented in GPArotation (yet).

The difference between biquartimin and bifactor is just that the latter is the orthogonal case which is documented in Jennrich and Bentler (2011). It seems as if these two functions are sensitive to the starting values and random restarts (modifying T) might be called for.

bifactor output for the 24 cognitive variable of Holzinger matches that of Jennrich and Bentler as does output for the Chen et al. problem when fm="mle" is used and the Jennrich and Bentler solution is rescaled from covariances to correlations.

Promax is a very direct adaptation of the stats::promax function. The addition is that it will return the interfactor correlations as well as the loadings and rotation matrix.

varimin implements the varimin criterion proposed by Suitbert Ertl (2013). Rather than maximize the varimax criterion, it minimizes it. For a discussion of the benefits of this procedure, consult Ertel (2013).

In addition, these functions will take output from either the factanal, fa or earlier (factor.pa, factor.minres or principal) functions and select just the loadings matrix for analysis.

equamax is just a call to GPArotation's cFT function (for the Crawford Ferguson family of rotations.

TargetQ implements Michael Browne's algorithm and allows specification of NA values. The Target input is a list (see examples). It is interesting to note how powerful specifying what a factor isn't works in defining a factor. That is, by specifying the pattern of 0s and letting most other elements be NA, the factor structure is still clearly defined.

The target.rot function is an adaptation of a function of Michael Browne's to do rotations to arbitrary target matrices. Suggested by Pat Shrout.

The default for target.rot is to rotate to an independent cluster structure (every items is assigned to a group with its highest loading.)

target.rot will not handle targets that have linear dependencies (e.g., a pure bifactor model where there is a g loading and a group factor for all variables).

Value

loadings

Oblique factor loadings

rotmat

The rotation matrix applied to the original loadings to produce the promax solution or the targeted matrix

Phi

The interfactor correlation matrix

References

Ertel, S. (2013). Factor analysis: healing an ailing model. Universitatsverlag Gottingen.

Hendrickson, A. E. and White, P. O, 1964, British Journal of Statistical Psychology, 17, 65-70.

Jennrich, Robert and Bentler, Peter (2011) Exploratory Bi-Factor Analysis. Psychometrika, 1-13

Note

A direct adaptation of the stats:promax function following suggestions to the R-help list by Ulrich Keller and John Fox. Further modified to do targeted rotation similar to a function of Michael Browne.

varimin is a direct application of the GPArotation GPForth function modified to do varimin.

See also

promax, fa, or principal for examples of data analysis and Holzinger or Bechtoldt for examples of bifactor data. factor.rotate for 'hand rotation'.

Examples

jen <- sim.hierarchical()
f3 <- fa(jen,3,rotate="varimax")
f3   #not a very clean solution

#> Factor Analysis using method =  minres
#> Call: fa(r = jen, nfactors = 3, rotate = "varimax")
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>     MR1  MR3  MR2   h2   u2 com
#> V1 0.70 0.30 0.26 0.64 0.36 1.6
#> V2 0.61 0.26 0.22 0.49 0.51 1.6
#> V3 0.52 0.22 0.19 0.36 0.64 1.6
#> V4 0.24 0.63 0.18 0.49 0.51 1.5
#> V5 0.21 0.54 0.16 0.36 0.64 1.5
#> V6 0.17 0.45 0.13 0.25 0.75 1.5
#> V7 0.17 0.15 0.56 0.36 0.64 1.3
#> V8 0.14 0.12 0.46 0.25 0.75 1.3
#> V9 0.11 0.10 0.37 0.16 0.84 1.3
#> 
#>                        MR1  MR3  MR2
#> SS loadings           1.32 1.15 0.89
#> Proportion Var        0.15 0.13 0.10
#> Cumulative Var        0.15 0.27 0.37
#> Proportion Explained  0.39 0.34 0.27
#> Cumulative Proportion 0.39 0.73 1.00
#> 
#> Mean item complexity =  1.5
#> Test of the hypothesis that 3 factors are sufficient.
#> 
#> The degrees of freedom for the null model are  36  and the objective function was  1.71
#> The degrees of freedom for the model are 12  and the objective function was  0 
#> 
#> The root mean square of the residuals (RMSR) is  0 
#> The df corrected root mean square of the residuals is  0 
#> 
#> Fit based upon off diagonal values = 1
#> Measures of factor score adequacy             
#>                                                 MR1  MR3   MR2
#> Correlation of scores with factors             0.78 0.73  0.67
#> Multiple R square of scores with factors       0.61 0.53  0.45
#> Minimum correlation of possible factor scores  0.22 0.07 -0.10

Promax(f3)

#> 
#> Call: NULL
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>     MR1  MR3  MR2   h2   u2
#> V1 0.76 0.04 0.03 0.64 0.36
#> V2 0.66 0.03 0.02 0.49 0.51
#> V3 0.57 0.03 0.02 0.36 0.64
#> V4 0.01 0.69 0.01 0.49 0.51
#> V5 0.01 0.59 0.01 0.36 0.64
#> V6 0.01 0.49 0.01 0.25 0.75
#> V7 0.00 0.00 0.60 0.36 0.64
#> V8 0.00 0.00 0.50 0.25 0.75
#> V9 0.00 0.00 0.40 0.16 0.84
#> 
#>                        MR1  MR3  MR2
#> SS loadings           1.43 1.13 0.80
#> Proportion Var        0.16 0.13 0.09
#> Cumulative Var        0.16 0.28 0.37
#> Proportion Explained  0.42 0.34 0.24
#> Cumulative Proportion 0.42 0.76 1.00
#>      MR1  MR3  MR2
#> MR1 1.00 0.68 0.60
#> MR3 0.68 1.00 0.55
#> MR2 0.60 0.55 1.00

target.rot(f3)

#> 
#> Call: NULL
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>    MR1 MR3 MR2   h2   u2
#> V1 0.8 0.0 0.0 0.64 0.36
#> V2 0.7 0.0 0.0 0.49 0.51
#> V3 0.6 0.0 0.0 0.36 0.64
#> V4 0.0 0.7 0.0 0.49 0.51
#> V5 0.0 0.6 0.0 0.36 0.64
#> V6 0.0 0.5 0.0 0.25 0.75
#> V7 0.0 0.0 0.6 0.36 0.64
#> V8 0.0 0.0 0.5 0.25 0.75
#> V9 0.0 0.0 0.4 0.16 0.84
#> 
#>                        MR1  MR3  MR2
#> SS loadings           1.49 1.10 0.77
#> Proportion Var        0.17 0.12 0.09
#> Cumulative Var        0.17 0.29 0.37
#> Proportion Explained  0.44 0.33 0.23
#> Cumulative Proportion 0.44 0.77 1.00
#>      MR1  MR3  MR2
#> MR1 1.00 0.72 0.63
#> MR3 0.72 1.00 0.56
#> MR2 0.63 0.56 1.00

m3 <- fa(jen,nfactors=3)
Promax(m3)  #example of taking the output from factanal

#> 
#> Call: NULL
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>    MR1 MR3 MR2   h2   u2
#> V1 0.8 0.0 0.0 0.64 0.36
#> V2 0.7 0.0 0.0 0.49 0.51
#> V3 0.6 0.0 0.0 0.36 0.64
#> V4 0.0 0.7 0.0 0.49 0.51
#> V5 0.0 0.6 0.0 0.36 0.64
#> V6 0.0 0.5 0.0 0.25 0.75
#> V7 0.0 0.0 0.6 0.36 0.64
#> V8 0.0 0.0 0.5 0.25 0.75
#> V9 0.0 0.0 0.4 0.16 0.84
#> 
#>                        MR1  MR3  MR2
#> SS loadings           1.49 1.10 0.77
#> Proportion Var        0.17 0.12 0.09
#> Cumulative Var        0.17 0.29 0.37
#> Proportion Explained  0.44 0.33 0.23
#> Cumulative Proportion 0.44 0.77 1.00
#>     MR1 MR3 MR2
#> MR1   1   0   0
#> MR3   0   1   0
#> MR2   0   0   1

#compare this rotation with the solution from a targeted rotation aimed for 
#an independent cluster solution
target.rot(m3)

#> 
#> Call: NULL
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>    MR1 MR3 MR2   h2   u2
#> V1 0.8 0.0 0.0 0.64 0.36
#> V2 0.7 0.0 0.0 0.49 0.51
#> V3 0.6 0.0 0.0 0.36 0.64
#> V4 0.0 0.7 0.0 0.49 0.51
#> V5 0.0 0.6 0.0 0.36 0.64
#> V6 0.0 0.5 0.0 0.25 0.75
#> V7 0.0 0.0 0.6 0.36 0.64
#> V8 0.0 0.0 0.5 0.25 0.75
#> V9 0.0 0.0 0.4 0.16 0.84
#> 
#>                        MR1  MR3  MR2
#> SS loadings           1.49 1.10 0.77
#> Proportion Var        0.17 0.12 0.09
#> Cumulative Var        0.17 0.29 0.37
#> Proportion Explained  0.44 0.33 0.23
#> Cumulative Proportion 0.44 0.77 1.00
#>     MR1 MR3 MR2
#> MR1   1   0   0
#> MR3   0   1   0
#> MR2   0   0   1

#now try a bifactor solution
fb <-fa(jen,3,rotate="bifactor")
fq <- fa(jen,3,rotate="biquartimin")
#Suitbert Ertel has suggested varimin
fm <-  fa(jen,3,rotate="varimin") #the Ertel varimin
fn <- fa(jen,3,rotate="none")  #just the unrotated factors
#compare them
factor.congruence(list(f3,fb,fq,fm,fn))

#>       MR1   MR3  MR2  MR1   MR3   MR2   MR1   MR3   MR2  MR1   MR2   MR3  MR1
#> MR1  1.00  0.71 0.66 0.87  0.93 -0.02  0.92  0.93 -0.01 0.87  0.78  0.63 0.92
#> MR3  0.71  1.00 0.61 0.92  0.42 -0.36  0.89  0.42 -0.34 0.88  0.75 -0.08 0.89
#> MR2  0.66  0.61 1.00 0.84  0.41  0.52  0.82  0.41  0.54 0.88  0.16  0.40 0.82
#> MR1  0.87  0.92 0.84 1.00  0.61  0.00  0.99  0.61  0.01 1.00  0.65  0.27 0.99
#> MR3  0.93  0.42 0.41 0.61  1.00  0.00  0.71  1.00  0.00 0.62  0.73  0.80 0.71
#> MR2 -0.02 -0.36 0.52 0.00  0.00  1.00 -0.01  0.00  1.00 0.09 -0.63  0.53 0.00
#> MR1  0.92  0.89 0.82 0.99  0.71 -0.01  1.00  0.71  0.01 0.99  0.70  0.37 1.00
#> MR3  0.93  0.42 0.41 0.61  1.00  0.00  0.71  1.00  0.00 0.62  0.73  0.80 0.71
#> MR2 -0.01 -0.34 0.54 0.01  0.00  1.00  0.01  0.00  1.00 0.11 -0.63  0.52 0.01
#> MR1  0.87  0.88 0.88 1.00  0.62  0.09  0.99  0.62  0.11 1.00  0.59  0.33 0.99
#> MR2  0.78  0.75 0.16 0.65  0.73 -0.63  0.70  0.73 -0.63 0.59  1.00  0.18 0.70
#> MR3  0.63 -0.08 0.40 0.27  0.80  0.53  0.37  0.80  0.52 0.33  0.18  1.00 0.37
#> MR1  0.92  0.89 0.82 0.99  0.71  0.00  1.00  0.71  0.01 0.99  0.70  0.37 1.00
#> MR2 -0.01 -0.36 0.52 0.00  0.02  1.00  0.00  0.02  1.00 0.09 -0.62  0.55 0.00
#> MR3 -0.38  0.28 0.25 0.13 -0.70  0.02  0.00 -0.70  0.03 0.12 -0.35 -0.75 0.00
#>       MR2   MR3
#> MR1 -0.01 -0.38
#> MR3 -0.36  0.28
#> MR2  0.52  0.25
#> MR1  0.00  0.13
#> MR3  0.02 -0.70
#> MR2  1.00  0.02
#> MR1  0.00  0.00
#> MR3  0.02 -0.70
#> MR2  1.00  0.03
#> MR1  0.09  0.12
#> MR2 -0.62 -0.35
#> MR3  0.55 -0.75
#> MR1  0.00  0.00
#> MR2  1.00  0.00
#> MR3  0.00  1.00

# compare an oblimin with a target rotation using the Browne algorithm
 #note that we are changing the factor #order (this is for demonstration only)
 Targ <- make.keys(9,list(f1=1:3,f2=7:9,f3=4:6))
 Targ <- scrub(Targ,isvalue=1)  #fix the 0s, allow the NAs to be estimated
 Targ <- list(Targ)  #input must be a list
#show the target
 Targ

#> [[1]]
#>       f1 f2 f3
#>  [1,] NA  0  0
#>  [2,] NA  0  0
#>  [3,] NA  0  0
#>  [4,]  0  0 NA
#>  [5,]  0  0 NA
#>  [6,]  0  0 NA
#>  [7,]  0 NA  0
#>  [8,]  0 NA  0
#>  [9,]  0 NA  0
#> 

 fa(Thurstone,3,rotate="TargetQ",Target=Targ)  #targeted rotation

#> Factor Analysis using method =  minres
#> Call: fa(r = Thurstone, nfactors = 3, rotate = "TargetQ", Target = Targ)
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>                     MR1   MR3   MR2   h2   u2 com
#> Sentences          0.86 -0.01  0.08 0.82 0.18 1.0
#> Vocabulary         0.85  0.10  0.00 0.84 0.16 1.0
#> Sent.Completion    0.80  0.06  0.04 0.74 0.26 1.0
#> First.Letters     -0.01  0.86  0.00 0.73 0.27 1.0
#> Four.Letter.Words -0.04  0.75  0.11 0.63 0.37 1.0
#> Suffixes           0.17  0.64 -0.08 0.50 0.50 1.2
#> Letter.Series     -0.05 -0.07  0.92 0.73 0.27 1.0
#> Pedigrees          0.32 -0.07  0.52 0.51 0.49 1.7
#> Letter.Group      -0.12  0.17  0.68 0.52 0.48 1.2
#> 
#>                        MR1  MR3  MR2
#> SS loadings           2.41 1.88 1.71
#> Proportion Var        0.27 0.21 0.19
#> Cumulative Var        0.27 0.48 0.67
#> Proportion Explained  0.40 0.31 0.28
#> Cumulative Proportion 0.40 0.72 1.00
#> 
#>  With factor correlations of 
#>      MR1  MR3  MR2
#> MR1 1.00 0.57 0.58
#> MR3 0.57 1.00 0.59
#> MR2 0.58 0.59 1.00
#> 
#> Mean item complexity =  1.1
#> Test of the hypothesis that 3 factors are sufficient.
#> 
#> The degrees of freedom for the null model are  36  and the objective function was  5.2
#> The degrees of freedom for the model are 12  and the objective function was  0.01 
#> 
#> The root mean square of the residuals (RMSR) is  0.01 
#> The df corrected root mean square of the residuals is  0.01 
#> 
#> Fit based upon off diagonal values = 1
#> Measures of factor score adequacy             
#>                                                 MR1  MR3  MR2
#> Correlation of scores with factors             0.96 0.92 0.92
#> Multiple R square of scores with factors       0.92 0.86 0.84
#> Minimum correlation of possible factor scores  0.84 0.71 0.69

#compare with oblimin
fa(Thurstone,3)

#> Factor Analysis using method =  minres
#> Call: fa(r = Thurstone, nfactors = 3)
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>                     MR1   MR2   MR3   h2   u2 com
#> Sentences          0.90 -0.03  0.04 0.82 0.18 1.0
#> Vocabulary         0.89  0.06 -0.03 0.84 0.16 1.0
#> Sent.Completion    0.84  0.03  0.00 0.74 0.26 1.0
#> First.Letters      0.00  0.85  0.00 0.73 0.27 1.0
#> Four.Letter.Words -0.02  0.75  0.10 0.63 0.37 1.0
#> Suffixes           0.18  0.63 -0.08 0.50 0.50 1.2
#> Letter.Series      0.03 -0.01  0.84 0.73 0.27 1.0
#> Pedigrees          0.38 -0.05  0.46 0.51 0.49 2.0
#> Letter.Group      -0.06  0.21  0.63 0.52 0.48 1.2
#> 
#>                        MR1  MR2  MR3
#> SS loadings           2.65 1.87 1.49
#> Proportion Var        0.29 0.21 0.17
#> Cumulative Var        0.29 0.50 0.67
#> Proportion Explained  0.44 0.31 0.25
#> Cumulative Proportion 0.44 0.75 1.00
#> 
#>  With factor correlations of 
#>      MR1  MR2  MR3
#> MR1 1.00 0.59 0.53
#> MR2 0.59 1.00 0.52
#> MR3 0.53 0.52 1.00
#> 
#> Mean item complexity =  1.2
#> Test of the hypothesis that 3 factors are sufficient.
#> 
#> The degrees of freedom for the null model are  36  and the objective function was  5.2
#> The degrees of freedom for the model are 12  and the objective function was  0.01 
#> 
#> The root mean square of the residuals (RMSR) is  0.01 
#> The df corrected root mean square of the residuals is  0.01 
#> 
#> Fit based upon off diagonal values = 1
#> Measures of factor score adequacy             
#>                                                 MR1  MR2  MR3
#> Correlation of scores with factors             0.96 0.92 0.90
#> Multiple R square of scores with factors       0.93 0.85 0.82
#> Minimum correlation of possible factor scores  0.86 0.71 0.63