RMSEA {psych}	R Documentation

Root Mean Squared Error of Approximation from chisq, df, and n

Description

Find the RMSEA from model chi square, degrees of freedom and number of observations. Show confidence intervals.

Usage

RMSEA(chisq, dof, n.obs, alpha = 0.1)

Arguments

chisq
The Chi square statistic from an analysis

dof
Degrees of freedom of the model

n.obs
Number of observations

alpha
alpha level for confidence intervals

Details

RMSEA is just a chisquare adjusted by its degrees of freedom and the sample size. √{chisq/(dof * (nobs)) -1/(n.obs -1)}. It is given in most of the appropriate functions (e.g.,fa) and given here for completeness.

Value

RMSEA
the estimated value

RMSEA-L
the lower bound

RMSEA-U
the upper bound

Author(s)

William Revelle

References

Steiger, J. H.,and Lind, J. C. (1980). Statistically based tests for the number of common factors. Paper presented at the Annual Meeting of the Psychometric Society, Iowa City, IA.

See Also

fa, omega , esem

[Package psych version 2.6.1 Index]