VSS.parallel {psych}	R Documentation

Compare real and random VSS solutions

Description

Another useful test for the number of factors is when the eigen values of a random matrix are greater than the eigen values of a a real matrix. Here we show VSS solutions to random data. A better test is probably fa.parallel.

Usage

VSS.parallel(ncases, nvariables,scree=FALSE,rotate="none")

Arguments

ncases
Number of simulated cases

nvariables
number of simulated variables

scree
Show a scree plot for random data – see omega

rotate
rotate="none" or rotate="varimax"

Value

VSS like output to be plotted by VSS.plot

Author(s)

William Revelle

References

Very Simple Structure (VSS)

See Also

fa.parallel, VSS.plot, ICLUST, omega

Examples

#VSS.plot(VSS.parallel(200,24))

[Package psych version 2.6.1 Index]