VSS.scree {psych}	R Documentation

Plot the successive eigen values for a scree test

Description

Cattell's scree test is one of most simple ways of testing the number of components or factors in a correlation matrix. Here we plot the eigen values of a correlation matrix as well as the eigen values of a factor analysis.

Usage

scree(rx,factors=TRUE,pc=TRUE,main="Scree plot",hline=NULL,add=FALSE,sqrt=FALSE) 
VSS.scree(rx, main = "scree plot",sqrt=FALSE)

Arguments

rx
a correlation matrix or a data matrix. If data, then correlations are found using pairwise deletions.

factors
If true, draw the scree for factors

pc
If true, draw the scree for components

hline
if null, draw a horizontal line at 1, otherwise draw it at hline (make negative to not draw it)

main
Title

add
Should multiple plots be drawn?

sqrt
If TRUE, take the sqrt of the eigen value before plotting

Details

Among the many ways to choose the optimal number of factors is the scree test. A better function to show the scree as well as compare it to randomly parallel solutions is found found in fa.parallel

Following a suggestion from Marco Del Giudice, I added the sqrt option for version 2.2.12. -

Author(s)

William Revelle

References

https://personality-project.org/r/vss.html

See Also

fa.parallel VSS.plot, ICLUST, omega

Examples

scree(attitude)
#VSS.scree(cor(attitude)

[Package psych version 2.6.1 Index]