Find two estimates of reliability: Cronbach's alpha and Guttman's Lambda 6.

Internal consistency measures of reliability range from \(\omega_h\) to \(\alpha\) to \(\omega_t\). This function reports two estimates: Cronbach's coefficient \(\alpha\) and Guttman's \(\lambda_6\). Also reported are item - whole correlations, \(\alpha\) if an item is omitted, and item means and standard deviations.

alpha(x, keys=NULL,cumulative=FALSE, title=NULL, max=10,na.rm = TRUE,
   check.keys=FALSE,n.iter=1,delete=TRUE,use="pairwise",warnings=TRUE,n.obs=NULL)

Arguments

x
A data.frame or matrix of data, or a covariance or correlation matrix

keys
If some items are to be reversed keyed, then either specify the direction of all items or just a vector of which items to reverse

title
Any text string to identify this run

cumulative
should means reflect the sum of items or the mean of the items. The default value is means.

max
the number of categories/item to consider if reporting category frequencies. Defaults to 10, passed to link{response.frequencies}

na.rm
The default is to remove missing values and find pairwise correlations

check.keys
if TRUE, then find the first principal component and reverse key items with negative loadings. Give a warning if this happens.

n.iter
Number of iterations if bootstrapped confidence intervals are desired

delete
Delete items with no variance and issue a warning

use
Options to pass to the cor function: "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs". The default is "pairwise"

warnings
By default print a warning and a message that items were reversed. Suppress the message if warnings = FALSE

n.obs
If using correlation matrices as input, by specify the number of observations, we can find confidence intervals

Details

Alpha is one of several estimates of the internal consistency reliability of a test.

Surprisingly, more than a century after Spearman (1904) introduced the concept of reliability to psychologists, there are still multiple approaches for measuring it. Although very popular, Cronbach's \(\alpha\) (1951) underestimates the reliability of a test and over estimates the first factor saturation. \(\alpha\) (Cronbach, 1951) is the same as Guttman's \(\lambda\)3 (Guttman, 1945) and may be found by $$ \lambda_3 = \frac{n}{n-1}\Bigl(1 - \frac{tr(\vec{V})_x}{V_x}\Bigr) = \frac{n}{n-1} \frac{V_x - tr(\vec{V}_x)}{V_x} = \alpha $$

Perhaps because it is so easy to calculate and is available in most commercial programs, alpha is without doubt the most frequently reported measure of internal consistency reliability. Alpha is the mean of all possible spit half reliabilities (corrected for test length). For a unifactorial test, it is a reasonable estimate of the first factor saturation, although if the test has any microstructure (i.e., if it is ``lumpy") coefficients \(\beta\) (Revelle, 1979; see ICLUST) and \(\omega_h\) (see omega) are more appropriate estimates of the general factor saturation. \(\omega_t\) (see omega) is a better estimate of the reliability of the total test.

Guttman's Lambda 6 (G6) considers the amount of variance in each item that can be accounted for the linear regression of all of the other items (the squared multiple correlation or smc), or more precisely, the variance of the errors, \(e_j^2\), and is $$ \lambda_6 = 1 - \frac{\sum e_j^2}{V_x} = 1 - \frac{\sum(1-r_{smc}^2)}{V_x} .$$

The squared multiple correlation is a lower bound for the item communality and as the number of items increases, becomes a better estimate.

G6 is also sensitive to lumpyness in the test and should not be taken as a measure of unifactorial structure. For lumpy tests, it will be greater than alpha. For tests with equal item loadings, alpha > G6, but if the loadings are unequal or if there is a general factor, G6 > alpha. alpha is a generalization of an earlier estimate of reliability for tests with dichotomous items developed by Kuder and Richardson, known as KR20, and a shortcut approximation, KR21. (See Revelle, in prep).

Alpha and G6 are both positive functions of the number of items in a test as well as the average intercorrelation of the items in the test. When calculated from the item variances and total test variance, as is done here, raw alpha is sensitive to differences in the item variances. Standardized alpha is based upon the correlations rather than the covariances.

A useful index of the quality of the test that is linear with the number of items and the average correlation is the Signal/Noise ratio where $$s/n = \frac{n \bar{r}}{1-n \bar{r}}$$ (Cronbach and Gleser, 1964; Revelle and Condon (in press)).

More complete reliability analyses of a single scale can be done using the omega function which finds \(\omega_h\) and \(\omega_t\) based upon a hierarchical factor analysis.

Alternative functions score.items and cluster.cor will also score multiple scales and report more useful statistics. ``Standardized" alpha is calculated from the inter-item correlations and will differ from raw alpha.

Four alternative item-whole correlations are reported, three are conventional, one unique. raw.r is the correlation of the item with the entire scale, not correcting for item overlap. std.r is the correlation of the item with the entire scale, if each item were standardized. r.drop is the correlation of the item with the scale composed of the remaining items. Although each of these are conventional statistics, they have the disadvantage that a) item overlap inflates the first and b) the scale is different for each item when an item is dropped. Thus, the fourth alternative, r.cor, corrects for the item overlap by subtracting the item variance but then replaces this with the best estimate of common variance, the smc. This is similar to a suggestion by Cureton (1966).

If some items are to be reversed keyed then they can be specified by either item name or by item location. (Look at the 3rd and 4th examples.) Automatic reversal can also be done, and this is based upon the sign of the loadings on the first principal component (Example 5). This requires the check.keys option to be TRUE. Previous versions defaulted to have check.keys=TRUE, but some users complained that this made it too easy to find alpha without realizing that some items had been reversed (even though a warning was issued!). Thus, I have set the default to be check.keys=FALSE with a warning that some items need to be reversed (if this is the case). To suppress these warnings, set warnings=FALSE.

Scores are based upon the simple averages (or totals) of the items scored. Reversed items are subtracted from the maximum + minimum item response for all the items.

When using raw data, standard errors for the raw alpha are calculated using equation 2 and 3 from Duhhachek and Iacobucci (2004). This is problematic because some simulations suggest these values are too small. It is probably better to use bootstrapped value

Bootstrapped resamples are found if n.iter > 1. These are returned as the boot object. They may be plotted or described.

Value

total

a list containing

raw_alpha

alpha based upon the covariances

std.alpha

The standarized alpha based upon the correlations

G6(smc)

Guttman's Lambda 6 reliability

average_r

The average interitem correlation

mean

For data matrices, the mean of the scale formed by summing the items

sd

For data matrices, the standard deviation of the total score

alpha.drop

A data frame with all of the above for the case of each item being removed one by one.

item.stats

A data frame including

n

number of complete cases for the item

raw.r

The correlation of each item with the total score, not corrected for item overlap.

std.r

The correlation of each item with the total score (not corrected for item overlap) if the items were all standardized

r.cor

Item whole correlation corrected for item overlap and scale reliability

r.drop

Item whole correlation for this item against the scale without this item

mean

for data matrices, the mean of each item

sd

For data matrices, the standard deviation of each item

response.freq

For data matrices, the frequency of each item response (if less than 20)

boot

a 6 column by n.iter matrix of boot strapped resampled values

Unidim

An index of unidimensionality

Fit

The fit of the off diagonal matrix

References

Cronbach, L.J. (1951) Coefficient alpha and the internal strucuture of tests. Psychometrika, 16, 297-334.

Cureton, E. (1966). Corrected item-test correlations. Psychometrika, 31(1):93-96.

Cronbach, L.J. and Gleser G.C. (1964)The signal/noise ratio in the comparison of reliability coefficients. Educational and Psychological Measurement, 24 (3) 467-480.

Duhachek, A. and Iacobucci, D. (2004). Alpha's standard error (ase): An accurate and precise confidence interval estimate. Journal of Applied Psychology, 89(5):792-808.

Guttman, L. (1945). A basis for analyzing test-retest reliability. Psychometrika, 10 (4), 255-282.

Revelle, W. (in preparation) An introduction to psychometric theory with applications in R. Springer. (Available online at http://personality-project.org/r/book).

Revelle, W. Hierarchical Cluster Analysis and the Internal Structure of Tests. Multivariate Behavioral Research, 1979, 14, 57-74.

Revelle, W. and Condon, D.C. Reliability. In Irwing, P., Booth, T. and Hughes, D. (Eds). the Wiley-Blackwell Handbook of Psychometric Testing (in press).

Revelle, W. and Zinbarg, R. E. (2009) Coefficients alpha, beta, omega and the glb: comments on Sijtsma. Psychometrika, 74 (1) 1145-154.

Note

By default, items that correlate negatively with the overall scale will be reverse coded. This option may be turned off by setting check.keys = FALSE. If items are reversed, then each item is subtracted from the minimum item response + maximum item response where min and max are taken over all items. Thus, if the items intentionally differ in range, the scores will be off by a constant. See scoreItems for a solution.

If the data have been preprocessed by the dplyr package, a strange error can occur. alpha expects either data.frames or matrix input. data.frames returned by dplyr have had three extra classes added to them which causes alpha to break. The solution is merely to change the class of the input to "data.frame".

Two experimental measures of Goodness of Fit are returned in the output: Unidim and Fit. They are not printed or displayed, but are available for analysis. The first is an index of how well the modeled average correlations actually reproduce the original correlation matrix. The second is how well the modeled correlations reproduce the off diagonal elements of the matrix. Both are indices of squared residuals compared to the squared original correlations. These two measures are under development and might well be modified or dropped in subsequent versions.

See also

omega, ICLUST, guttman, scoreItems, cluster.cor

Examples

set.seed(42) #keep the same starting values
#four congeneric measures
r4 <- sim.congeneric()
alpha(r4)

#> 
#> Reliability analysis   
#> Call: alpha(x = r4)
#> 
#>   raw_alpha std.alpha G6(smc) average_r S/N
#>       0.74      0.74    0.69      0.42 2.9
#> 
#>  Reliability if an item is dropped:
#>    raw_alpha std.alpha G6(smc) average_r S/N
#> V1      0.62      0.62    0.53      0.36 1.7
#> V2      0.66      0.66    0.57      0.39 1.9
#> V3      0.70      0.70    0.62      0.44 2.3
#> V4      0.74      0.74    0.66      0.49 2.8
#> 
#>  Item statistics 
#>       r r.cor r.drop
#> V1 0.81  0.74   0.64
#> V2 0.78  0.67   0.57
#> V3 0.73  0.59   0.51
#> V4 0.68  0.50   0.43

#nine hierarchical measures -- should actually use omega
r9 <- sim.hierarchical()
alpha(r9)

#> 
#> Reliability analysis   
#> Call: alpha(x = r9)
#> 
#>   raw_alpha std.alpha G6(smc) average_r S/N
#>       0.76      0.76    0.76      0.26 3.2
#> 
#>  Reliability if an item is dropped:
#>    raw_alpha std.alpha G6(smc) average_r S/N
#> V1      0.71      0.71    0.70      0.24 2.5
#> V2      0.72      0.72    0.71      0.25 2.6
#> V3      0.74      0.74    0.73      0.26 2.8
#> V4      0.73      0.73    0.72      0.25 2.7
#> V5      0.74      0.74    0.73      0.26 2.9
#> V6      0.75      0.75    0.74      0.27 3.0
#> V7      0.75      0.75    0.74      0.27 3.0
#> V8      0.76      0.76    0.75      0.28 3.1
#> V9      0.77      0.77    0.76      0.29 3.3
#> 
#>  Item statistics 
#>       r r.cor r.drop
#> V1 0.72  0.71   0.61
#> V2 0.67  0.63   0.54
#> V3 0.61  0.55   0.47
#> V4 0.65  0.59   0.51
#> V5 0.59  0.52   0.45
#> V6 0.53  0.43   0.38
#> V7 0.56  0.46   0.40
#> V8 0.50  0.39   0.34
#> V9 0.45  0.32   0.28

# examples of two independent factors that produce reasonable alphas
#this is a case where alpha is a poor indicator of unidimensionality
two.f <- sim.item(8)
#specify which items to reverse key by name
 alpha(two.f,keys=c("V1","V2","V7","V8"))

#> 
#> Reliability analysis   
#> Call: alpha(x = two.f, keys = c("V1", "V2", "V7", "V8"))
#> 
#>   raw_alpha std.alpha G6(smc) average_r S/N   ase  mean   sd
#>       0.58      0.58    0.62      0.15 1.4 0.029 0.031 0.51
#> 
#>  lower alpha upper     95% confidence boundaries
#> 0.53 0.58 0.64 
#> 
#>  Reliability if an item is dropped:
#>     raw_alpha std.alpha G6(smc) average_r S/N alpha se
#> V1-      0.55      0.55    0.58      0.15 1.2    0.031
#> V2-      0.53      0.53    0.57      0.14 1.1    0.033
#> V3       0.55      0.55    0.58      0.15 1.2    0.031
#> V4       0.55      0.54    0.58      0.15 1.2    0.032
#> V5       0.56      0.56    0.59      0.15 1.3    0.031
#> V6       0.56      0.56    0.59      0.15 1.3    0.031
#> V7-      0.53      0.53    0.56      0.14 1.1    0.033
#> V8-      0.56      0.57    0.59      0.16 1.3    0.030
#> 
#>  Item statistics 
#>       n raw.r std.r r.cor r.drop     mean   sd
#> V1- 500  0.50  0.50  0.38   0.28  0.09148 1.01
#> V2- 500  0.55  0.55  0.46   0.34  0.10501 1.00
#> V3  500  0.50  0.50  0.39   0.27 -0.04114 1.05
#> V4  500  0.51  0.51  0.41   0.30 -0.04698 0.99
#> V5  500  0.48  0.48  0.36   0.26  0.00021 1.01
#> V6  500  0.48  0.48  0.36   0.26 -0.00965 1.00
#> V7- 500  0.56  0.56  0.48   0.35  0.08102 1.00
#> V8- 500  0.45  0.46  0.33   0.23  0.07120 0.96

 #by location
 alpha(two.f,keys=c(1,2,7,8))

#> 
#> Reliability analysis   
#> Call: alpha(x = two.f, keys = c(1, 2, 7, 8))
#> 
#>   raw_alpha std.alpha G6(smc) average_r S/N   ase  mean   sd
#>       0.58      0.58    0.62      0.15 1.4 0.029 0.031 0.51
#> 
#>  lower alpha upper     95% confidence boundaries
#> 0.53 0.58 0.64 
#> 
#>  Reliability if an item is dropped:
#>     raw_alpha std.alpha G6(smc) average_r S/N alpha se
#> V1-      0.55      0.55    0.58      0.15 1.2    0.031
#> V2-      0.53      0.53    0.57      0.14 1.1    0.033
#> V3       0.55      0.55    0.58      0.15 1.2    0.031
#> V4       0.55      0.54    0.58      0.15 1.2    0.032
#> V5       0.56      0.56    0.59      0.15 1.3    0.031
#> V6       0.56      0.56    0.59      0.15 1.3    0.031
#> V7-      0.53      0.53    0.56      0.14 1.1    0.033
#> V8-      0.56      0.57    0.59      0.16 1.3    0.030
#> 
#>  Item statistics 
#>       n raw.r std.r r.cor r.drop     mean   sd
#> V1- 500  0.50  0.50  0.38   0.28  0.09148 1.01
#> V2- 500  0.55  0.55  0.46   0.34  0.10501 1.00
#> V3  500  0.50  0.50  0.39   0.27 -0.04114 1.05
#> V4  500  0.51  0.51  0.41   0.30 -0.04698 0.99
#> V5  500  0.48  0.48  0.36   0.26  0.00021 1.01
#> V6  500  0.48  0.48  0.36   0.26 -0.00965 1.00
#> V7- 500  0.56  0.56  0.48   0.35  0.08102 1.00
#> V8- 500  0.45  0.46  0.33   0.23  0.07120 0.96

 #automatic reversal base upon first component
alpha(two.f)

#> Warning: Some items were negatively correlated with the total scale and probably 
#> should be reversed.  
#> To do this, run the function again with the 'check.keys=TRUE' option

#> Some items ( V3 V4 V5 V6 ) were negatively correlated with the total scale and 
#> probably should be reversed.  
#> To do this, run the function again with the 'check.keys=TRUE' option

#> Warning: NaNs produced

#> 
#> Reliability analysis   
#> Call: alpha(x = two.f)
#> 
#>   raw_alpha std.alpha G6(smc) average_r   S/N  ase    mean   sd
#>      -0.58     -0.58   -0.17    -0.048 -0.37 0.11 -0.0042 0.29
#> 
#>  lower alpha upper     95% confidence boundaries
#> -0.8 -0.58 -0.37 
#> 
#>  Reliability if an item is dropped:
#>    raw_alpha std.alpha G6(smc) average_r   S/N alpha se
#> V1     -0.54     -0.53  -0.147    -0.052 -0.35    0.106
#> V2     -0.52     -0.52  -0.149    -0.051 -0.34    0.105
#> V3     -0.37     -0.36  -0.041    -0.039 -0.27    0.094
#> V4     -0.40     -0.37  -0.049    -0.040 -0.27    0.097
#> V5     -0.49     -0.48  -0.119    -0.049 -0.32    0.103
#> V6     -0.56     -0.56  -0.172    -0.054 -0.36    0.107
#> V7     -0.39     -0.40  -0.083    -0.043 -0.29    0.097
#> V8     -0.57     -0.57  -0.177    -0.055 -0.36    0.109
#> 
#>  Item statistics 
#>      n raw.r std.r r.cor r.drop     mean   sd
#> V1 500  0.33  0.33   NaN -0.112  0.01170 1.01
#> V2 500  0.32  0.32   NaN -0.123 -0.00183 1.00
#> V3 500  0.24  0.21   NaN -0.219 -0.04114 1.05
#> V4 500  0.22  0.22   NaN -0.204 -0.04698 0.99
#> V5 500  0.30  0.30   NaN -0.144  0.00021 1.01
#> V6 500  0.34  0.34   NaN -0.100 -0.00965 1.00
#> V7 500  0.23  0.24   NaN -0.205  0.02217 1.00
#> V8 500  0.33  0.35   NaN -0.091  0.03198 0.96

#an example with discrete item responses  -- show the frequencies
items <- sim.congeneric(N=500,short=FALSE,low=-2,high=2,
        categorical=TRUE) #500 responses to 4 discrete items with 5 categories
a4 <- alpha(items$observed)  #item response analysis of congeneric measures
a4

#> 
#> Reliability analysis   
#> Call: alpha(x = items$observed)
#> 
#>   raw_alpha std.alpha G6(smc) average_r S/N  ase   mean   sd
#>       0.73      0.73    0.68       0.4 2.7 0.02 -0.013 0.76
#> 
#>  lower alpha upper     95% confidence boundaries
#> 0.69 0.73 0.77 
#> 
#>  Reliability if an item is dropped:
#>    raw_alpha std.alpha G6(smc) average_r S/N alpha se
#> V1      0.61      0.61    0.52      0.34 1.6    0.031
#> V2      0.64      0.64    0.55      0.37 1.8    0.028
#> V3      0.68      0.68    0.60      0.41 2.1    0.025
#> V4      0.73      0.73    0.65      0.48 2.8    0.021
#> 
#>  Item statistics 
#>      n raw.r std.r r.cor r.drop   mean   sd
#> V1 500  0.80  0.80  0.73   0.62  0.050 1.00
#> V2 500  0.77  0.77  0.67   0.57 -0.022 1.03
#> V3 500  0.72  0.73  0.58   0.50 -0.028 0.99
#> V4 500  0.67  0.66  0.46   0.40 -0.050 1.05
#> 
#> Non missing response frequency for each item
#>      -2   -1    0    1    2 miss
#> V1 0.06 0.24 0.38 0.25 0.07    0
#> V2 0.07 0.26 0.35 0.25 0.07    0
#> V3 0.05 0.27 0.38 0.22 0.07    0
#> V4 0.10 0.22 0.39 0.22 0.07    0

#summary just gives Alpha
summary(a4)

#> 
#> Reliability analysis   
#>  raw_alpha std.alpha G6(smc) average_r S/N  ase   mean   sd
#>       0.73      0.73    0.68       0.4 2.7 0.02 -0.013 0.76