bassAckward {psych}	R Documentation

The Bass-Ackward factoring algorithm discussed by Goldberg

Description

Goldberg (2006) described a hierarchical factor structure organization from the “top down". The original idea was to do successive factor analyses from 1 to nf factors organized by factor score correlations from one level to the next. Waller (2007) discussed a simple way of doing this for components without finding the scores. Using the factor correlations (from Gorsuch) to organize factors hierarchically results may be organized at many different levels. The algorithm may be applied to principal components (pca) or to true factor analysis.

Usage

bassAckward(r, nfactors = 1, fm = "minres", rotate = "oblimin", scores = "tenBerge",
adjust=TRUE, plot=TRUE,cut=.3, use = "pairwise", cor = "cor", weight = NULL,
correct = 0.5,...)
bassAckward.diagram(x,digits=2,cut = .3,labels=NULL,marg=c(1.5,.5,1.0,.5),
main="BassAckward",items=TRUE,sort=TRUE,lr=TRUE,curves=FALSE,organize=TRUE,...) 

Arguments

r
A correlation matrix or a data matrix suitable for factoring

nfactors
Factors from 1 to nfactors will be extracted. If nfactors is a a vector, then just the number of factors specified in the vector will be extracted. (See examples).

fm
Factor method. The default is 'minres' factoring. Although to be consistent with the original Goldberg article, we can also do principal components (fm ="pca").

rotate
What type of rotation to apply. The default for factors is oblimin, for pca is varimax.

scores
What factor scoring algorithm should be used. The default is "tenBerge", other possibilities include "regression", or "bartlett"

adjust
If using any other scoring proceure that "tenBerge" should we adjust the correlations for the lack of factor score fit?

plot
By default draw a bassAckward diagram

use
How to treat missing data. Use='pairwise" finds pairwise complete correlations.

cor
What kind of correlation to find. The default is Pearson.

weight
Should cases be weighted? Default, no.

correct
If finding tetrachoric or polychoric correlations, what correction should be applied to empty cells (defaults to .5)

x
The object returned by bassAckward

digits
Number of digits to display on each path

cut
Values greater than the abs(cut) will be displayed in a path diagram.

labels
Labels may be taken from the output of the bassAckward function or can be specified as a list.

marg
Margins are set to be slightly bigger than normal to allow for a cleaner diagram

main
The main title for the figure

items
if TRUE, show the items associated with the factors

sort
if TRUE, sort the items by factor loadings

lr
Should the graphic be drawn left to right or top to bottom

curves
Should we show the correlations between factors at the same level

organize
Rename and sort the factors at two lowest levels for a more pleasing figure

...
Other graphic parameters (e.g., cex)

Details

This is essentially a wrapper to the fa and pca combined with the faCor functions. They are called repeatedly and then the weights from the resulting solutions are used to find the factor/component correlations.

Although the default is do all factor solutions from 1 to the nfactors, this can be simplified by specifying just some of the factor solutions. Thus, for the 135 items of the spi, it is more reasonable to ask for 3,5, and 27 item solutions.

The function bassAckward.diagram may be called using the diagram function or may be called directly.

The output from bassAckward.diagram is the sorted factor structure suitable for using fa.lookup.

Although not particularly pretty, it is possible to do Schmid-Leiman rotations at each level. Specify the rotation as rotate="schmid".

Value

Call
Echo the call

fm
Echos the factor method used

itemfaA list of the factor loadings at each level

bass.ack
A list of the factor correlations at each level

summary
The factors at each level

sumnames
Summary of the factor names

labels
Factor labels including items for each level

r
The correlation matrix analyzed

Phi
The factor correlations at each level

fa
The factor analysis loadings at each level, now includes Phi values

Note

Goldberg calculated factor/component scores and then correlated these. Waller suggests just looking at the unrotated components and then examining the correlations when rotating different numbers of components. I do not follow the Waller procedure, but rather find successive factors and then find factor/component correlations following Gorsuch.

Author(s)

William Revelle

References

Goldberg, L.R. (2006) Doing it all Bass-Ackwards: The development of hierarchical factor structures from the top down. Journal of Research in Personality, 40, 4, 347-358.

Gorsuch, Richard, (1983) Factor Analysis. Lawrence Erlebaum Associates.

Revelle, William. (in prep) An introduction to psychometric theory with applications in R. Springer. Working draft available at https://personality-project.org/r/book/

Waller, N. (2007), A general method for computing hierarchical component structures by Goldberg's Bass-Ackwards method, Journal of Research in Personality, 41, 4, 745-752, DOI: 10.1016/j.jrp.2006.08.005

See Also

fa, pca, omega and iclust for alternative hierarchical solutions.

Examples

bassAckward(Thurstone,4,main="Thurstone data set")
print(bassAckward(psychTools::bfi[1:25],c(2,3,5),main="bfi data set"),short=FALSE)
 #do pca instead of factors  just summarize, don't plot
summary(bassAckward(psychTools::bfi[1:25],c(1,3,5,7),fm="pca",main="Components",plot=FALSE))
##not run, but useful example
#sp5 <- bassAckward(psychTools::spi[11:145], c(3,4,5,27))

[Package psych version 1.9.12 Index]