11 emotional variables from Burt (1915)

Cyril Burt reported an early factor analysis with a circumplex structure of 11 emotional variables in 1915. 8 of these were subsequently used by Harman in his text on factor analysis. Unfortunately, it seems as if Burt made a mistake for the matrix is not positive definite. With one change from .87 to .81 the matrix is positive definite.

data(burt)

Format

A correlation matrix based upon 172 "normal school age children aged 9-12".

Sociality

Sociality

Sorrow

Sorrow

Tenderness

Tenderness

Joy

Joy

Wonder

Wonder

Elation

Elation

Disgust

Disgust

Anger

Anger

Sex

Sex

Fear

Fear

Subjection

Subjection

Details

The Burt data set is interesting for several reasons. It seems to be an early example of the organizaton of emotions into an affective circumplex, a subset of it has been used for factor analysis examples (see Harman.Burt, and it is an example of how typos affect data. The original data matrix has one negative eigenvalue. With the replacement of the correlation between Sorrow and Tenderness from .87 to .81, the matrix is positive definite.

Alternatively, using cor.smooth, the matrix can be made positive definite as well, although cor.smooth makes more (but smaller) changes.

Source

(retrieved from the web at http://www.biodiversitylibrary.org/item/95822#790) Following a suggestion by Jan DeLeeuw.

References

Burt, C.General and Specific Factors underlying the Primary Emotions. Reports of the British Association for the Advancement of Science, 85th meeting, held in Manchester, September 7-11, 1915. London, John Murray, 1916, p. 694-696 (retrieved from the web at http://www.biodiversitylibrary.org/item/95822#790)

See also

Harman.Burt in the Harman dataset and cor.smooth

Examples

data(burt)
eigen(burt)$values  #one is negative!

#>  [1]  5.16688557  1.79046638  0.96961472  0.77670247  0.69415362  0.62296025
#>  [7]  0.51271989  0.34702229  0.12987550  0.01413376 -0.02453446

burt.new <- burt
burt.new[2,3] <- burt.new[3,2] <- .81
eigen(burt.new)$values  #all are positive

#>  [1] 5.151669e+00 1.782279e+00 9.692796e-01 7.792475e-01 6.935441e-01
#>  [6] 6.177734e-01 5.042254e-01 3.423256e-01 1.319504e-01 2.770323e-02
#> [11] 2.646875e-06

bs <- cor.smooth(burt)

#> Warning: Matrix was not positive definite, smoothing was done

round(burt.new - bs,3)

#>            Sociality Sorrow Tenderness   Joy Wonder Elation Disgust  Anger
#> Sociality      0.000  0.000      0.010 0.002  0.003   0.001   0.003  0.003
#> Sorrow         0.000  0.000     -0.044 0.004  0.005   0.002   0.006  0.006
#> Tenderness     0.010 -0.044      0.000 0.001 -0.001   0.002  -0.002 -0.003
#> Joy            0.002  0.004      0.001 0.000  0.000   0.000   0.000  0.000
#> Wonder         0.003  0.005     -0.001 0.000  0.000   0.000  -0.001  0.000
#> Elation        0.001  0.002      0.002 0.000  0.000   0.000   0.000  0.001
#> Disgust        0.003  0.006     -0.002 0.000 -0.001   0.000   0.000 -0.001
#> Anger          0.003  0.006     -0.003 0.000  0.000   0.001  -0.001  0.000
#> Sex            0.000 -0.001      0.004 0.001  0.001   0.000   0.001  0.002
#> Fear           0.000  0.002      0.001 0.000  0.000   0.000   0.000  0.000
#> Subjection     0.000  0.001      0.002 0.000  0.000   0.000   0.000  0.000
#>               Sex  Fear Subjection
#> Sociality   0.000 0.000      0.000
#> Sorrow     -0.001 0.002      0.001
#> Tenderness  0.004 0.001      0.002
#> Joy         0.001 0.000      0.000
#> Wonder      0.001 0.000      0.000
#> Elation     0.000 0.000      0.000
#> Disgust     0.001 0.000      0.000
#> Anger       0.002 0.000      0.000
#> Sex         0.000 0.000      0.000
#> Fear        0.000 0.000      0.000
#> Subjection  0.000 0.000      0.000