Find item by cluster correlations, corrected for overlap and reliability

Given a n x n correlation matrix and a n x c matrix of -1,0,1 cluster weights for those n items on c clusters, find the correlation of each item with each cluster. If the item is part of the cluster, correct for item overlap. Part of the ICLUST set of functions, but useful for many item analysis problems.

cluster.loadings(keys, r.mat, correct = TRUE,SMC=TRUE)

Arguments

keys
Cluster keys: a matrix of -1,0,1 cluster weights

r.mat
A correlation matrix

correct
Correct for reliability

SMC
Use the squared multiple correlation as a communality estimate, otherwise use the greatest correlation for each variable

Details

Given a set of items to be scored as (perhaps overlapping) clusters and the intercorrelation matrix of the items, find the clusters and then the correlations of each item with each cluster. Correct for item overlap by replacing the item variance with its average within cluster inter-item correlation.

Although part of ICLUST, this may be used in any SAPA (http://sapa-project.org) application where we are interested in item- whole correlations of items and composite scales.

These loadings are particularly interpretable when sorted by absolute magnitude for each cluster (see ICLUST.sort).

Value

loadings

A matrix of item-cluster correlations (loadings)

cor

Correlation matrix of the clusters

corrected

Correlation matrix of the clusters, raw correlations below the diagonal, alpha on diagonal, corrected for reliability above the diagonal

sd

Cluster standard deviations

alpha

alpha reliabilities of the clusters

G6

G6* Modified estimated of Guttman Lambda 6

count

Number of items in the cluster

References

ICLUST: http://personality-project.org/r/r.ICLUST.html

Note

Although part of ICLUST, this may be used in any SAPA application where we are interested in item- whole correlations of items and composite scales.

See also

ICLUST, factor2cluster, cluster.cor

Examples

 r.mat<- Harman74.cor$cov
 clusters <- matrix(c(1,1,1,rep(0,24),1,1,1,1,rep(0,17)),ncol=2)
 cluster.loadings(clusters,r.mat)

#> Call: cluster.loadings(keys = clusters, r.mat = r.mat)
#> 
#> (Standardized) Alpha:
#> [1] 0.61 0.79
#> 
#> (Standardized) G6*:
#> [1] 0.66 0.84
#> 
#> Average item correlation:
#> [1] 0.35 0.48
#> 
#> Number of items:
#> [1] 3 4
#> 
#> Scale intercorrelations corrected for attenuation 
#>  raw correlations below the diagonal, alpha on the diagonal 
#>  corrected correlations above the diagonal:
#>      [,1] [,2]
#> [1,] 0.61 0.69
#> [2,] 0.48 0.79
#> 
#> Item by scale intercorrelations
#>  corrected for item overlap and scale reliability
#>                         [,1] [,2]
#> VisualPerception       0.675 0.50
#> Cubes                  0.512 0.32
#> PaperFormBoard         0.636 0.36
#> Flags                  0.550 0.45
#> GeneralInformation     0.468 0.76
#> PargraphComprehension  0.459 0.82
#> SentenceCompletion     0.375 0.84
#> WordClassification     0.478 0.74
#> WordMeaning            0.387 0.85
#> Addition               0.054 0.30
#> Code                   0.301 0.36
#> CountingDots           0.328 0.23
#> StraightCurvedCapitals 0.575 0.46
#> WordRecognition        0.222 0.30
#> NumberRecognition      0.238 0.27
#> FigureRecognition      0.520 0.34
#> ObjectNumber           0.196 0.31
#> NumberFigure           0.457 0.29
#> FigureWord             0.380 0.28
#> Deduction              0.523 0.57
#> NumericalPuzzles       0.461 0.43
#> ProblemReasoning       0.491 0.56
#> SeriesCompletion       0.661 0.56
#> ArithmeticProblems     0.382 0.54