Convert correlations to distances (necessary to do multidimensional scaling of correlation data)

A minor helper function to convert correlations (ranging from -1 to 1) to distances (ranging from 0 to 2). \(d = \sqrt{(2(1-r))}\).

cor2dist(x)

Arguments

x
If square, then assumed to be a correlation matrix, otherwise the correlations are found first.

Value

dist: a square matrix of distances.

References

Revelle, William. (in prep) An introduction to psychometric theory with applications in R. Springer. Working draft available at http://personality-project.org/r/book/

Note

For an example of doing multidimensional scaling on data that are normally factored, see Revelle (in prep)