fa.poly {psych}	R Documentation

Deprecated Exploratory Factor analysis functions. Please use fa

Description

After 6 years, it is time to stop using these deprecated functions! Please see fa which includes all of the functionality of these older functions.

Usage

fa.poly(x,nfactors=1,n.obs = NA, n.iter=1, rotate="oblimin", SMC=TRUE,  missing=FALSE,
 impute="median", min.err = .001, max.iter=50, symmetric=TRUE, warnings=TRUE,
 fm="minres",alpha=.1, p =.05,scores="regression", oblique.scores=TRUE,
        weight=NULL,global=TRUE,...)  #deprecated
        
factor.minres(r, nfactors=1, residuals = FALSE, rotate = "varimax",n.obs = NA,
scores = FALSE,SMC=TRUE, missing=FALSE,impute="median",min.err = 0.001, digits = 2,
 max.iter = 50,symmetric=TRUE,warnings=TRUE,fm="minres")    #deprecated

factor.wls(r,nfactors=1,residuals=FALSE,rotate="varimax",n.obs = NA,
scores=FALSE,SMC=TRUE,missing=FALSE,impute="median", min.err = .001,
 digits=2,max.iter=50,symmetric=TRUE,warnings=TRUE,fm="wls")  #deprecated

Arguments

r
deprecated.

x
deprecated

nfactors
deprecated

n.obs
deprecated

rotate
deprecated

n.iter
deprecated

residuals
deprecated

scores
deprecated

SMC
deprecated

missing
deprecated

impute
deprecated

max.iter
deprecated

symmetric
deprecated

warnings
deprecated

fm
deprecated

alpha
deprecated

p
deprecated

oblique.scores
deprecated

weight
deprecated

global
deprecated

digits
deprecated

min.err
deprecated

...
deprecated

Details

Please see the writeup for fa for all of the functionality in these older functions.

Value

Pleases see the writeup for fa

Note

These functions have been deprecated for 8 years. Don't use them.

Author(s)

William Revelle

References

Revelle, William. (in prep) An introduction to psychometric theory with applications in R. Springer. Working draft available at https://personality-project.org/r/book/

Examples

#none, you should see fa
#using the Harman 24 mental tests, compare a principal factor with a principal components solution

[Package psych version 2.6.1 Index]