Convert a data frame, correlation matrix, or factor analysis output to a LaTeX table

A set of handy helper functions to convert data frames or matrices to LaTeX tables. Although Sweave is the preferred means of converting R output to LaTeX, it is sometimes useful to go directly from a data.frame or matrix to a LaTeX table. cor2latex will find the correlations and then create a lower (or upper) triangular matrix for latex output. fa2latex will create the latex commands for showing the loadings and factor intercorrelations. As the default option, tables are prepared in an approximation of APA format.

df2latex(x,digits=2,rowlabels=TRUE,apa=TRUE,short.names=TRUE,font.size ="scriptsize",
       big.mark=NULL,drop.na=TRUE, heading="A table from the psych package in R",
   caption="df2latex",label="default", char=FALSE,
    stars=FALSE,silent=FALSE,file=NULL,append=FALSE,cut=0,big=0)
cor2latex(x,use = "pairwise", method="pearson", adjust="holm",stars=FALSE,
       digits=2,rowlabels=TRUE,lower=TRUE,apa=TRUE,short.names=TRUE,
     font.size ="scriptsize", heading="A correlation table from the psych package in R.",
      caption="cor2latex",label="default",silent=FALSE,file=NULL,append=FALSE,cut=0,big=0)
fa2latex(f,digits=2,rowlabels=TRUE,apa=TRUE,short.names=FALSE,cumvar=FALSE,
       cut=0,big=.3,alpha=.05,font.size ="scriptsize",
       heading="A factor analysis table from the psych package in R",
       caption="fa2latex",label="default",silent=FALSE,file=NULL,append=FALSE)
omega2latex(f,digits=2,rowlabels=TRUE,apa=TRUE,short.names=FALSE,cumvar=FALSE,cut=.2,
        font.size ="scriptsize",
        heading="An omega analysis table from the psych package in R",
        caption="omega2latex",label="default",silent=FALSE,file=NULL,append=FALSE)

irt2latex(f,digits=2,rowlabels=TRUE,apa=TRUE,short.names=FALSE,
       font.size ="scriptsize", heading="An IRT factor analysis table from R",
       caption="fa2latex",label="default",silent=FALSE,file=NULL,append=FALSE)
ICC2latex(icc,digits=2,rowlabels=TRUE,apa=TRUE,ci=TRUE,
   font.size ="scriptsize",big.mark=NULL, drop.na=TRUE,
    heading="A table from the psych package in R",
   caption="ICC2latex",label="default",char=FALSE,silent=FALSE,file=NULL,append=FALSE)

Arguments

x
A data frame or matrix to convert to LaTeX. If non-square, then correlations will be found prior to printing in cor2latex

digits
Round the output to digits of accuracy. NULL for formatting character data

rowlabels
If TRUE, use the row names from the matrix or data.frame

short.names
Name the columns with abbreviated rownames to save space

apa
If TRUE formats table in APA style

cumvar
For factor analyses, should we show the cumulative variance accounted for?

font.size
e.g., "scriptsize", "tiny" or anyother acceptable LaTeX font size.

heading
The label appearing at the top of the table

caption
The table caption

lower
in cor2latex, just show the lower triangular matrix

f
The object returned from a factor analysis using fa or irt.fa.

label
The label for the table

big.mark
Comma separate numbers large numbers (big.mark=",")

drop.na
Do not print NA values

method
When finding correlations, which method should be used (pearson)

use
use="pairwise" is the default when finding correlations in cor2latex

adjust
If showing probabilities, which adjustment should be used (holm)

stars
Should probability asterixs be displayed in cor2latex (FALSE)

char
char=TRUE allows printing tables with character information, but does not allow for putting in commas into numbers

cut
In omega2latex, df2latex and fa2latex, do not print abs(values) < cut

big
In fa2latex and df2latex boldface those abs(values) > big

alpha
If fa has returned confidence intervals, then what values of loadings should be boldfaced?

icc
Either the output of an ICC, or the data to be analyzed.

ci
Should confidence intervals of the ICC be displayed

silent
If TRUE, do not print any output, just return silently -- useful if using Sweave

file
If specified, write the output to this file

append
If file is specified, then should we append (append=TRUE) or just write to the file

Value

A LaTeX table. Note that if showing "stars" for correlations, then one needs to use the siunitx package in LaTex. The entire LaTeX output is also returned invisibly. If using Sweave to create tables, then the silent option should be set to TRUE and the returned object saved as a file. See the last example.

See also

The many LaTeX conversion routines in Hmisc.

Examples

df2latex(Thurstone,rowlabels=FALSE,apa=FALSE,short.names=FALSE,
        caption="Thurstone Correlation matrix")

#> % df2latex % Thurstone % FALSE % FALSE % FALSE % Thurstone Correlation matrix 
#> \begin{table}[htpb]\caption{Thurstone Correlation matrix}
#> \begin{center}
#> \begin{scriptsize} 
#> \begin{tabular} { r r r r r r r r r r }
#> Sentences  &  Vocabulary  &  Sent.Completion  &  First.Letters  &  Four.Letter.Words  &  Suffixes  &  Letter.Series  &  Pedigrees  &  Letter.Group \cr 
#>  \cr 
#>   &  1.00  &  0.83  &  0.78  &  0.44  &  0.43  &  0.45  &  0.45  &  0.54  &  0.38 \cr 
#>    &  0.83  &  1.00  &  0.78  &  0.49  &  0.46  &  0.49  &  0.43  &  0.54  &  0.36 \cr 
#>    &  0.78  &  0.78  &  1.00  &  0.46  &  0.42  &  0.44  &  0.40  &  0.53  &  0.36 \cr 
#>    &  0.44  &  0.49  &  0.46  &  1.00  &  0.67  &  0.59  &  0.38  &  0.35  &  0.42 \cr 
#>    &  0.43  &  0.46  &  0.42  &  0.67  &  1.00  &  0.54  &  0.40  &  0.37  &  0.45 \cr 
#>    &  0.45  &  0.49  &  0.44  &  0.59  &  0.54  &  1.00  &  0.29  &  0.32  &  0.32 \cr 
#>    &  0.45  &  0.43  &  0.40  &  0.38  &  0.40  &  0.29  &  1.00  &  0.56  &  0.60 \cr 
#>    &  0.54  &  0.54  &  0.53  &  0.35  &  0.37  &  0.32  &  0.56  &  1.00  &  0.45 \cr 
#>    &  0.38  &  0.36  &  0.36  &  0.42  &  0.45  &  0.32  &  0.60  &  0.45  &  1.00 \cr 
#>  \hline 
#> \end{tabular}
#> \end{scriptsize}
#> \end{center}
#> \label{default}
#> \end{table} 
#> 

df2latex(Thurstone,heading="Thurstone Correlation matrix in APA style")

#> % df2latex % Thurstone % Thurstone Correlation matrix in APA style 
#> \begin{table}[htpb]\caption{df2latex}
#> \begin{center}
#> \begin{scriptsize} 
#> \begin{tabular} {l r r r r r r r r r }
#>  \multicolumn{ 9 }{l}{ Thurstone Correlation matrix in APA style } \cr 
#>  \hline Variable  &   Sntnc  &  Vcblr  &  Snt.C  &  Frs.L  &  F.L.W  &  Sffxs  &  Ltt.S  &  Pdgrs  &  Ltt.G \cr 
#>   \hline 
#> Sentences   &  1.00  &  0.83  &  0.78  &  0.44  &  0.43  &  0.45  &  0.45  &  0.54  &  0.38 \cr 
#>  Vocabulary   &  0.83  &  1.00  &  0.78  &  0.49  &  0.46  &  0.49  &  0.43  &  0.54  &  0.36 \cr 
#>  Sent.Completion   &  0.78  &  0.78  &  1.00  &  0.46  &  0.42  &  0.44  &  0.40  &  0.53  &  0.36 \cr 
#>  First.Letters   &  0.44  &  0.49  &  0.46  &  1.00  &  0.67  &  0.59  &  0.38  &  0.35  &  0.42 \cr 
#>  Four.Letter.Words   &  0.43  &  0.46  &  0.42  &  0.67  &  1.00  &  0.54  &  0.40  &  0.37  &  0.45 \cr 
#>  Suffixes   &  0.45  &  0.49  &  0.44  &  0.59  &  0.54  &  1.00  &  0.29  &  0.32  &  0.32 \cr 
#>  Letter.Series   &  0.45  &  0.43  &  0.40  &  0.38  &  0.40  &  0.29  &  1.00  &  0.56  &  0.60 \cr 
#>  Pedigrees   &  0.54  &  0.54  &  0.53  &  0.35  &  0.37  &  0.32  &  0.56  &  1.00  &  0.45 \cr 
#>  Letter.Group   &  0.38  &  0.36  &  0.36  &  0.42  &  0.45  &  0.32  &  0.60  &  0.45  &  1.00 \cr 
#>  \hline 
#> \end{tabular}
#> \end{scriptsize}
#> \end{center}
#> \label{default}
#> \end{table} 
#> 

df2latex(describe(sat.act)[2:10],short.names=FALSE)

#> % df2latex % describe(sat.act)[2:10] % FALSE 
#> \begin{table}[htpb]\caption{df2latex}
#> \begin{center}
#> \begin{scriptsize} 
#> \begin{tabular} {l r r r r r r r r r }
#>  \multicolumn{ 9 }{l}{ A table from the psych package in R } \cr 
#>  \hline Variable  &   n  &  mean  &  sd  &  median  &  trimmed  &  mad  &  min  &  max  &  range \cr 
#>   \hline 
#> gender   &  700  &    1.65  &    0.48  &    2  &    1.68  &    0.00  &    1  &    2  &    1 \cr 
#>  education   &  700  &    3.16  &    1.43  &    3  &    3.31  &    1.48  &    0  &    5  &    5 \cr 
#>  age   &  700  &   25.59  &    9.50  &   22  &   23.86  &    5.93  &   13  &   65  &   52 \cr 
#>  ACT   &  700  &   28.55  &    4.82  &   29  &   28.84  &    4.45  &    3  &   36  &   33 \cr 
#>  SATV   &  700  &  612.23  &  112.90  &  620  &  619.45  &  118.61  &  200  &  800  &  600 \cr 
#>  SATQ   &  687  &  610.22  &  115.64  &  620  &  617.25  &  118.61  &  200  &  800  &  600 \cr 
#>  \hline 
#> \end{tabular}
#> \end{scriptsize}
#> \end{center}
#> \label{default}
#> \end{table} 
#> 

cor2latex(Thurstone)

#> % df2latex % R % digits % rowlabels % apa % short.names % font.size % heading % caption % label % TRUE % stars % silent % file % append % cut % big 
#> \begin{table}[htpb]\caption{cor2latex}
#> \begin{center}
#> \begin{scriptsize} 
#> \begin{tabular} {l r r r r r r r r r }
#>  \multicolumn{ 9 }{l}{ A correlation table from the psych package in R. } \cr 
#>  \hline Variable  &   Sntnc  &  Vcblr  &  Snt.C  &  Frs.L  &  F.L.W  &  Sffxs  &  Ltt.S  &  Pdgrs  &  Ltt.G \cr 
#>   \hline 
#> Sentences   &  1.00   &    &    &    &    &    &    &    &   \cr 
#>  Vocabulary   &  0.83  &  1.00   &    &    &    &    &    &    &   \cr 
#>  Sent.Completion   &  0.78  &  0.78  &  1.00   &    &    &    &    &    &   \cr 
#>  First.Letters   &  0.44  &  0.49  &  0.46  &  1.00   &    &    &    &    &   \cr 
#>  Four.Letter.Words   &  0.43  &  0.46  &  0.42  &  0.67  &  1.00   &    &    &    &   \cr 
#>  Suffixes   &  0.45  &  0.49  &  0.44  &  0.59  &  0.54  &  1.00   &    &    &   \cr 
#>  Letter.Series   &  0.45  &  0.43  &  0.40  &  0.38  &  0.40  &  0.29  &  1.00   &    &   \cr 
#>  Pedigrees   &  0.54  &  0.54  &  0.53  &  0.35  &  0.37  &  0.32  &  0.56  &  1.00   &   \cr 
#>  Letter.Group   &  0.38  &  0.36  &  0.36  &  0.42  &  0.45  &  0.32  &  0.60  &  0.45  &  1.00  \cr 
#>  \hline 
#> \end{tabular}
#> \end{scriptsize}
#> \end{center}
#> \label{default}
#> \end{table} 
#> 

cor2latex(sat.act,short.names=FALSE)

#> % df2latex % R % digits % rowlabels % apa % short.names % font.size % heading % caption % label % TRUE % stars % silent % file % append % cut % big 
#> \begin{table}[htpb]\caption{cor2latex}
#> \begin{center}
#> \begin{scriptsize} 
#> \begin{tabular} {l r r r r r r }
#>  \multicolumn{ 6 }{l}{ A correlation table from the psych package in R. } \cr 
#>  \hline Variable  &   gender  &  education  &  age  &  ACT  &  SATV  &  SATQ \cr 
#>   \hline 
#> gender   &   1.00   &    &    &    &    &   \cr 
#>  education   &   0.09  &   1.00   &    &    &    &   \cr 
#>  age   &  -0.02  &   0.55  &   1.00   &    &    &   \cr 
#>  ACT   &  -0.04  &   0.15  &   0.11  &   1.00   &    &   \cr 
#>  SATV   &  -0.02  &   0.05  &  -0.04  &   0.56  &   1.00   &   \cr 
#>  SATQ   &  -0.17  &   0.03  &  -0.03  &   0.59  &   0.64  &   1.00  \cr 
#>  \hline 
#> \end{tabular}
#> \end{scriptsize}
#> \end{center}
#> \label{default}
#> \end{table} 
#> 

fa2latex(fa(Thurstone,3),heading="Factor analysis from R in quasi APA style")

#> % Called in the psych package  fa2latex % Called in the psych package  fa(Thurstone, 3) % Called in the psych package  Factor analysis from R in quasi APA style 
#> \begin{table}[htpb]\caption{fa2latex}
#> \begin{center}
#> \begin{scriptsize} 
#> \begin{tabular} {l r r r r r r }
#>  \multicolumn{ 6 }{l}{ Factor analysis from R in quasi APA style } \cr 
#>  \hline Variable  &   MR1  &  MR2  &  MR3  &  h2  &  u2  &  com \cr 
#>   \hline 
#> Sentences   &  \bf{ 0.90}  &  -0.03  &   0.04  &  0.82  &  0.18  &  1.01 \cr 
#>  Vocabulary   &  \bf{ 0.89}  &   0.06  &  -0.03  &  0.84  &  0.16  &  1.01 \cr 
#>  Sent.Completion   &  \bf{ 0.84}  &   0.03  &   0.00  &  0.74  &  0.26  &  1.00 \cr 
#>  First.Letters   &   0.00  &  \bf{ 0.85}  &   0.00  &  0.73  &  0.27  &  1.00 \cr 
#>  Four.Letter.Words   &  -0.02  &  \bf{ 0.75}  &   0.10  &  0.63  &  0.37  &  1.04 \cr 
#>  Suffixes   &   0.18  &  \bf{ 0.63}  &  -0.08  &  0.50  &  0.50  &  1.20 \cr 
#>  Letter.Series   &   0.03  &  -0.01  &  \bf{ 0.84}  &  0.73  &  0.27  &  1.00 \cr 
#>  Pedigrees   &  \bf{ 0.38}  &  -0.05  &  \bf{ 0.46}  &  0.51  &  0.49  &  1.96 \cr 
#>  Letter.Group   &  -0.06  &   0.21  &  \bf{ 0.63}  &  0.52  &  0.48  &  1.25 \cr 
#> \hline \cr SS loadings & 2.65 &  1.87 &  1.49 &  \cr  
#> \cr 
#>             \hline \cr 
#> MR1   & 1.00 & 0.59 & 0.53 \cr 
#>  MR2   & 0.59 & 1.00 & 0.52 \cr 
#>  MR3   & 0.53 & 0.52 & 1.00 \cr 
#>  \hline 
#> \end{tabular}
#> \end{scriptsize}
#> \end{center}
#> \label{default}
#> \end{table} 
#> 

#If using Sweave to create a LateX table as a separate file then set silent=TRUE
#e.g., 
#LaTex preamble 
#....
#<<print=FALSE,echo=FALSE>>= 
#f3 <- fa(Thurstone,3)
#fa2latex(f3,silent=TRUE,file='testoutput.tex')
#@
#
#\input{testoutput.tex}