kaiser {psych}	R Documentation

Apply the Kaiser normalization when rotating factors

Description

Kaiser (1958) suggested normalizing factor loadings before rotating them, and then denormalizing them after rotation. The GPArotation package does not (by default) normalize, nor does the fa function. Then, to make it more confusing, varimax in stats does,Varimax in GPArotation does not. kaiser will take the output of a non-normalized solution and report the normalized solution.

Usage

kaiser(f, rotate = "oblimin",m=4,pro.m=4)

Arguments

f
A factor analysis output from fa or a factor loading matrix.

rotate
Any of the standard rotations avaialable in the GPArotation package.

m
a parameter to pass to Promax

pro.m
A redundant parameter, which is used to replace m in calls to Promax

Details

Best results if called from an unrotated solution. Repeated calls using a rotated solution will produce incorrect estimates of the correlations between the factors.

Value

See the values returned by GPArotation functions

Note

Prepared in response to a question about why fa oblimin results are different from SPSS.

Author(s)

William Revelle

References

Kaiser, H. F. (1958) The varimax criterion for analytic rotation in factor analysis. Psychometrika 23, 187-200.

See Also

fa, Promax

Examples

f3 <- fa(Thurstone,3)
f3n <- kaiser(fa(Thurstone,3,rotate="none"))
f3p <- kaiser(fa(Thurstone,3,rotate="none"),rotate="Promax",m=3)
factor.congruence(list(f3,f3n,f3p))

[Package psych version 1.9.12 Index]