Find Cohen's kappa and weighted kappa coefficients for correlation of two raters

Cohen's kappa (Cohen, 1960) and weighted kappa (Cohen, 1968) may be used to find the agreement of two raters when using nominal scores.

weighted.kappa is (probability of observed matches - probability of expected matches)/(1 - probability of expected matches). Kappa just considers the matches on the main diagonal. Weighted kappa considers off diagonal elements as well.

cohen.kappa(x, w=NULL,n.obs=NULL,alpha=.05,levels=NULL)
wkappa(x, w = NULL)    #deprectated

Arguments

x
Either a two by n data with categorical values from 1 to p or a p x p table. If a data array, a table will be found.

w
A p x p matrix of weights. If not specified, they are set to be 0 (on the diagonal) and (distance from diagonal) off the diagonal)^2.

n.obs
Number of observations (if input is a square matrix.

alpha
Probability level for confidence intervals

levels
Specify the number of levels if some levels of x or y are completely missing

Details

When cateogorical judgments are made with two cateories, a measure of relationship is the phi coefficient. However, some categorical judgments are made using more than two outcomes. For example, two diagnosticians might be asked to categorize patients three ways (e.g., Personality disorder, Neurosis, Psychosis) or to categorize the stages of a disease. Just as base rates affect observed cell frequencies in a two by two table, they need to be considered in the n-way table (Cohen, 1960).

Kappa considers the matches on the main diagonal. A penalty function (weight) may be applied to the off diagonal matches. If the weights increase by the square of the distance from the diagonal, weighted kappa is similar to an Intra Class Correlation (ICC).

Derivations of weighted kappa are sometimes expressed in terms of similarities, and sometimes in terms of dissimilarities. In the latter case, the weights on the diagonal are 1 and the weights off the diagonal are less than one. In this, if the weights are 1 - squared distance from the diagonal / k, then the result is similar to the ICC (for any positive k).

cohen.kappa may use either similarity weighting (diagonal = 0) or dissimilarity weighting (diagonal = 1) in order to match various published examples.

The input may be a two column data.frame or matrix with columns representing the two judges and rows the subjects being rated. Alternatively, the input may be a square n x n matrix of counts or proportion of matches. If proportions are used, it is necessary to specify the number of observations (n.obs) in order to correctly find the confidence intervals.

The confidence intervals are based upon the variance estimates discussed by Fleiss, Cohen, and Everitt who corrected the formulae of Cohen (1968) and Blashfield.

Some data sets will include data with numeric categories with some category values missing completely. In the sense that kappa is a measure of category relationship, this should not matter. But when finding weighted kappa, the number of categories weighted will be less than the number of categories potentially in the data. This can be remedied by specifying the levels parameter.

Value

kappa

Unweighted kappa

weighted.kappa

The default weights are quadratric.

var.kappa

Variance of kappa

var.weighted

Variance of weighted kappa

n.obs

number of observations

weight

The weights used in the estimation of weighted kappa

confid

The alpha/2 confidence intervals for unweighted and weighted kappa

plevel

The alpha level used in determining the confidence limits

References

Banerjee, M., Capozzoli, M., McSweeney, L and Sinha, D. (1999) Beyond Kappa: A review of interrater agreement measures The Canadian Journal of Statistics / La Revue Canadienne de Statistique, 27, 3-23

Cohen, J. (1960). A coefficient of agreement for nominal scales. Educational and Psychological Measurement, 20 37-46

Cohen, J. (1968). Weighted kappa: Nominal scale agreement provision for scaled disagreement or partial credit. Psychological Bulletin, 70, 213-220.

Fleiss, J. L., Cohen, J. and Everitt, B.S. (1969) Large sample standard errors of kappa and weighted kappa. Psychological Bulletin, 72, 332-327.

Zwick, R. (1988) Another look at interrater agreement. Psychological Bulletin, 103, 374 - 378.

Note

As is true of many R functions, there are alternatives in other packages. The Kappa function in the vcd package estimates unweighted and weighted kappa and reports the variance of the estimate. The input is a square matrix. The ckappa and wkappa functions in the psy package take raw data matrices.

To avoid confusion with Kappa (from vcd) or the kappa function from base, the function was originally named wkappa. With additional features modified from psy::ckappa to allow input with a different number of categories, the function has been renamed cohen.kappa.

Unfortunately, to make it more confusing, the weights described by Cohen are a function of the reciprocals of those discucssed by Fleiss and Cohen. The cohen.kappa function uses the appropriate formula for Cohen or Fleiss-Cohen weights.

There are some cases where the large sample size approximation of Fleiss et al. will produce confidence intervals exceeding +/- 1. Clearly, for these cases, the upper (or lower for negative values) should be set to 1. Boot strap resampling shows the problem is that the values are not symmetric. See the last (unrun) example.

Examples

#rating data (with thanks to Tim Bates)
rater1 = c(1,2,3,4,5,6,7,8,9) # rater one's ratings
rater2 = c(1,3,1,6,1,5,5,6,7) # rater one's ratings
cohen.kappa(x=cbind(rater1,rater2))

#> Call: cohen.kappa1(x = x, w = w, n.obs = n.obs, alpha = alpha, levels = levels)
#> 
#> Cohen Kappa and Weighted Kappa correlation coefficients and confidence boundaries 
#>                  lower estimate upper
#> unweighted kappa -0.18     0.00  0.18
#> weighted kappa    0.43     0.68  0.93
#> 
#>  Number of subjects = 9 

#data matrix taken from Cohen
cohen <- matrix(c(
0.44, 0.07, 0.09,
0.05, 0.20, 0.05,
0.01, 0.03, 0.06),ncol=3,byrow=TRUE)

#cohen.weights  weight differences
cohen.weights <- matrix(c(
0,1,3,
1,0,6,
3,6,0),ncol=3)

cohen.kappa(cohen,cohen.weights,n.obs=200)

#> Warning: upper or lower confidence interval exceed  abs(1)  and set to +/- 1. 

#> Call: cohen.kappa1(x = x, w = w, n.obs = n.obs, alpha = alpha, levels = levels)
#> 
#> Cohen Kappa and Weighted Kappa correlation coefficients and confidence boundaries 
#>                  lower estimate upper
#> unweighted kappa  0.39     0.49  0.59
#> weighted kappa   -0.39     0.35  1.00
#> 
#>  Number of subjects = 200 

#cohen reports .492 and .348 

#another set of weights
#what if the weights are non-symmetric
wc <- matrix(c(
0,1,4,
1,0,6,
2,2,0),ncol=3,byrow=TRUE)
cohen.kappa(cohen,wc)

#> Warning: upper or lower confidence interval exceed  abs(1)  and set to +/- 1. 

#> Call: cohen.kappa1(x = x, w = w, n.obs = n.obs, alpha = alpha, levels = levels)
#> 
#> Cohen Kappa and Weighted Kappa correlation coefficients and confidence boundaries 
#>                  lower estimate upper
#> unweighted kappa -0.92     0.49     1
#> weighted kappa   -1.00     0.35     1
#> 
#>  Number of subjects = 1 

#Cohen reports kw = .353

cohen.kappa(cohen,n.obs=200)  #this uses the squared weights

#> Call: cohen.kappa1(x = x, w = w, n.obs = n.obs, alpha = alpha, levels = levels)
#> 
#> Cohen Kappa and Weighted Kappa correlation coefficients and confidence boundaries 
#>                  lower estimate upper
#> unweighted kappa  0.39     0.49  0.59
#> weighted kappa    0.32     0.45  0.58
#> 
#>  Number of subjects = 200 

fleiss.cohen <- 1 - cohen.weights/9
cohen.kappa(cohen,fleiss.cohen,n.obs=200)

#> Call: cohen.kappa1(x = x, w = w, n.obs = n.obs, alpha = alpha, levels = levels)
#> 
#> Cohen Kappa and Weighted Kappa correlation coefficients and confidence boundaries 
#>                  lower estimate upper
#> unweighted kappa  0.39     0.49  0.59
#> weighted kappa    0.20     0.35  0.50
#> 
#>  Number of subjects = 200 

#however, Fleiss, Cohen and Everitt weight similarities
fleiss <- matrix(c(
106, 10,4,
22,28, 10,
2, 12,  6),ncol=3,byrow=TRUE)

#Fleiss weights the similarities
weights <- matrix(c(
 1.0000, 0.0000, 0.4444,
 0.0000, 1.0000, 0.6667,
 0.4444, 0.6667, 1.0000),ncol=3)

 cohen.kappa(fleiss,weights,n.obs=200)

#> Call: cohen.kappa1(x = x, w = w, n.obs = n.obs, alpha = alpha, levels = levels)
#> 
#> Cohen Kappa and Weighted Kappa correlation coefficients and confidence boundaries 
#>                  lower estimate upper
#> unweighted kappa  0.32     0.43  0.53
#> weighted kappa    0.40     0.51  0.62
#> 
#>  Number of subjects = 200 

 #another example is comparing the scores of two sets of twins
 #data may be a 2 column matrix
 #compare weighted and unweighted
 #also look at the ICC for this data set.
 twins <- matrix(c(
    1, 2,
    2, 3,
    3, 4,
    5, 6,
    6, 7), ncol=2,byrow=TRUE)
  cohen.kappa(twins)

#> Call: cohen.kappa1(x = x, w = w, n.obs = n.obs, alpha = alpha, levels = levels)
#> 
#> Cohen Kappa and Weighted Kappa correlation coefficients and confidence boundaries 
#>                  lower estimate  upper
#> unweighted kappa -0.23    -0.14 -0.046
#> weighted kappa    0.80     0.87  0.942
#> 
#>  Number of subjects = 5 

#data may be explicitly categorical
x <- c("red","yellow","blue","red")
y <- c("red",  "blue", "blue" ,"red")
xy.df <- data.frame(x,y)
ck <- cohen.kappa(xy.df)

#> Warning: upper or lower confidence interval exceed  abs(1)  and set to +/- 1. 

ck

#> Call: cohen.kappa1(x = x, w = w, n.obs = n.obs, alpha = alpha, levels = levels)
#> 
#> Cohen Kappa and Weighted Kappa correlation coefficients and confidence boundaries 
#>                   lower estimate upper
#> unweighted kappa  0.098      0.6  1.00
#> weighted kappa   -0.693      0.0  0.69
#> 
#>  Number of subjects = 4 

ck$agree

#>         x2f
#> x1f      blue  red yellow
#>   blue   0.25 0.00   0.00
#>   red    0.00 0.50   0.00
#>   yellow 0.25 0.00   0.00

#The problem of missing categories (from Amy Finnegan)
numbers <- data.frame(rater1=c(6,3,7,8,7),
                      rater2=c(6,1,8,5,10))
cohen.kappa(numbers)  #compare with the next analysis

#> Warning: upper or lower confidence interval exceed  abs(1)  and set to +/- 1. 

#> Call: cohen.kappa1(x = x, w = w, n.obs = n.obs, alpha = alpha, levels = levels)
#> 
#> Cohen Kappa and Weighted Kappa correlation coefficients and confidence boundaries 
#>                   lower estimate upper
#> unweighted kappa -0.161     0.13  0.42
#> weighted kappa   -0.037     0.53  1.00
#> 
#>  Number of subjects = 5 

cohen.kappa(numbers,levels=1:10)  #specify the number of levels 

#> Warning: upper or lower confidence interval exceed  abs(1)  and set to +/- 1. 

#> Call: cohen.kappa1(x = x, w = w, n.obs = n.obs, alpha = alpha, levels = levels)
#> 
#> Cohen Kappa and Weighted Kappa correlation coefficients and confidence boundaries 
#>                  lower estimate upper
#> unweighted kappa -0.16     0.13  0.42
#> weighted kappa    0.23     0.62  1.00
#> 
#>  Number of subjects = 5 

              #   these leads to slightly higher weighted kappa

#finally, input can be a data.frame of ratings from more than two raters
ratings <- matrix(rep(1:5,4),ncol=4)
ratings[1,2] <- ratings[2,3] <- ratings[3,4] <- NA
ratings[2,1] <- ratings[3,2] <- ratings[4,3] <- 1
cohen.kappa(ratings)

#> Warning: upper or lower confidence interval exceed  abs(1)  and set to +/- 1. 

#> Warning: upper or lower confidence interval exceed  abs(1)  and set to +/- 1. 

#> Warning: upper or lower confidence interval exceed  abs(1)  and set to +/- 1. 

#> Warning: upper or lower confidence interval exceed  abs(1)  and set to +/- 1. 

#> Warning: upper or lower confidence interval exceed  abs(1)  and set to +/- 1. 

#> 
#> Cohen Kappa (below the diagonal) and Weighted Kappa (above the diagonal) 
#> For confidence intervals and detail print with all=TRUE
#>      R1   R2   R3   R4
#> R1 1.00 0.74 0.67 0.92
#> R2 0.38 1.00 0.48 1.00
#> R3 0.67 0.14 1.00 0.80
#> R4 0.67 1.00 0.50 1.00

 #In the case of confidence intervals being artificially truncated to +/- 1, it is 
 #helpful to compare the results of a boot strap resample
 #ck.boot <-function(x,s=1:nrow(x)) {cohen.kappa(x[s,])$kappa}
 #library(boot)
 #ckb <- boot(x,ck.boot,R=1000)
 #hist(ckb$t)