mat.sort {psych}	R Documentation

Sort the elements of a correlation matrix to reflect factor loadings

Description

To see the structure of a correlation matrix, it is helpful to organize the items so that the similar items are grouped together. One such grouping technique is factor analysis. mat.sort will sort the items by a factor model (if specified), or any other order, or by the loadings on the first factor (if unspecified)

Usage

mat.sort(m, f = NULL)
matSort(m, f = NULL)

Arguments

m
A correlation matrix

f
A factor analysis output (i.e., one with a loadings matrix) or a matrix of weights

Details

The factor analysis output is sorted by size of the largest factor loading for each variable and then the matrix items are organized by those loadings. The default is to sort by the loadings on the first factor. Alternatives allow for ordering based upon any vector or matrix.

Value

A sorted correlation matrix, suitable for showing with corPlot.

Author(s)

William Revelle

See Also

fa, corPlot

Examples

data(Bechtoldt.1)
sorted <- mat.sort(Bechtoldt.1,fa(Bechtoldt.1,5))
corPlot(sorted,xlas=2) #vertical xaxis names

[Package psych version 1.9.12 Index]