Miscellaneous helper functions for the psych package

This is a set of minor, if not trivial, helper functions. lowerCor finds the correlation of x variables and then prints them using lowerMat which is a trivial, but useful, function to round off and print the lower triangle of a matrix. reflect reflects the output of a factor analysis or principal components analysis so that one or more factors is reflected. (Requested by Alexander Weiss.) progressBar prints out ... as a calling routine (e.g., tetrachoric) works through a tedious calculation. shannon finds the Shannon index (H) of diversity or of information. test.all tests all the examples in a package. best.items sorts a factor matrix for absolute values and displays the expanded items names. fa.lookup returns sorted factor analysis output with item labels. cor2 correlates two data.frames (of equal length). levels2numeric and char2numeric convert dataframe columns that are categorical/levels to numeric values.

psych.misc()
lowerCor(x,digits=2,use="pairwise",method="pearson")
cor2(x,y,digits=2,use="pairwise",method="pearson")
lowerMat(R, digits = 2)
tableF(x,y)
reflect(f,flip=NULL)
progressBar(value,max,label=NULL)
shannon(x,correct=FALSE,base=2)
test.all(pl,package="psych",dependencies
      = c("Depends", "Imports", "LinkingTo"),find=FALSE,skip=NULL)
 levels2numeric(x)
 char2numeric(x)
 isCorrelation(x) #test if an object is a symmetric matrix

Arguments

R
A rectangular matrix or data frame (probably a correlation matrix)

x
A data matrix or data frame or a vector depending upon the function.

y
A data matrix or data frame or a vector

f
The object returned from either a factor analysis (fa) or a principal components analysis (principal)

digits
round to digits

use
Should pairwise deletion be done, or one of the other options to cor

method
"pearson", "kendall", "spearman"

value
the current value of some looping variable

max
The maximum value the loop will achieve

label
what function is looping

flip
The factor or components to be reversed keyed (by factor number)

correct
Correct for the maximum possible information in this item

base
What is the base for the log function (default=2, e implies base = exp(1))

pl
The name of a package (or list of packages) to be activated and then have all the examples tested.

package
Find the dependencies for this package, e.g., psych

dependencies
Which type of dependency to examine?

find
Look up the dependencies, and then test all of their examples

skip
Do not test these dependencies

Value

tableF is fast alternative to the table function for creating two way tables of numeric variables. It does not have any of the elegant checks of the table function and thus is much faster. Used in the tetrachoric and polychoric functions to maximize speed. lowerCor Finds and prints (using lowerMat) the lower diagonal correlation matrix but returns (invisibly) the full correlation matrix found with the use and method parameters. The default values are for pairwise deletion of variables, and to print to 2 decimal places. lowerMatShows the lower triangle of a matrix, rounded to digits with titles abbreviated to digits + 3 progressBar Display a series of dots as we progress through a slow loop (removed from anything using multicores). tableF (for tableFast) is a cut down version of table that does no error checking, nor returns pretty output, but is significantly faster than table. It will just work on two integer vectors. This is used in polychoric an tetrachoric for about a 50% speed improvement for large problems. shannon finds Shannon's H index of information. Used for estimating the complexity or diversity of the distribution of responses in a vector or matrix. $$H = -\sum{p_i log(p_i) }$$ test.all allows one to test all the examples in specified package. This allows us to make sure that those examples work when other packages (e.g., psych) are also loaded. This is used when developing revisions to the psych package to make sure the the other packages work. Some packages will not work and/or crash the system (e.g., DeducerPlugInScaling requires Java and even with Java, crashes when loaded, even if psych is not there!). Alternatively, if testing a long list of dependencies, you can skip the first part by specifying them by name. cor2 will find and display the correlations between two sets of variables, rounded to digits, using the other options. If x is a list of multiple sets (two or more), then all sets are correlated. levels2numericconverts character data with levels to numeric data. Used in the SAPA analyses where we code some variables, (e.g., gender, education) with character codes to help in the documentation of files, but want to do analyses of correlations with other categorical variables. char2numericconverts character data with levels to numeric data. Used for cases when data from questionnaires include the response categories rathere than numeric data. Unless the levels of the data are in meaningful order, the numeric results are not useful. Most useful if doing polychoric analyses.

Details

lowerCor prints out the lower off diagonal matrix rounded to digits with column names abbreviated to digits + 3 characters, but also returns the full and unrounded matrix. By default, it uses pairwise deletion of variables. It in turn calls lowerMat which does the pretty printing.

It is important to remember to not call lowerCor when all you need is lowerMat!

See also

corr.test to find correlations, count the pairwise occurrences, and to give significance tests for each correlation. r.test for a number of tests of correlations, including tests of the difference between correlations. lowerUpper will display the differences between two matrices.

Examples

lowerMat(Thurstone)

#>                   Sntnc Vcblr Snt.C Frs.L F.L.W Sffxs Ltt.S Pdgrs Ltt.G
#> Sentences         1.00                                                 
#> Vocabulary        0.83  1.00                                           
#> Sent.Completion   0.78  0.78  1.00                                     
#> First.Letters     0.44  0.49  0.46  1.00                               
#> Four.Letter.Words 0.43  0.46  0.42  0.67  1.00                         
#> Suffixes          0.45  0.49  0.44  0.59  0.54  1.00                   
#> Letter.Series     0.45  0.43  0.40  0.38  0.40  0.29  1.00             
#> Pedigrees         0.54  0.54  0.53  0.35  0.37  0.32  0.56  1.00       
#> Letter.Group      0.38  0.36  0.36  0.42  0.45  0.32  0.60  0.45  1.00 

lb <- lowerCor(bfi[1:10])  #finds and prints the lower correlation matrix, 

#>    A1    A2    A3    A4    A5    C1    C2    C3    C4    C5   
#> A1  1.00                                                      
#> A2 -0.34  1.00                                                
#> A3 -0.27  0.49  1.00                                          
#> A4 -0.15  0.34  0.36  1.00                                    
#> A5 -0.18  0.39  0.50  0.31  1.00                              
#> C1  0.03  0.09  0.10  0.09  0.12  1.00                        
#> C2  0.02  0.14  0.14  0.23  0.11  0.43  1.00                  
#> C3 -0.02  0.19  0.13  0.13  0.13  0.31  0.36  1.00            
#> C4  0.13 -0.15 -0.12 -0.15 -0.13 -0.34 -0.38 -0.34  1.00      
#> C5  0.05 -0.12 -0.16 -0.24 -0.17 -0.25 -0.30 -0.34  0.48  1.00

  # returns the square matrix.
#fiml <- corFiml(bfi[1:10])     #FIML correlations require lavaan package
#lowerMat(fiml)  #to get pretty output
f3 <- fa(Thurstone,3)
f3r <- reflect(f3,2)  #reflect the second factor
#find the complexity of the response patterns of the iqitems.
round(shannon(iqitems),2)

#>  reason.4 reason.16 reason.17 reason.19  letter.7 letter.33 letter.34 letter.58 
#>      1.78      1.52      1.62      1.86      1.88      2.01      1.89      2.29 
#> matrix.45 matrix.46 matrix.47 matrix.55  rotate.3  rotate.4  rotate.6  rotate.8 
#>      2.03      2.10      1.92      2.40      2.89      2.76      2.72      2.93 

#test.all('BinNor')  #Does the BinNor package work when we are using other packages
bestItems(lb,3,cut=.1)

#>       A3
#> A3  1.00
#> A5  0.50
#> A2  0.49
#> A4  0.36
#> A1 -0.27
#> C5 -0.16
#> C2  0.14
#> C3  0.13
#> C4 -0.12

#to make this a latex table 
#df2latex(bestItems(lb,2,cut=.2))
#
data(bfi.dictionary)
f2 <- fa(bfi[1:10],2)
fa.lookup(f2,bfi.dictionary)

#>      MR1   MR2  com   h2 ItemLabel                                      Item
#> C4 -0.65  0.00 1.00 0.42     q_626           Do things in a half-way manner.
#> C2  0.64 -0.01 1.00 0.40     q_530     Continue until everything is perfect.
#> C1  0.57 -0.06 1.02 0.31     q_124                   Am exacting in my work.
#> C5 -0.56 -0.06 1.02 0.34    q_1949                            Waste my time.
#> C3  0.54  0.03 1.01 0.31     q_619            Do things according to a plan.
#> A3 -0.03  0.76 1.00 0.56    q_1206               Know how to comfort others.
#> A2  0.01  0.68 1.00 0.46    q_1162         Inquire about others' well-being.
#> A5  0.03  0.60 1.00 0.37    q_1419                 Make people feel at ease.
#> A4  0.14  0.44 1.22 0.25    q_1364                            Love children.
#> A1  0.08 -0.41 1.08 0.15     q_146 Am indifferent to the feelings of others.
#>       Giant3              Big6        Little12 Keying IPIP100
#> C4 Stability Conscientiousness Industriousness     -1    B5:C
#> C2 Stability Conscientiousness     Orderliness      1    B5:C
#> C1 Stability Conscientiousness     Orderliness      1    B5:C
#> C5 Stability Conscientiousness Industriousness     -1    B5:C
#> C3 Stability Conscientiousness     Orderliness      1    B5:C
#> A3  Cohesion     Agreeableness      Compassion      1    B5:A
#> A2  Cohesion     Agreeableness      Compassion      1    B5:A
#> A5  Cohesion     Agreeableness      Compassion      1    B5:A
#> A4  Cohesion     Agreeableness      Compassion      1    B5:A
#> A1  Cohesion     Agreeableness      Compassion     -1    B5:A

sa1 <-sat.act[1:2]
sa2 <- sat.act[3:4]
sa3 <- sat.act[5:6]
cor2(sa1,sa2)

#>             age   ACT
#> gender    -0.02 -0.04
#> education  0.55  0.15

cor2(list(sa1,sa2))  #show within set and between set cors

#>           gendr edctn age   ACT  
#> gender     1.00                  
#> education  0.09  1.00            
#> age       -0.02  0.55  1.00      
#> ACT       -0.04  0.15  0.11  1.00

cor2(list(sa1,sa2,sa3))

#>           gendr edctn age   ACT   SATV  SATQ 
#> gender     1.00                              
#> education  0.09  1.00                        
#> age       -0.02  0.55  1.00                  
#> ACT       -0.04  0.15  0.11  1.00            
#> SATV      -0.02  0.05 -0.04  0.56  1.00      
#> SATQ      -0.17  0.03 -0.03  0.59  0.64  1.00