Find von Neuman's Mean Square of Successive Differences

Von Neuman et al. (1941) discussed the Mean Square of Successive Differences as a measure of variability that takes into account gradual shifts in mean. This is appropriate when studying errors in ballistics or variability and stability in mood when studying affect. For random data, this will be twice the variance, but for data with a sequential order and a positive autocorrelation, this will be much smaller. Since the mssd is just twice the variance - the autocorrelation, it is thus possible to also find the autocorrelation for a particular lag.

mssd(x,group=NULL, lag = 1,na.rm=TRUE)
rmssd(x,group=NULL, lag=1, na.rm=TRUE)
autoR(x,group=NULL,lag=1,na.rm=TRUE)

Arguments

x
a vector, data.frame or matrix

lag
the lag to use when finding diff

group
A column of the x data.frame to be used for grouping

na.rm
Should missing data be removed?

Details

When examining multiple measures within subjects, it is sometimes useful to consider the variability of trial by trial observations in addition to the over all between trial variation. The Mean Square of Successive Differences (mssd) and root mean square of successive differences (rmssd) find the variance or standard deviation of the trial to trial differences. \(\sigma^2 = \Sigma(x_i - x_{i+1})^2 /(n-1) \)

In the case of multiple subjects (groups) with multiple observations per subject (group), specify the grouping variable will produce output for each group.

Similar functions are available in the matrixStats package. This is just the variance and standard deviation applied to the result from the diff function.

Perhaps useful when studying mood, the autoR function finds the autocorrelation for each item for the specified lag. It also returns the rmssd (root means square successive difference).

Value

The average squared successive difference (mssd) and the square root of the average squared successive difference (rmssd). Note that this is not the same as the standard deviation of the lagged differences.

References

Jahng, Seungmin and Wood, Phillip K and Trull, Timothy J. Analysis of affective instability in ecological momentary assessment: Indices using successive difference and group comparison via multilevel modeling. Psychological methods (2008) 13, 354-375.

Von Neumann, J., Kent, R., Bellinson, H., and Hart, B. (1941). The mean square successive difference. The Annals of Mathematical Statistics, pages 153-162.

See also

See Also rmssd for the standard deviation or describe for more conventional statistics. describeBy and statsBy give group level statistics. See also link{mlr}, link{mlreliability}, link{mlPlot} for other ways of examining within person variability over time.

Examples

t <- seq(-pi, pi, .1)
trial <- 1: length(t)
gr <- trial %% 8
c <- cos(t)
ts <- sample(t,length(t))
cs <- cos(ts)
x.df <- data.frame(trial,gr,t,c,ts,cs)
rmssd(x.df)

#>      trial         gr          t          c         ts         cs 
#> 1.00816340 2.55432773 0.10081634 0.07173326 2.61687462 0.97101878 

rmssd(x.df,gr)

#>      trial gr         t         c       ts        cs
#> 0 8.763561  0 0.8763561 0.6692496 3.636207 1.1905379
#> 1 8.640988  0 0.8640988 0.6231704 2.217356 0.9511735
#> 2 8.640988  0 0.8640988 0.6270824 3.093811 1.2830240
#> 3 8.640988  0 0.8640988 0.6297245 2.597755 0.6183497
#> 4 8.640988  0 0.8640988 0.6310078 2.071634 0.9417185
#> 5 8.640988  0 0.8640988 0.6308895 2.132291 1.1630731
#> 6 8.640988  0 0.8640988 0.6293734 1.771534 0.8758153
#> 7 8.640988  0 0.8640988 0.6265103 2.415229 1.3500736

autoR(x.df,gr)

#> 
#> Autocorrelations 
#> Call: autoR(x = x.df, group = gr)
#>   trial  gr    t    c    ts    cs
#> 0  0.86 NaN 0.86 0.55 -0.24  0.25
#> 1  0.89 NaN 0.89 0.69  0.43 -0.05
#> 2  0.89 NaN 0.89 0.68  0.01 -0.24
#> 3  0.89 NaN 0.89 0.67 -0.48  0.54
#> 4  0.89 NaN 0.89 0.67 -0.21 -0.11
#> 5  0.89 NaN 0.89 0.67 -0.16 -0.31
#> 6  0.89 NaN 0.89 0.67  0.44  0.40
#> 7  0.89 NaN 0.89 0.68 -0.01 -0.08

describe(x.df)

#>       vars  n  mean    sd median trimmed   mad   min   max range  skew kurtosis
#> trial    1 63 32.00 18.33  32.00   32.00 23.72  1.00 63.00  62.0  0.00    -1.26
#> gr       2 63  3.56  2.28   4.00    3.57  2.97  0.00  7.00   7.0 -0.01    -1.28
#> t        3 63 -0.04  1.83  -0.04   -0.04  2.37 -3.14  3.06   6.2  0.00    -1.26
#> c        4 63  0.00  0.71   0.01    0.00  1.06 -1.00  1.00   2.0  0.00    -1.55
#> ts       5 63 -0.04  1.83  -0.04   -0.04  2.37 -3.14  3.06   6.2  0.00    -1.26
#> cs       6 63  0.00  0.71   0.01    0.00  1.06 -1.00  1.00   2.0  0.00    -1.55
#>         se
#> trial 2.31
#> gr    0.29
#> t     0.23
#> c     0.09
#> ts    0.23
#> cs    0.09

#pairs.panels(x.df)
#mlPlot(x.df,grp="gr",Time="t",items=c(4:6))