NEO correlation matrix from the NEO_PI_R manual

The NEO.PI.R is a widely used personality test to assess 5 broad factors (Neuroticism, Extraversion, Openness, Agreeableness and Conscientiousness) with six facet scales for each factor. The correlation matrix of the facets is reported in the NEO.PI.R manual for 1000 subjects.

data(neo)

Format

A data frame of a 30 x 30 correlation matrix with the following 30 variables.

N1

Anxiety

N2

AngryHostility

N3

Depression

N4

Self-Consciousness

N5

Impulsiveness

N6

Vulnerability

E1

Warmth

E2

Gregariousness

E3

Assertiveness

E4

Activity

E5

Excitement-Seeking

E6

PositiveEmotions

O1

Fantasy

O2

Aesthetics

O3

Feelings

O4

Ideas

O5

Actions

O6

Values

A1

Trust

A2

Straightforwardness

A3

Altruism

A4

Compliance

A5

Modesty

A6

Tender-Mindedness

C1

Competence

C2

Order

C3

Dutifulness

C4

AchievementStriving

C5

Self-Discipline

C6

Deliberation

Details

The past thirty years of personality research has led to a general consensus on the identification of major dimensions of personality. Variously known as the ``Big 5" or the ``Five Factor Model", the general solution represents 5 broad domains of personal and interpersonal experience. Neuroticism and Extraversion are thought to reflect sensitivity to negative and positive cues from the environment and the tendency to withdraw or approach. Openness is sometimes labeled as Intellect and reflects an interest in new ideas and experiences. Agreeableness and Conscientiousness reflect tendencies to get along with others and to want to get ahead.

The factor structure of the NEO suggests five correlated factors as well as two higher level factors. The NEO was constructed with 6 ``facets" for each of the five broad factors.

Source

Costa, Paul T. and McCrae, Robert R. (1992) (NEO PI-R) professional manual. Psychological Assessment Resources, Inc. Odessa, FL. (with permission of the author and the publisher)

References

Digman, John M. (1990) Personality structure: Emergence of the five-factor model. Annual Review of Psychology. 41, 417-440.

John M. Digman (1997) Higher-order factors of the Big Five. Journal of Personality and Social Psychology, 73, 1246-1256.

McCrae, Robert R. and Costa, Paul T., Jr. (1999) A Five-Factor theory of personality. In Pervin, Lawrence A. and John, Oliver P. (eds) Handbook of personality: Theory and research (2nd ed.) 139-153. Guilford Press, New York. N.Y.

Revelle, William (1995), Personality processes, Annual Review of Psychology, 46, 295-328.

Joshua Wilt and William Revelle (2009) Extraversion and Emotional Reactivity. In Mark Leary and Rick H. Hoyle (eds). Handbook of Individual Differences in Social Behavior. Guilford Press, New York, N.Y.

Examples

data(neo)
n5 <- fa(neo,5)
neo.keys <- make.keys(30,list(N=c(1:6),E=c(7:12),O=c(13:18),A=c(19:24),C=c(25:30)))
n5p <- target.rot(n5,neo.keys) #show a targeted rotation for simple structure
n5p

#> 
#> Call: NULL
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>      MR1   MR4   MR3   MR2   MR5   h2   u2
#> N1  0.79  0.00  0.03  0.08  0.07 0.62 0.38
#> N2  0.61  0.02  0.02 -0.40  0.07 0.54 0.46
#> N3  0.77 -0.12  0.09  0.05 -0.06 0.63 0.37
#> N4  0.67 -0.18 -0.01  0.09  0.01 0.50 0.50
#> N5  0.43  0.33 -0.01 -0.08 -0.27 0.37 0.63
#> N6  0.62 -0.15 -0.01  0.10 -0.22 0.49 0.51
#> E1 -0.03  0.57  0.08  0.46 -0.02 0.52 0.48
#> E2 -0.11  0.56 -0.03  0.16 -0.14 0.35 0.65
#> E3 -0.22  0.43  0.11 -0.26  0.19 0.39 0.61
#> E4  0.16  0.51  0.05 -0.16  0.34 0.43 0.57
#> E5  0.04  0.55  0.01 -0.24 -0.13 0.39 0.61
#> E6  0.04  0.67  0.06  0.21 -0.04 0.49 0.51
#> O1  0.10  0.10  0.51 -0.09 -0.29 0.38 0.62
#> O2  0.13 -0.10  0.69  0.15  0.14 0.52 0.48
#> O3  0.39  0.30  0.43  0.08  0.11 0.42 0.58
#> O4 -0.17  0.14  0.43  0.04 -0.09 0.26 0.74
#> O5 -0.15 -0.12  0.70 -0.12  0.12 0.55 0.45
#> O6 -0.13  0.06  0.34 -0.06 -0.15 0.17 0.83
#> A1 -0.30  0.13  0.11  0.50 -0.07 0.37 0.63
#> A2  0.02 -0.21 -0.05  0.58  0.24 0.46 0.54
#> A3  0.06  0.45 -0.12  0.61  0.16 0.57 0.43
#> A4 -0.15 -0.18  0.05  0.69 -0.02 0.55 0.45
#> A5  0.15 -0.17 -0.09  0.51 -0.04 0.34 0.66
#> A6  0.06  0.14  0.11  0.58 -0.04 0.37 0.63
#> C1 -0.26  0.13  0.07  0.00  0.55 0.42 0.58
#> C2  0.09  0.07 -0.18  0.00  0.64 0.46 0.54
#> C3 -0.07 -0.08  0.02  0.22  0.63 0.47 0.53
#> C4  0.09  0.19  0.10 -0.12  0.71 0.57 0.43
#> C5 -0.14  0.16 -0.13  0.02  0.69 0.56 0.44
#> C6 -0.14 -0.27 -0.01  0.12  0.53 0.40 0.60
#> 
#>                        MR1  MR4  MR3  MR2  MR5
#> SS loadings           3.22 2.68 1.88 2.79 2.97
#> Proportion Var        0.11 0.09 0.06 0.09 0.10
#> Cumulative Var        0.11 0.20 0.26 0.35 0.45
#> Proportion Explained  0.24 0.20 0.14 0.21 0.22
#> Cumulative Proportion 0.24 0.44 0.57 0.78 1.00
#>       MR1   MR4   MR3   MR2   MR5
#> MR1  1.00 -0.05 -0.04 -0.02 -0.08
#> MR4 -0.05  1.00  0.09 -0.07  0.03
#> MR3 -0.04  0.09  1.00  0.00 -0.03
#> MR2 -0.02 -0.07  0.00  1.00  0.01
#> MR5 -0.08  0.03 -0.03  0.01  1.00