Test the difference between (un)paired correlations

Test the difference between two (paired or unpaired) correlations. Given 3 variables, x, y, z, is the correlation between xy different than that between xz? If y and z are independent, this is a simple t-test of the z transformed rs. But, if they are dependent, it is a bit more complicated.

paired.r(xy, xz, yz=NULL, n, n2=NULL,twotailed=TRUE)

Arguments

xy
r(xy)

xz
r(xz)

yz
r(yz)

n
Number of subjects for first group

n2
Number of subjects in second group (if not equal to n)

twotailed
Calculate two or one tailed probability values

Details

To find the z of the difference between two independent correlations, first convert them to z scores using the Fisher r-z transform and then find the z of the difference between the two correlations. The default assumption is that the group sizes are the same, but the test can be done for different size groups by specifying n2.

If the correlations are not independent (i.e., they are from the same sample) then the correlation with the third variable r(yz) must be specified. Find a t statistic for the difference of thee two dependent correlations.

Value

a list containing the calculated t or z values and the associated two (or one) tailed probability.

t

t test of the difference between two dependent correlations

p

probability of the t or of the z

z

z test of the difference between two independent correlations

See also

r.test for more tests of independent as well as dependent (paired) tests. p.rep.r for the probability of replicating a particular correlation. cor.test from stats for testing a single correlation and corr.test for finding the values and probabilities of multiple correlations. See also set.cor to do multiple correlations from matrix input.

Examples

paired.r(.5,.3, .4, 100) #dependent correlations

#> Call: paired.r(xy = 0.5, xz = 0.3, yz = 0.4, n = 100)
#> [1] "test of difference between two correlated  correlations"
#> t = 2.06  With probability =  0.04

paired.r(.5,.3,NULL,100) #independent correlations same sample size

#> Call: paired.r(xy = 0.5, xz = 0.3, yz = NULL, n = 100)
#> [1] "test of difference between two independent correlations"
#> z = 1.67  With probability =  0.09

paired.r(.5,.3,NULL, 100, 64) # independent correlations, different sample sizes

#> Call: paired.r(xy = 0.5, xz = 0.3, yz = NULL, n = 100, n2 = 64)
#> [1] "test of difference between two independent correlations"
#> z = 1.47  With probability =  0.14