Print and summary functions for the psych class

Give limited output (print) or somewhat more detailed (summary) for most of the functions in psych.

# S3 method for psych
print(x,digits=2,all=FALSE,cut=NULL,sort=FALSE,short=TRUE,lower=TRUE,signif=NULL,...)

# S3 method for psych
summary(object,digits=2,items=FALSE,...)
# S3 method for psych
anova(object,object2,...)

Arguments

x
Output from a psych function (e.g., factor.pa, omega,ICLUST, score.items, cluster.cor

object
Output from a psych function

items
items=TRUE (default) does not print the item whole correlations

digits
Number of digits to use in printing

all
if all=TRUE, then the object is declassed and all output from the function is printed

cut
Cluster loadings < cut will not be printed. For the factor analysis functions (fa and factor.pa etc.), cut defaults to 0, for ICLUST to .3, for omega to .2.

sort
Cluster loadings are in sorted order

short
Controls how much to print

lower
For square matrices, just print the lower half of the matrix

object2
Another object from fa or omega

signif
If not NULL, a numeric value, show just signif number of leading digits for describe output

...
More options to pass to summary and print

Details

Most of the psych functions produce too much output. print.psych and summary.psych use generic methods for printing just the highlights. To see what else is available, ask for the structure of the particular object: (str(theobject) ).

Alternatively, to get complete output, unclass(theobject) and then print it. This may be done by using the all=TRUE option.

As an added feature, if the promax function is applied to a factanal loadings matrix, the normal output just provides the rotation matrix. print.psych will provide the factor correlations. (Following a suggestion by John Fox and Uli Keller to the R-help list). The alternative is to just use the Promax function directly on the factanal object.

Value

Various psych functions produce copious output. This is a way to summarize the most important parts of the output of the score.items, cluster.scores, and ICLUST functions. See those ( score.items, cluster.cor, cluster.loadings, or ICLUST) for details on what is produced.

The signf option is available for the output from describe to adjust the number of digits shown for all columns. This is slightly different from what happens if you specify digits, which rounds all output to the number of digits. print(x,signif=3) will print just the 3 largest digits of x, which will frequently result in scientific notation for any column where that would be appropriate for at least one row.

Note

See score.items, cluster.cor, cluster.loadings, or ICLUSTfor details on what is printed.

Examples

data(bfi)
 keys.list <- list(agree=c(-1,2:5),conscientious=c(6:8,-9,-10),
 extraversion=c(-11,-12,13:15),neuroticism=c(16:20),openness = c(21,-22,23,24,-25))
 keys <- make.keys(25,keys.list,item.labels=colnames(bfi[1:25]))
 scores <- score.items(keys,bfi[1:25])

#> score.items has been replaced by scoreItems, please change your call

 scores

#> Call: scoreItems(keys = keys, items = items, totals = totals, ilabels = ilabels, 
#>     missing = missing, impute = impute, delete = delete, min = min, 
#>     max = max, digits = digits, select = select)
#> 
#> (Unstandardized) Alpha:
#>       agree conscientious extraversion neuroticism openness
#> alpha   0.7          0.72         0.76        0.81      0.6
#> 
#> Standard errors of unstandardized Alpha:
#>       agree conscientious extraversion neuroticism openness
#> ASE   0.014         0.014        0.013       0.011    0.017
#> 
#> Average item correlation:
#>           agree conscientious extraversion neuroticism openness
#> average.r  0.32          0.34         0.39        0.46     0.23
#> 
#>  Guttman 6* reliability: 
#>          agree conscientious extraversion neuroticism openness
#> Lambda.6   0.7          0.72         0.76        0.81      0.6
#> 
#> Signal/Noise based upon av.r : 
#>              agree conscientious extraversion neuroticism openness
#> Signal/Noise   2.3           2.6          3.2         4.3      1.5
#> 
#> Scale intercorrelations corrected for attenuation 
#>  raw correlations below the diagonal, alpha on the diagonal 
#>  corrected correlations above the diagonal:
#>               agree conscientious extraversion neuroticism openness
#> agree          0.70          0.36         0.63      -0.245     0.23
#> conscientious  0.26          0.72         0.35      -0.305     0.30
#> extraversion   0.46          0.26         0.76      -0.284     0.32
#> neuroticism   -0.18         -0.23        -0.22       0.812    -0.12
#> openness       0.15          0.19         0.22      -0.086     0.60
#> 
#>  In order to see the item by scale loadings and frequency counts of the data
#>  print with the short option = FALSE

 summary(scores)

#> Call: scoreItems(keys = keys, items = items, totals = totals, ilabels = ilabels, 
#>     missing = missing, impute = impute, delete = delete, min = min, 
#>     max = max, digits = digits, select = select)
#> 
#> Scale intercorrelations corrected for attenuation 
#>  raw correlations below the diagonal, (unstandardized) alpha on the diagonal 
#>  corrected correlations above the diagonal:
#>               agree conscientious extraversion neuroticism openness
#> agree          0.70          0.36         0.63      -0.245     0.23
#> conscientious  0.26          0.72         0.35      -0.305     0.30
#> extraversion   0.46          0.26         0.76      -0.284     0.32
#> neuroticism   -0.18         -0.23        -0.22       0.812    -0.12
#> openness       0.15          0.19         0.22      -0.086     0.60