residuals.psych {psych}	R Documentation

Extract residuals from various psych objects

Description

Residuals in the various psych functions are extracted and then may be "pretty" printed.

Usage

## S3 method for class 'psych'
residuals(object,diag=TRUE,...)
## S3 method for class 'psych'
resid(object,diag=TRUE,...)

Arguments

object
The object returned by a psych function.

diag
if FALSE, then convert the diagonal of the residuals to NA

...
Other parameters to be passed to residual (ignored but required by the generic function)

Details

Currently implemented for fa, principal, omega, irt.fa, and fa.extension.

Value

residuals: a matrix of residual estimates

Author(s)

William Revelle

Examples

f3 <- fa(Thurstone,3)
residuals(f3)
sum(residuals(f3)^2) #include diagonal
sum(residuals(f3,diag=FALSE)^2,na.rm=TRUE) #drop diagonal

[Package psych version 1.9.12 Index]