scaling.fits {psych}	R Documentation

Test the adequacy of simple choice, logistic, or Thurstonian scaling.

Description

Given a matrix of choices and a vector of scale values, how well do the scale values capture the choices? That is, what is size of the squared residuals given the model versus the size of the squared choice values?

Usage

scaling.fits(model, data, test = "logit", digits = 2, rowwise = TRUE)

Arguments

model
A vector of scale values

data
A matrix or dataframe of choice frequencies

test
"choice", "logistic", "normal"

digits
Precision of answer

rowwise
Are the choices ordered by column over row (TRUE) or row over column False)

Details

How well does a model fit the data is the classic problem of all of statistics. One fit statistic for scaling is the just the size of the residual matrix compared to the original estimates.

Value

GF
Goodness of fit of the model

original
Sum of squares for original data

resid
Sum of squares for residuals given the data and the model

residual
Residual matrix

Note

Mainly for demonstration purposes for a course on psychometrics

Author(s)

William Revelle

References

Revelle, W. (in preparation) Introduction to psychometric theory with applications in R, Springer. https://personality-project.org/r/book

See Also

thurstone, vegetables

[Package psych version 1.9.12 Index]