Score item composite scales and find Cronbach's alpha, Guttman lambda 6 and item whole correlations

Given a list of items and the direction to score them (a keys.list) or a matrix or data.frame of k keys for n items (-1, 0, 1), and a matrix or data.frame of items scores for m items and N people, find the sum scores or average scores for each person and each scale. In addition, report Cronbach's alpha, Guttman's Lambda 6, the average r, the scale intercorrelations, and the item by scale correlations (raw and corrected for item overlap). Replace missing values with the item median or mean if desired. Will adjust scores for reverse scored items. Although prior versions used a keys matrix, it is now possible to just use a list of scoring keys. See make.keys for a convenient way to make the keys file. If the input is a square matrix, then it is assumed that the input is a covariance or correlation matix and scores are not found, but the item statistics are reported. (Similar functionality to cluster.cor). response.frequencies reports the frequency of item endorsements fore each response category for polytomous or multiple choice items.

scoreItems(keys, items, totals = FALSE, ilabels = NULL,missing=TRUE, impute="median",
 delete=TRUE, min = NULL, max = NULL, digits = 2,n.obs=NULL,select=TRUE)
 score.items(keys, items, totals = FALSE, ilabels = NULL,missing=TRUE, impute="median",
 delete=TRUE, min = NULL, max = NULL, digits = 2,select=TRUE)
 scoreFast(keys, items, totals = FALSE, ilabels = NULL,missing=TRUE, impute="none",
 delete=TRUE, min = NULL, max = NULL, digits = 2)
  response.frequencies(items,max=10,uniqueitems=NULL)

Arguments

keys
A list of scoring keys or a matrix or dataframe of -1, 0, or 1 weights for each item on each scale which may be created by hand, or by using make.keys. Just using a list of scoring keys (see example) is probably more convenient.

items
Matrix or dataframe of raw item scores

totals
if TRUE find total scores, if FALSE (default), find average scores

ilabels
a vector of item labels.

missing
missing = TRUE is the normal case and data are imputed according to the impute option. missing=FALSE, only complete cases are scored.

impute
impute="median" replaces missing values with the item median, impute = "mean" replaces values with the mean response. impute="none" the subject's scores are based upon the average of the keyed, but non missing scores. impute = "none" is probably more appropriate for a large number of missing cases (e.g., SAPA data).

delete
if delete=TRUE, automatically delete items with no variance (and issue a warning)

min
May be specified as minimum item score allowed, else will be calculated from data. min and max should be specified if items differ in their possible minima or maxima. See notes for details.

max
May be specified as maximum item score allowed, else will be calculated from data. Alternatively, in response frequencies, it is maximum number of alternative responses to count.

uniqueitems
If specified, the set of possible unique response categories

digits
Number of digits to report

n.obs
If scoring from a correlation matrix, specify the number of subjects allows for the calculation of the confidence intervals for alpha.

select
By default, just find the statistics of those items that are included in scoring keys. This allows scoring of data sets that have bad data for some items that are not included in the scoring keys.

Details

The process of finding sum or average scores for a set of scales given a larger set of items is a typical problem in applied psychometrics and in psychometric research. Although the structure of scales can be determined from the item intercorrelations, to find scale means, variances, and do further analyses, it is typical to find scores based upon the sum or the average item score. For some strange reason, personality scale scores are typically given as totals, but attitude scores as averages. The default for scoreItems is the average as it would seem to make more sense to report scale scores in the metric of the item.

When scoring more than one scale, it is convenient to have list of the items on each scale and the direction to score the items. This may be converted to a keys.matrix using make.keys or may be entered as a keys.list directly.

Various estimates of scale reliability include ``Cronbach's alpha", Guttman's Lambda 6, and the average interitem correlation. For k = number of items in a scale, and av.r = average correlation between items in the scale, alpha = k * av.r/(1+ (k-1)*av.r). Thus, alpha is an increasing function of test length as well as the test homeogeneity.

Surprisingly, more than a century after Spearman (1904) introduced the concept of reliability to psychologists, there are still multiple approaches for measuring it. Although very popular, Cronbach's \(\alpha\) (1951) underestimates the reliability of a test and over estimates the first factor saturation. \(\alpha\) (Cronbach, 1951) is the same as Guttman's \(\lambda_3\) (Guttman, 1945) and may be found by $$ \lambda_3 = \frac{n}{n-1}\Bigl(1 - \frac{tr(\vec{V})_x}{V_x}\Bigr) = \frac{n}{n-1} \frac{V_x - tr(\vec{V}_x)}{V_x} = \alpha $$

Perhaps because it is so easy to calculate and is available in most commercial programs, alpha is without doubt the most frequently reported measure of internal consistency reliability. Alpha is the mean of all possible spit half reliabilities (corrected for test length). For a unifactorial test, it is a reasonable estimate of the first factor saturation, although if the test has any microstructure (i.e., if it is ``lumpy") coefficients \(\beta\) (Revelle, 1979; see ICLUST) and \(\omega_h\) (see omega) (McDonald, 1999; Revelle and Zinbarg, 2009) are more appropriate estimates of the general factor saturation. \(\omega_t\) (see omega) is a better estimate of the reliability of the total test.

Guttman's Lambda 6 (G6) considers the amount of variance in each item that can be accounted for the linear regression of all of the other items (the squared multiple correlation or smc), or more precisely, the variance of the errors, \(e_j^2\), and is $$ \lambda_6 = 1 - \frac{\sum e_j^2}{V_x} = 1 - \frac{\sum(1-r_{smc}^2)}{V_x} .$$

The squared multiple correlation is a lower bound for the item communality and as the number of items increases, becomes a better estimate.

G6 is also sensitive to lumpyness in the test and should not be taken as a measure of unifactorial structure. For lumpy tests, it will be greater than alpha. For tests with equal item loadings, alpha > G6, but if the loadings are unequal or if there is a general factor, G6 > alpha. Although it is normal when scoring just a single scale to calculate G6 from just those items within the scale, logically it is appropriate to estimate an item reliability from all items available. This is done here and is labeled as G6* to identify the subtle difference.

Alpha and G6* are both positive functions of the number of items in a test as well as the average intercorrelation of the items in the test. When calculated from the item variances and total test variance, as is done here, raw alpha is sensitive to differences in the item variances. Standardized alpha is based upon the correlations rather than the covariances. alpha is a generalization of an earlier estimate of reliability for tests with dichotomous items developed by Kuder and Richardson, known as KR20, and a shortcut approximation, KR21. (See Revelle, in prep; Revelle and Condon, in press.).

A useful index is the ratio of reliable variance to unreliable variance and is known as the Signal/Noise ratio. This is just $$s/n = \frac{n \bar{r}}{1-n \bar{r}}$$ (Cronbach and Gleser, 1964; Revelle and Condon (in press)).

Standard errors for unstandardized alpha are reported using the formula from Duhachek and Iacobucci (2005).

More complete reliability analyses of a single scale can be done using the omega function which finds \(\omega_h\) and \(\omega_t\) based upon a hierarchical factor analysis. Alternative estimates of the Greatest Lower Bound for the reliability are found in the guttman function.

Alpha is a poor estimate of the general factor saturation of a test (see Revelle and Zinbarg, 2009; Zinbarg et al., 2005) for it can seriously overestimate the size of a general factor, and a better but not perfect estimate of total test reliability because it underestimates total reliability. None the less, it is a common statistic to report. In general, the use of alpha should be discouraged and the use of more appropriate estimates (\(\omega_h\) and \(\omega_t\)) should be encouraged.

Correlations between scales are attenuated by a lack of reliability. Correcting correlations for reliability (by dividing by the square roots of the reliabilities of each scale) sometimes help show structure.

There are several alternative ways to treat missing values. By default, missing values are replaced with the corresponding median value for that item. Means can be used instead (impute="mean"), or subjects with missing data can just be dropped (missing = FALSE). For data with a great deal of missingness, yet another option is to just find the average of the available responses (impute="none"). This is useful for findings means for scales for the SAPA project (see https://sapa-project.org) where most scales are estimated from random sub samples of the items from the scale. In this case, the alpha reliabilities are seriously overinflated because they are based upon the total number of items in each scale. The "alpha observed" values are based upon the average number of items answered in each scale using the standard form for alpha a function of inter-item correlation and number of items. The number of missing responses for each person for each scale is reported in the missing object. One possibility is to drop scores just for those scales with missing responses. This may be done adding the code:

scores$scores[scores$missing >0] <- NA

This is shown in the last example.

scoreItems can be applied to correlation matrices to find just the reliability statistics. This will be done automatically if the items matrix is square and none of the values in the matrix are less than -1 or greater than 1. scoreFast just finds the scores (with or without imputation) and does not report other statistics. It is much faster!

Value

scores

Sum or average scores for each subject on the k scales

alpha

Cronbach's coefficient alpha. A simple (but non-optimal) measure of the internal consistency of a test. See also beta and omega. Set to 1 for scales of length 1.

av.r

The average correlation within a scale, also known as alpha 1, is a useful index of the internal consistency of a domain. Set to 1 for scales with 1 item.

G6

Guttman's Lambda 6 measure of reliability

G6*

A generalization of Guttman's Lambda 6 measure of reliability using all the items to find the smc.

n.items

Number of items on each scale

item.cor

The correlation of each item with each scale. Because this is not corrected for item overlap, it will overestimate the amount that an item correlates with the other items in a scale.

cor

The intercorrelation of all the scales based upon the interitem correlations (see note for why these differ from the correlations of the observed scales themselves).

corrected

The correlations of all scales (below the diagonal), alpha on the diagonal, and the unattenuated correlations (above the diagonal)

item.corrected

The item by scale correlations for each item, corrected for item overlap by replacing the item variance with the smc for that item

response.freq

The response frequency (based upon number of non-missing responses) for each alternative.

missing

How many items were not answered for each scale

num.ob.item

The average number of items with responses on a scale. Used in calculating the alpha.observed-- relevant for SAPA type data structures.

Note

It is important to recognize in the case of massively missing data (e.g., data from a Synthetic Aperture Personality Assessment (https://sapa-project.org) study where perhaps only 10-50% of the items per scale are given to any one subject)) that the number of items per scale, and hence standardized alpha, is not the nominal value and hence alpha of the observed scales will be overestimated. For this case (impute="none"), an additional alpha (alpha.ob) is reported.

More importantly in this case of massively missing data, there is a difference between the correlations of the composite scales based upon the correlations of the items and the correlations of the scored scales based upon the observed data. That is, the cor object will have correlations as if all items had been given, while the correlation of the scores object will reflect the actual correlation of the scores. For SAPA data, it is recommended to use the cor object. Confidence of these correlations may be found using the cor.ci function.

Further note that the inter-scale correlations are based upon the correlations of scales formed from the covariance matrix of the items. This will differ from the correlation of scales based upon the correlation of the items. Thus, although scoreItems will produce reliabilities and intercorrelations from either the raw data or from a correlation matrix, these values will differ slightly. In addition, with a great deal of missing data, the scale intercorrelations will differ from the correlations of the scores produced, for the latter will be attenuated.

An alternative to classical test theory scoring is to use score.irt to find score estimates based upon Item Response Theory. This is particularly useful in the case of SAPA data which tend to be massively missing. It is also useful to find scores based upon polytomous items following a factor analysis of the polychoric correlation matrix (see irt.fa).

When reverse scoring items from a set where items differ in their possible minima or maxima, it is important to specify the min and max values. Items are reversed by subtracting them from max + min. Thus, if items range from 1 to 6, items are reversed by subtracting them from 7. But, if the data set includes other variables, (say an id field) that far exceeds the item min or max, then the max id will incorrectly be used to reverse key. min and max can either be single values, or vectors for all items. Compare two examples of scoring the bfi data set.

References

Cronbach, L.J. and Gleser G.C. (1964)The signal/noise ratio in the comparison of reliability coefficients. Educational and Psychological Measurement, 24 (3) 467-480.

Duhachek, A. and Iacobucci, D. (2004). Alpha's standard error (ase): An accurate and precise confidence interval estimate. Journal of Applied Psychology, 89(5):792-808.

McDonald, R. P. (1999). Test theory: A unified treatment. L. Erlbaum Associates, Mahwah, N.J.

Revelle, W. (in preparation) An introduction to psychometric theory with applications in R. http://personality-project.org/r/book

Revelle, W. and Condon, D.C. Reliability. In Irwing, P., Booth, T. and Hughes, D. (Eds). the Wiley-Blackwell Handbook of Psychometric Testing (in press).

Revelle W. and R.E. Zinbarg. (2009) Coefficients alpha, beta, omega and the glb: comments on Sijtsma. Psychometrika, 74(1):145-154.

Zinbarg, R. E., Revelle, W., Yovel, I. and Li, W. (2005) Cronbach's alpha, Revelle's beta, and McDonald's omega h, Their relations with each other and two alternative conceptualizations of reliability, Psychometrika, 70, 123-133.

See also

make.keys for a convenient way to create the keys file, score.multiple.choice for multiple choice items, alpha, correct.cor, cluster.cor , cluster.loadings, omega, guttman for item/scale analysis.

If scales are formed from overlapping sets of items, their correlations will be inflated. This is corrected for when using the scoreOverlap function which, although it will not produce scores, will report scale intercorrelations corrected for item overlap.

In addition, the irt.fa function provides an alternative way of examining the structure of a test and emphasizes item response theory approaches to the information returned by each item and the total test. Associated with these IRT parameters is the score.irt function for finding IRT based scores as well as irt.responses to show response curves for the alternatives in a multiple choice test. score.irt will find both IRT based estimates as well as average item response scores. These latter correlate perfectly with those found by scoreItems. If using a keys matrix, the score.irt results are based upon the item difficulties with the assumption that all items are equally discriminating (effectively a Rasch model). These scores are probably most useful in the case of massively missing data because they can take into account the item difficulties.

Examples

#see  the example including the bfi data set
data(bfi)
keys.list <- list(agree=c("-A1","A2","A3","A4","A5"),
  conscientious=c("C1","C2","C3","-C4","-C5"),extraversion=c("-E1","-E2","E3","E4","E5"),
  neuroticism=c("N1","N2","N3","N4","N5"), openness = c("O1","-O2","O3","O4","-O5"))
  keys <- make.keys(bfi,keys.list)  #no longer necessary
 scores <- scoreItems(keys,bfi,min=1,max=6)  #using a keys matrix 
 scores <- scoreItems(keys.list,bfi,min=1,max=6)  # or just use the keys.list
 summary(scores)

#> Call: scoreItems(keys = keys.list, items = bfi, min = 1, max = 6)
#> 
#> Scale intercorrelations corrected for attenuation 
#>  raw correlations below the diagonal, (unstandardized) alpha on the diagonal 
#>  corrected correlations above the diagonal:
#>               agree conscientious extraversion neuroticism openness
#> agree          0.70          0.36         0.63      -0.245     0.23
#> conscientious  0.26          0.72         0.35      -0.305     0.30
#> extraversion   0.46          0.26         0.76      -0.284     0.32
#> neuroticism   -0.18         -0.23        -0.22       0.812    -0.12
#> openness       0.15          0.19         0.22      -0.086     0.60

 #to get the response frequencies, we need to not use the age variable
 scores <- scoreItems(keys[1:25,],bfi[1:25]) #we do not need to specify min or max if
           #there are no values (such as age) outside the normal item range.
 scores

#> Call: scoreItems(keys = keys[1:25, ], items = bfi[1:25])
#> 
#> (Unstandardized) Alpha:
#>       agree conscientious extraversion neuroticism openness
#> alpha   0.7          0.72         0.76        0.81      0.6
#> 
#> Standard errors of unstandardized Alpha:
#>       agree conscientious extraversion neuroticism openness
#> ASE   0.014         0.014        0.013       0.011    0.017
#> 
#> Average item correlation:
#>           agree conscientious extraversion neuroticism openness
#> average.r  0.32          0.34         0.39        0.46     0.23
#> 
#>  Guttman 6* reliability: 
#>          agree conscientious extraversion neuroticism openness
#> Lambda.6   0.7          0.72         0.76        0.81      0.6
#> 
#> Signal/Noise based upon av.r : 
#>              agree conscientious extraversion neuroticism openness
#> Signal/Noise   2.3           2.6          3.2         4.3      1.5
#> 
#> Scale intercorrelations corrected for attenuation 
#>  raw correlations below the diagonal, alpha on the diagonal 
#>  corrected correlations above the diagonal:
#>               agree conscientious extraversion neuroticism openness
#> agree          0.70          0.36         0.63      -0.245     0.23
#> conscientious  0.26          0.72         0.35      -0.305     0.30
#> extraversion   0.46          0.26         0.76      -0.284     0.32
#> neuroticism   -0.18         -0.23        -0.22       0.812    -0.12
#> openness       0.15          0.19         0.22      -0.086     0.60
#> 
#>  In order to see the item by scale loadings and frequency counts of the data
#>  print with the short option = FALSE

 #The scores themselves are available in the scores$scores object.  I.e.,
 describe(scores$scores)

#>               vars    n mean   sd median trimmed  mad min max range  skew
#> agree            1 2800 4.65 0.89    4.8    4.73 0.89 1.0   6   5.0 -0.77
#> conscientious    2 2800 4.27 0.95    4.4    4.31 0.89 1.0   6   5.0 -0.41
#> extraversion     3 2800 4.15 1.05    4.2    4.20 1.19 1.0   6   5.0 -0.48
#> neuroticism      4 2800 3.16 1.19    3.0    3.13 1.19 1.0   6   5.0  0.22
#> openness         5 2800 4.59 0.80    4.6    4.62 0.89 1.2   6   4.8 -0.34
#>               kurtosis   se
#> agree             0.44 0.02
#> conscientious    -0.16 0.02
#> extraversion     -0.19 0.02
#> neuroticism      -0.65 0.02
#> openness         -0.28 0.02

 #compare this output to that for the impute="none" option for SAPA type data
 #first make many of the items missing in a missing pattern way
 missing.bfi <- bfi
 missing.bfi[1:1000,3:8] <- NA
 missing.bfi[1001:2000,c(1:2,9:10)] <- NA
 scores <- scoreItems(keys.list,missing.bfi,impute="none",min=1,max=6)
 scores

#> Call: scoreItems(keys = keys.list, items = missing.bfi, impute = "none", 
#>     min = 1, max = 6)
#> 
#> (Standardized) Alpha:
#>       agree conscientious extraversion neuroticism openness
#> alpha  0.72          0.72         0.76        0.81      0.6
#> 
#> Standard errors of unstandardized Alpha:
#>       agree conscientious extraversion neuroticism openness
#> ASE   0.014         0.014        0.013       0.011    0.017
#> 
#> Standardized Alpha of observed scales:
#>      agree conscientious extraversion neuroticism openness
#> [1,]  0.62          0.62         0.76        0.81      0.6
#> 
#> Average item correlation:
#>           agree conscientious extraversion neuroticism openness
#> average.r  0.34          0.34         0.39        0.47     0.23
#> 
#>  Guttman 6* reliability: 
#>          agree conscientious extraversion neuroticism openness
#> Lambda.6  0.73          0.72         0.77        0.81     0.61
#> 
#> Signal/Noise based upon av.r : 
#>              agree conscientious extraversion neuroticism openness
#> Signal/Noise   2.5           2.6          3.2         4.4      1.5
#> 
#> Scale intercorrelations corrected for attenuation 
#>  raw correlations below the diagonal, alpha on the diagonal 
#>  corrected correlations above the diagonal:
#> 
#> Note that these are the correlations of the complete scales based on the correlation matrix,
#>  not the observed scales based on the raw items.
#>               agree conscientious extraversion neuroticism openness
#> agree          0.72          0.35         0.64      -0.243     0.25
#> conscientious  0.25          0.72         0.33      -0.275     0.28
#> extraversion   0.47          0.24         0.76      -0.284     0.32
#> neuroticism   -0.19         -0.21        -0.22       0.814    -0.12
#> openness       0.17          0.19         0.22      -0.086     0.60
#> 
#>  In order to see the item by scale loadings and frequency counts of the data
#>  print with the short option = FALSE

 describe(scores$scores)  #the actual scores themselves

#>               vars    n mean   sd median trimmed  mad min max range  skew
#> agree            1 2800 4.67 1.00   5.00    4.77 0.99 1.0   6   5.0 -0.82
#> conscientious    2 2800 4.27 1.09   4.33    4.33 1.09 1.0   6   5.0 -0.47
#> extraversion     3 2800 4.15 1.06   4.20    4.20 1.19 1.0   6   5.0 -0.48
#> neuroticism      4 2800 3.16 1.20   3.00    3.13 1.48 1.0   6   5.0  0.21
#> openness         5 2800 4.59 0.81   4.60    4.62 0.89 1.2   6   4.8 -0.34
#>               kurtosis   se
#> agree             0.42 0.02
#> conscientious    -0.24 0.02
#> extraversion     -0.21 0.02
#> neuroticism      -0.67 0.02
#> openness         -0.29 0.02

 #If we want to delete scales scores for people who did not answer some items for one 
 #(or more) scales, we can do the following:

  scores <- scoreItems(keys.list,missing.bfi,totals=TRUE,min=1,max=6) #find total scores
  describe(scores$scores) #note that missing data were replaced with median for the item

#>               vars    n  mean   sd median trimmed  mad min max range  skew
#> agree            1 2800 23.89 3.34     25   24.24 2.97   8  30    22 -1.14
#> conscientious    2 2800 22.37 3.48     23   22.60 2.97   6  30    24 -0.78
#> extraversion     3 2800 20.73 5.27     21   21.02 5.93   5  30    25 -0.48
#> neuroticism      4 2800 15.81 5.93     15   15.64 5.93   5  30    25  0.22
#> openness         5 2800 22.95 4.02     23   23.10 4.45   6  30    24 -0.34
#>               kurtosis   se
#> agree             1.93 0.06
#> conscientious     1.22 0.07
#> extraversion     -0.19 0.10
#> neuroticism      -0.65 0.11
#> openness         -0.28 0.08

  scores$scores[scores$missing > 0] <- NA  #get rid of cases with missing data
  describe(scores$scores)

#>               vars    n  mean   sd median trimmed  mad min max range  skew
#> agree            1  776 23.17 4.56     24   23.55 4.45   8  30    22 -0.75
#> conscientious    2  769 21.52 4.72     22   21.76 4.45   6  30    24 -0.49
#> extraversion     3 2713 20.72 5.30     21   21.01 5.93   5  30    25 -0.47
#> neuroticism      4 2694 15.82 5.97     15   15.65 7.41   5  30    25  0.22
#> openness         5 2726 22.97 4.04     23   23.12 4.45   6  30    24 -0.35
#>               kurtosis   se
#> agree             0.31 0.16
#> conscientious     0.01 0.17
#> extraversion     -0.20 0.10
#> neuroticism      -0.66 0.12
#> openness         -0.28 0.08