small.msq {psych}	R Documentation

A small example data set taken from a larger data set

Description

A small subset of 200 cases and 14 variables from the msqR data set in the psychTools package. Saved here to allow vignettes to work without using psychTools

Usage

data("small.msq")

Format

A data frame with 200 observations on the following 14 variables.

active

a numeric vector

alert

a numeric vector

aroused

a numeric vector

sleepy

a numeric vector

tired

a numeric vector

drowsy

a numeric vector

anxious

a numeric vector

jittery

a numeric vector

nervous

a numeric vector

calm

a numeric vector

relaxed

a numeric vector

at.ease

a numeric vector

gender

a numeric vector

drug

a numeric vector

Details

See the detailed discussion of the msqR in the psychTools package

References

Revelle, W. and Anderson, K.J. (1998) Personality, motivation and cognitive performance: Final report to the Army Research Institute on contract MDA 903-93-K-0008. (https://www.personality-project.org/revelle/publications/ra.ari.98.pdf).

Examples

data(small.msq)

[Package psych version 2.6.1 Index]