ICLUST.cluster {psych}	R Documentation

Function to form hierarchical cluster analysis of items

Description

The guts of the ICLUST algorithm. Called by ICLUST See ICLUST for description.

Usage

ICLUST.cluster(r.mat, ICLUST.options,smc.items)

Arguments

r.mat
A correlation matrix

ICLUST.options
A list of options (see ICLUST)

smc.items
passed from the main program to speed up processing

Details

See ICLUST

Value

A list of cluster statistics, described more fully in ICLUST

comp1
Description of 'comp1'

comp2
Description of 'comp2'

...

Note

Although the main code for ICLUST is here in ICLUST.cluster, the more extensive documentation is for ICLUST.

Author(s)

William Revelle

References

Revelle, W. 1979, Hierarchical Cluster Analysis and the Internal Structure of Tests. Multivariate Behavioral Research, 14, 57-74. https://personality-project.org/revelle/publications/iclust.pdf
See also more extensive documentation at https://personality-project.org/r/r.ICLUST.html

See Also

ICLUST.graph,ICLUST, cluster.fit , VSS, omega

[Package psych version 1.9.12 Index]