Function to form hierarchical cluster analysis of items

The guts of the ICLUST algorithm. Called by ICLUST See ICLUST for description.

ICLUST.cluster(r.mat, ICLUST.options,smc.items)

Arguments

r.mat
A correlation matrix

ICLUST.options
A list of options (see ICLUST)

smc.items
passed from the main program to speed up processing

Details

See ICLUST

Value

A list of cluster statistics, described more fully in ICLUST

comp1

Description of 'comp1'

comp2

Description of 'comp2'

...
References

Revelle, W. 1979, Hierarchical Cluster Analysis and the Internal Structure of Tests. Multivariate Behavioral Research, 14, 57-74. http://personality-project.org/revelle/publications/iclust.pdf See also more extensive documentation at http://personality-project.org/r/r.ICLUST.html

Note

Although the main code for ICLUST is here in ICLUST.cluster, the more extensive documentation is for ICLUST.

See also

ICLUST.graph,ICLUST, cluster.fit , VSS, omega