Find the Kaiser, Meyer, Olkin Measure of Sampling Adequacy

Henry Kaiser (1970) introduced an Measure of Sampling Adequacy (MSA) of factor analytic data matrices. Kaiser and Rice (1974) then modified it. This is just a function of the squared elements of the `image' matrix compared to the squares of the original correlations. The overall MSA as well as estimates for each item are found. The index is known as the Kaiser-Meyer-Olkin (KMO) index.

KMO(r)

Arguments

r
A correlation matrix or a data matrix (correlations will be found)

Details

Let \(S^2 = diag(R^{-1})^{-1} \) and \(Q = SR^{-1}S\). Then Q is said to be the anti-image intercorrelation matrix. Let \(sumr2 = \sum{R^2}\) and \(sumq2 = \sum{Q^2}\) for all off diagonal elements of R and Q, then \(SMA=sumr2/(sumr2 + sumq2)\). Although originally MSA was 1 - sumq2/sumr2 (Kaiser, 1970), this was modified in Kaiser and Rice, (1974) to be \(SMA=sumr2/(sumr2 + sumq2)\). This is the formula used by Dziuban and Shirkey (1974) and by SPSS. In his delightfully flamboyant style, Kaiser (1975) suggested that KMO > .9 were marvelous, in the .80s, mertitourious, in the .70s, middling, in the .60s, medicore, in the 50s, miserable, and less than .5, unacceptable. An alternative measure of whether the matrix is factorable is the Bartlett test cortest.bartlett which tests the degree that the matrix deviates from an identity matrix.

Value

MSA The overall Measure of Sampling Adequacy

MSAi The measure of sampling adequacy for each item itemImage The Image correlation matrix (Q)

References

H.~F. Kaiser. (1970) A second generation little jiffy. Psychometrika, 35(4):401--415.

H.~F. Kaiser and J.~Rice. (1974) Little jiffy, mark iv. Educational and Psychological Measurement, 34(1):111--117.

H.F. Kaiser. 1974) An index of factor simplicity. Psychometrika, 39 (1) 31-36.

Dziuban, Charles D. and Shirkey, Edwin C. (1974) When is a correlation matrix appropriate for factor analysis? Some decision rules. Psychological Bulletin, 81 (6) 358 - 361.

See also

See Also as fa, cortest.bartlett, Harman.political.

Examples

KMO(Thurstone)

#> Kaiser-Meyer-Olkin factor adequacy
#> Call: KMO(r = Thurstone)
#> Overall MSA =  0.88
#> MSA for each item = 
#>         Sentences        Vocabulary   Sent.Completion     First.Letters 
#>              0.86              0.86              0.90              0.86 
#> Four.Letter.Words          Suffixes     Letter.Series         Pedigrees 
#>              0.88              0.92              0.85              0.93 
#>      Letter.Group 
#>              0.87 

KMO(Harman.political)   #compare to the results in Dziuban and Shirkey (1974)

#> Kaiser-Meyer-Olkin factor adequacy
#> Call: KMO(r = Harman.political)
#> Overall MSA =  0.81
#> MSA for each item = 
#>         Lewis     Roosevelt  Party Voting Median Rental Homeownership 
#>          0.73          0.76          0.84          0.87          0.53 
#>  Unemployment      Mobility     Education 
#>          0.93          0.78          0.86