25 Personality items representing 5 factors

25 personality self report items taken from the International Personality Item Pool (ipip.ori.org) were included as part of the Synthetic Aperture Personality Assessment (SAPA) web based personality assessment project. The data from 2800 subjects are included here as a demonstration set for scale construction, factor analysis, and Item Response Theory analysis. Three additional demographic variables (sex, education, and age) are also included.

data(bfi)
data(bfi.dictionary)

Format

A data frame with 2800 observations on the following 28 variables. (The q numbers are the SAPA item numbers).

A1

Am indifferent to the feelings of others. (q_146)

A2

Inquire about others' well-being. (q_1162)

A3

Know how to comfort others. (q_1206)

A4

Love children. (q_1364)

A5

Make people feel at ease. (q_1419)

C1

Am exacting in my work. (q_124)

C2

Continue until everything is perfect. (q_530)

C3

Do things according to a plan. (q_619)

C4

Do things in a half-way manner. (q_626)

C5

Waste my time. (q_1949)

E1

Don't talk a lot. (q_712)

E2

Find it difficult to approach others. (q_901)

E3

Know how to captivate people. (q_1205)

E4

Make friends easily. (q_1410)

E5

Take charge. (q_1768)

N1

Get angry easily. (q_952)

N2

Get irritated easily. (q_974)

N3

Have frequent mood swings. (q_1099

N4

Often feel blue. (q_1479)

N5

Panic easily. (q_1505)

O1

Am full of ideas. (q_128)

O2

Avoid difficult reading material.(q_316)

O3

Carry the conversation to a higher level. (q_492)

O4

Spend time reflecting on things. (q_1738)

O5

Will not probe deeply into a subject. (q_1964)

gender

Males = 1, Females =2

education

1 = HS, 2 = finished HS, 3 = some college, 4 = college graduate 5 = graduate degree

age

age in years

Details

The first 25 items are organized by five putative factors: Agreeableness, Conscientiousness, Extraversion, Neuroticism, and Opennness. The scoring key is created using make.keys, the scores are found using score.items.

These five factors are a useful example of using irt.fa to do Item Response Theory based latent factor analysis of the polychoric correlation matrix. The endorsement plots for each item, as well as the item information functions reveal that the items differ in their quality.

The item data were collected using a 6 point response scale: 1 Very Inaccurate 2 Moderately Inaccurate 3 Slightly Inaccurate 4 Slightly Accurate 5 Moderately Accurate 6 Very Accurate

as part of the Synthetic Apeture Personality Assessment (SAPA http://sapa-project.org) project. To see an example of the data collection technique, visit http://SAPA-project.org. The items given were sampled from the International Personality Item Pool of Lewis Goldberg using the sampling technique of SAPA. This is a sample data set taken from the much larger SAPA data bank.

Source

The items are from the ipip (Goldberg, 1999). The data are from the SAPA project (Revelle, Wilt and Rosenthal, 2010) , collected Spring, 2010 ( http://sapa-project.org).

References

Goldberg, L.R. (1999) A broad-bandwidth, public domain, personality inventory measuring the lower-level facets of several five-factor models. In Mervielde, I. and Deary, I. and De Fruyt, F. and Ostendorf, F. (eds) Personality psychology in Europe. 7. Tilburg University Press. Tilburg, The Netherlands.

Revelle, W., Wilt, J., and Rosenthal, A. (2010) Individual Differences in Cognition: New Methods for examining the Personality-Cognition Link In Gruszka, A. and Matthews, G. and Szymura, B. (Eds.) Handbook of Individual Differences in Cognition: Attention, Memory and Executive Control, Springer.

See also

bi.bars to show the data by age and gender, irt.fa for item factor analysis applying the irt model.

Note

The bfi data set and items should not be confused with the BFI (Big Five Inventory) of Oliver John and colleagues (John, O. P., Donahue, E. M., & Kentle, R. L. (1991). The Big Five Inventory--Versions 4a and 54. Berkeley, CA: University of California,Berkeley, Institute of Personality and Social Research.)

Examples

data(bfi)
describe(bfi)

#>           vars    n  mean    sd median trimmed   mad min max range  skew
#> A1           1 2784  2.41  1.41      2    2.23  1.48   1   6     5  0.83
#> A2           2 2773  4.80  1.17      5    4.98  1.48   1   6     5 -1.12
#> A3           3 2774  4.60  1.30      5    4.79  1.48   1   6     5 -1.00
#> A4           4 2781  4.70  1.48      5    4.93  1.48   1   6     5 -1.03
#> A5           5 2784  4.56  1.26      5    4.71  1.48   1   6     5 -0.85
#> C1           6 2779  4.50  1.24      5    4.64  1.48   1   6     5 -0.85
#> C2           7 2776  4.37  1.32      5    4.50  1.48   1   6     5 -0.74
#> C3           8 2780  4.30  1.29      5    4.42  1.48   1   6     5 -0.69
#> C4           9 2774  2.55  1.38      2    2.41  1.48   1   6     5  0.60
#> C5          10 2784  3.30  1.63      3    3.25  1.48   1   6     5  0.07
#> E1          11 2777  2.97  1.63      3    2.86  1.48   1   6     5  0.37
#> E2          12 2784  3.14  1.61      3    3.06  1.48   1   6     5  0.22
#> E3          13 2775  4.00  1.35      4    4.07  1.48   1   6     5 -0.47
#> E4          14 2791  4.42  1.46      5    4.59  1.48   1   6     5 -0.82
#> E5          15 2779  4.42  1.33      5    4.56  1.48   1   6     5 -0.78
#> N1          16 2778  2.93  1.57      3    2.82  1.48   1   6     5  0.37
#> N2          17 2779  3.51  1.53      4    3.51  1.48   1   6     5 -0.08
#> N3          18 2789  3.22  1.60      3    3.16  1.48   1   6     5  0.15
#> N4          19 2764  3.19  1.57      3    3.12  1.48   1   6     5  0.20
#> N5          20 2771  2.97  1.62      3    2.85  1.48   1   6     5  0.37
#> O1          21 2778  4.82  1.13      5    4.96  1.48   1   6     5 -0.90
#> O2          22 2800  2.71  1.57      2    2.56  1.48   1   6     5  0.59
#> O3          23 2772  4.44  1.22      5    4.56  1.48   1   6     5 -0.77
#> O4          24 2786  4.89  1.22      5    5.10  1.48   1   6     5 -1.22
#> O5          25 2780  2.49  1.33      2    2.34  1.48   1   6     5  0.74
#> gender      26 2800  1.67  0.47      2    1.71  0.00   1   2     1 -0.73
#> education   27 2577  3.19  1.11      3    3.22  1.48   1   5     4 -0.05
#> age         28 2800 28.78 11.13     26   27.43 10.38   3  86    83  1.02
#>           kurtosis   se
#> A1           -0.31 0.03
#> A2            1.05 0.02
#> A3            0.44 0.02
#> A4            0.04 0.03
#> A5            0.16 0.02
#> C1            0.30 0.02
#> C2           -0.14 0.03
#> C3           -0.13 0.02
#> C4           -0.62 0.03
#> C5           -1.22 0.03
#> E1           -1.09 0.03
#> E2           -1.15 0.03
#> E3           -0.47 0.03
#> E4           -0.30 0.03
#> E5           -0.09 0.03
#> N1           -1.01 0.03
#> N2           -1.05 0.03
#> N3           -1.18 0.03
#> N4           -1.09 0.03
#> N5           -1.06 0.03
#> O1            0.43 0.02
#> O2           -0.81 0.03
#> O3            0.30 0.02
#> O4            1.08 0.02
#> O5           -0.24 0.03
#> gender       -1.47 0.01
#> education    -0.32 0.02
#> age           0.56 0.21

 keys.list <-
  list(agree=c("-A1","A2","A3","A4","A5"),conscientious=c("C1","C2","C3","-C4","-C5"),
extraversion=c("-E1","-E2","E3","E4","E5"),neuroticism=c("N1","N2","N3","N4","N5"),
openness = c("O1","-O2","O3","O4","-O5"))
 scores <- scoreItems(keys.list,bfi,min=1,max=6) #specify the minimum and maximum values
 scores

#> Call: scoreItems(keys = keys.list, items = bfi, min = 1, max = 6)
#> 
#> (Unstandardized) Alpha:
#>       agree conscientious extraversion neuroticism openness
#> alpha   0.7          0.72         0.76        0.81      0.6
#> 
#> Standard errors of unstandardized Alpha:
#>       agree conscientious extraversion neuroticism openness
#> ASE   0.014         0.014        0.013       0.011    0.017
#> 
#> Average item correlation:
#>           agree conscientious extraversion neuroticism openness
#> average.r  0.32          0.34         0.39        0.46     0.23
#> 
#>  Guttman 6* reliability: 
#>          agree conscientious extraversion neuroticism openness
#> Lambda.6   0.7          0.72         0.76        0.81      0.6
#> 
#> Signal/Noise based upon av.r : 
#>              agree conscientious extraversion neuroticism openness
#> Signal/Noise   2.3           2.6          3.2         4.3      1.5
#> 
#> Scale intercorrelations corrected for attenuation 
#>  raw correlations below the diagonal, alpha on the diagonal 
#>  corrected correlations above the diagonal:
#>               agree conscientious extraversion neuroticism openness
#> agree          0.70          0.36         0.63      -0.245     0.23
#> conscientious  0.26          0.72         0.35      -0.305     0.30
#> extraversion   0.46          0.26         0.76      -0.284     0.32
#> neuroticism   -0.18         -0.23        -0.22       0.812    -0.12
#> openness       0.15          0.19         0.22      -0.086     0.60
#> 
#>  In order to see the item by scale loadings and frequency counts of the data
#>  print with the short option = FALSE

 #show the use of the fa.lookup with a dictionary
 keys.lookup(keys.list,bfi.dictionary[,1:4])

#>     ItemLabel                                      Item     Giant3
#> A1-     q_146 Am indifferent to the feelings of others.   Cohesion
#> A2     q_1162         Inquire about others' well-being.   Cohesion
#> A3     q_1206               Know how to comfort others.   Cohesion
#> A4     q_1364                            Love children.   Cohesion
#> A5     q_1419                 Make people feel at ease.   Cohesion
#> C1      q_124                   Am exacting in my work.  Stability
#> C2      q_530     Continue until everything is perfect.  Stability
#> C3      q_619            Do things according to a plan.  Stability
#> C4-     q_626           Do things in a half-way manner.  Stability
#> C5-    q_1949                            Waste my time.  Stability
#> E1-     q_712                         Don't talk a lot. Plasticity
#> E2-     q_901     Find it difficult to approach others. Plasticity
#> E3     q_1205             Know how to captivate people. Plasticity
#> E4     q_1410                      Make friends easily. Plasticity
#> E5     q_1768                              Take charge. Plasticity
#> N1      q_952                         Get angry easily.  Stability
#> N2      q_974                     Get irritated easily.  Stability
#> N3     q_1099                Have frequent mood swings.  Stability
#> N4     q_1479                          Often feel blue.  Stability
#> N5     q_1505                             Panic easily.  Stability
#> O1      q_128                         Am full of ideas. Plasticity
#> O2-     q_316         Avoid difficult reading material. Plasticity
#> O3      q_492 Carry the conversation to a higher level. Plasticity
#> O4     q_1738          Spend time reflecting on things. Plasticity
#> O5-    q_1964     Will not probe deeply into a subject. Plasticity
#>                    Big6
#> A1-       Agreeableness
#> A2        Agreeableness
#> A3        Agreeableness
#> A4        Agreeableness
#> A5        Agreeableness
#> C1    Conscientiousness
#> C2    Conscientiousness
#> C3    Conscientiousness
#> C4-   Conscientiousness
#> C5-   Conscientiousness
#> E1-        Extraversion
#> E2-        Extraversion
#> E3         Extraversion
#> E4         Extraversion
#> E5         Extraversion
#> N1  Emotional Stability
#> N2  Emotional Stability
#> N3  Emotional Stability
#> N4  Emotional Stability
#> N5  Emotional Stability
#> O1             Openness
#> O2-            Openness
#> O3             Openness
#> O4             Openness
#> O5-            Openness