comorbidity {psych}	R Documentation

Convert base rates of two diagnoses and their comorbidity into phi, Yule, and tetrachorics

Description

In medicine and clinical psychology, diagnoses tend to be categorical (someone is depressed or not, someone has an anxiety disorder or not). Cooccurrence of both of these symptoms is called comorbidity. Diagnostic categories vary in their degree of comorbidity with other diagnostic categories. From the point of view of correlation, comorbidity is just a name applied to one cell in a four fold table. It is thus possible to analyze comorbidity rates by considering the probability of the separate diagnoses and the probability of the joint diagnosis. This gives the two by two table needed for a phi, Yule, or tetrachoric correlation.

Usage

comorbidity(d1, d2, com, labels = NULL)

Arguments

d1
Proportion of diagnostic category 1

d2
Proportion of diganostic category 2

com
Proportion of comorbidity (diagnostic category 1 and 2)

labels
Names of categories 1 and 2

Value

twobytwo
The two by two table implied by the input

phi
Phi coefficient of the two by two table

Yule
Yule coefficient of the two by two table

tetra
Tetrachoric coefficient of the two by two table

Author(s)

William Revelle

See Also

phi, phi2tetra ,Yule, Yule.inv Yule2phi, tetrachoric and polychoric, as well as AUC for graphical displays

Examples

comorbidity(.2,.15,.1,c("Anxiety","Depression")) 

[Package psych version 1.9.12 Index]