cor2dist {psych}	R Documentation

Convert correlations to distances (necessary to do multidimensional scaling of correlation data)

Description

A minor helper function to convert correlations (ranging from -1 to 1) to distances (ranging from 0 to 2). d = √{(2(1-r))}.

Usage

cor2dist(x)

Arguments

x
If square, then assumed to be a correlation matrix, otherwise the correlations are found first.

Value

dist: a square matrix of distances.

Note

For an example of doing multidimensional scaling on data that are normally factored, see Revelle (in prep)

Author(s)

William Revelle

References

Revelle, William. (in prep) An introduction to psychometric theory with applications in R. Springer. Working draft available at https://personality-project.org/r/book/

[Package psych version 1.9.12 Index]