Count number of pairwise cases for a data set with missing (NA) data.

When doing cor(x, use= "pairwise"), it is nice to know the number of cases for each pairwise correlation. This is particularly useful when doing SAPA type analyses.

count.pairwise(x, y = NULL,diagonal=TRUE)
pairwiseDescribe(x,diagonal=FALSE,...)

Arguments

x
An input matrix, typically a data matrix ready to be correlated.

y
An optional second input matrix

diagonal
if TRUE, then report the diagonal, else fill the diagonals with NA

...
Other parameters to pass to describe

Value

result = matrix of counts of pairwise observations

Examples

not_run({
  x <- matrix(rnorm(1000),ncol=6)
  y <- matrix(rnorm(500),ncol=3)
  x[x < 0] <- NA
  y[y > 1] <- NA

  count.pairwise(x)
  count.pairwise(y)
  count.pairwise(x,y)
  count.pairwise(x,diagonal=FALSE)
  pairwiseDescribe(x,quant=c(.1,.25,.5,.75,.9))
})