Deprecated Exploratory Factor analysis functions. Please use fa

After 4 years, it is time to stop using these deprecated functions! Please see fa which includes all of the functionality of these older functions.

fa.poly(x,nfactors=1,n.obs = NA, n.iter=1, rotate="oblimin", SMC=TRUE,  missing=FALSE,
 impute="median", min.err = .001, max.iter=50, symmetric=TRUE, warnings=TRUE,
 fm="minres",alpha=.1, p =.05,scores="regression", oblique.scores=TRUE,
        weight=NULL,global=TRUE,...)  #deprecated

factor.minres(r, nfactors=1, residuals = FALSE, rotate = "varimax",n.obs = NA,
scores = FALSE,SMC=TRUE, missing=FALSE,impute="median",min.err = 0.001, digits = 2,
 max.iter = 50,symmetric=TRUE,warnings=TRUE,fm="minres")    #deprecated

factor.wls(r,nfactors=1,residuals=FALSE,rotate="varimax",n.obs = NA,
scores=FALSE,SMC=TRUE,missing=FALSE,impute="median", min.err = .001,
 digits=2,max.iter=50,symmetric=TRUE,warnings=TRUE,fm="wls")  #deprecated

Arguments

r
deprecated.

x
deprecated

nfactors
deprecated

n.obs
deprecated

rotate
deprecated

n.iter
deprecated

residuals
deprecated

scores
deprecated

SMC
deprecated

missing
deprecated

impute
deprecated

max.iter
deprecated

symmetric
deprecated

warnings
deprecated

fm
deprecated

alpha
deprecated

p
deprecated

oblique.scores
deprecated

weight
deprecated

global
deprecated

digits
deprecated

min.err
deprecated

...
deprecated

Details

Please see the writeup for fa for all of the functionality in these older functions.

Value

values

Eigen values of the common factor solution

e.values

Eigen values of the original matrix

communality

Communality estimates for each item. These are merely the sum of squared factor loadings for that item.

communalities

If using minrank factor analysis, these are the communalities reflecting the total amount of common variance. They will exceed the communality (above) which is the model estimated common variance.

rotation

which rotation was requested?

loadings

An item by factor (pattern) loading matrix of class ``loadings" Suitable for use in other programs (e.g., GPA rotation or factor2cluster. To show these by sorted order, use print.psych with sort=TRUE

complexity

Hoffman's index of complexity for each item. This is just \(\frac{(\Sigma a_i^2)^2}{\Sigma a_i^4}\) where a_i is the factor loading on the ith factor. From Hofmann (1978), MBR. See also Pettersson and Turkheimer (2010).

Structure

An item by factor structure matrix of class ``loadings". This is just the loadings (pattern) matrix times the factor intercorrelation matrix.

fit

How well does the factor model reproduce the correlation matrix. This is just \(\frac{\Sigma r_{ij}^2 - \Sigma r^{*2}_{ij} }{\Sigma r_{ij}^2} \) (See VSS, ICLUST, and principal for this fit statistic.

fit.off

how well are the off diagonal elements reproduced?

dof

Degrees of Freedom for this model. This is the number of observed correlations minus the number of independent parameters. Let n=Number of items, nf = number of factors then \(dof = n * (n-1)/2 - n * nf + nf*(nf-1)/2\)

objective

Value of the function that is minimized by a maximum likelihood procedures. This is reported for comparison purposes and as a way to estimate chi square goodness of fit. The objective function is \(f = log(trace ((FF'+U2)^{-1} R) - log(|(FF'+U2)^{-1} R|) - n.items\). When using MLE, this function is minimized. When using OLS (minres), although we are not minimizing this function directly, we can still calculate it in order to compare the solution to a MLE fit.

STATISTIC

If the number of observations is specified or found, this is a chi square based upon the objective function, f (see above). Using the formula from factanal(which seems to be Bartlett's test) : \(\chi^2 = (n.obs - 1 - (2 * p + 5)/6 - (2 * factors)/3)) * f \)

PVAL

If n.obs > 0, then what is the probability of observing a chisquare this large or larger?

Phi

If oblique rotations (using oblimin from the GPArotation package or promax) are requested, what is the interfactor correlation.

communality.iterations

The history of the communality estimates (For principal axis only.) Probably only useful for teaching what happens in the process of iterative fitting.

residual

The matrix of residual correlations after the factor model is applied. To display it conveniently, use the residuals command.

chi

When normal theory fails (e.g., in the case of non-positive definite matrices), it useful to examine the empirically derived \(\chi^2\) based upon the sum of the squared residuals * N. This will differ slightly from the MLE estimate which is based upon the fitting function rather than the actual residuals.

rms

This is the sum of the squared (off diagonal residuals) divided by the degrees of freedom. Comparable to an RMSEA which, because it is based upon \(\chi^2\), requires the number of observations to be specified. The rms is an empirical value while the RMSEA is based upon normal theory and the non-central \(\chi^2\) distribution. That is to say, if the residuals are particularly non-normal, the rms value and the associated \(\chi^2\) and RMSEA can differ substantially.

crms

rms adjusted for degrees of freedom

RMSEA

The Root Mean Square Error of Approximation is based upon the non-central \(\chi^2\) distribution and the \(\chi^2\) estimate found from the MLE fitting function. With normal theory data, this is fine. But when the residuals are not distributed according to a noncentral \(\chi^2\), this can give very strange values. (And thus the confidence intervals can not be calculated.) The RMSEA is a conventional index of goodness (badness) of fit but it is also useful to examine the actual rms values.

TLI

The Tucker Lewis Index of factoring reliability which is also known as the non-normed fit index.

BIC

Based upon \(\chi^2\) with the assumption of normal theory and using the \(\chi^2\) found using the objective function defined above. This is just \(\chi^2 - 2 df\)

eBIC

When normal theory fails (e.g., in the case of non-positive definite matrices), it useful to examine the empirically derived eBIC based upon the empirical \(\chi^2\) - 2 df.

R2

The multiple R square between the factors and factor score estimates, if they were to be found. (From Grice, 2001). Derived from R2 is is the minimum correlation between any two factor estimates = 2R2-1.

r.scores

The correlations of the factor score estimates using the specified model, if they were to be found. Comparing these correlations with that of the scores themselves will show, if an alternative estimate of factor scores is used (e.g., the tenBerge method), the problem of factor indeterminacy. For these correlations will not necessarily be the same.

weights

The beta weights to find the factor score estimates. These are also used by the predict.psych function to find predicted factor scores for new cases.

scores

The factor scores as requested. Note that these scores reflect the choice of the way scores should be estimated (see scores in the input). That is, simple regression ("Thurstone"), correlaton preserving ("tenBerge") as well as "Anderson" and "Bartlett" using the appropriate algorithms (see factor.scores). The correlation between factor score estimates (r.scores) is based upon using the regression/Thurstone approach. The actual correlation between scores will reflect the rotation algorithm chosen and may be found by correlating those scores.

valid

The validity coffiecient of course coded (unit weighted) factor score estimates (From Grice, 2001)

score.cor

The correlation matrix of course coded (unit weighted) factor score estimates, if they were to be found, based upon the loadings matrix rather than the weights matrix.

rot.mat

The rotation matrix as returned from GPArotation.

References

Gorsuch, Richard, (1983) Factor Analysis. Lawrence Erlebaum Associates.

Grice, James W. (2001), Computing and evaluating factor scores. Psychological Methods, 6, 430-450

Harman, Harry and Jones, Wayne (1966) Factor analysis by minimizing residuals (minres), Psychometrika, 31, 3, 351-368.

Hofmann, R. J. ( 1978 ) . Complexity and simplicity as objective indices descriptive of factor solutions. Multivariate Behavioral Research, 13, 247-250.

Pettersson E, Turkheimer E. (2010) Item selection, evaluation, and simple structure in personality data. Journal of research in personality, 44(4), 407-420.

Revelle, William. (in prep) An introduction to psychometric theory with applications in R. Springer. Working draft available at http://personality-project.org/r/book/

Shapiro, A. and ten Berge, Jos M. F, (2002) Statistical inference of minimum rank factor analysis. Psychometika, (67) 79-84.

ten Berge, Jos M. F. and Kiers, Henk A. L. (1991). A numerical approach to the approximate and the exact minimum rank of a covariance matrix. Psychometrika, (56) 309-315.

Note

Thanks to Erich Studerus for some very helpful suggestions about various rotation and factor scoring algorithms, and to Gumundur Arnkelsson for suggestions about factor scores for singular matrices.

The fac function is the original fa function which is now called by fa repeatedly to get confidence intervals.

SPSS will sometimes use a Kaiser normalization before rotating. This will lead to different solutions than reported here. To get the Kaiser normalized loadings, use kaiser.

The communality for a variable is the amount of variance accounted for by all of the factors. That is to say, for orthogonal factors, it is the sum of the squared factor loadings (rowwise). The communality is insensitive to rotation. However, if an oblique solution is found, then the communality is not the sum of squared pattern coefficients. In both cases (oblique or orthogonal) the communality is the diagonal of the reproduced correlation matrix where \(_nR_n = _{n}P_{k k}\Phi_{k k}P_n'\) where P is the pattern matrix and \(\Phi\) is the factor intercorrelation matrix. This is the same, of course to multiplying the pattern by the structure: \(R = P S'\) R = PS' where the Structure matrix is \(S = \Phi P\). Similarly, the eigen values are the diagonal of the product \( _k\Phi_{kk}P'_{nn}P_{k} \).

A frequently asked question is why are the factor names of the rotated solution not in ascending order? That is, for example, if factoring the 25 items of the bfi, the factor names are MR2 MR3 MR5 MR1 MR4, rather than the seemingly more logical "MR1" "MR2" "MR3" "MR4" "MR5". This is for pedagogical reasons, in that factors as extracted are orthogonal and are in order of amount of variance accounted for. But when rotated (orthogonally) or transformed (obliquely) the simple structure solution does not preserve that order. The factor names are, of course, arbitrary, and are kept with the original names to show the effect of rotation/transformation. To give them names associated with their ordinal position, simply paste("F", 1:nf, sep="") where nf is the number of factors. See the last example.

Correction to documentation: as of September, 2014, the oblique.scores option is correctly explained. (It had been backwards.) The default (oblique.scores=FALSE) finds scores based upon the Structure matrix, while oblique.scores=TRUE finds them based upon the pattern matrix. The latter case matches factanal. This error was detected by Mark Seeto.

See also

principal for principal components analysis (PCA). PCA will give very similar solutions to factor analysis when there are many variables. The differences become more salient as the number variables decrease. The PCA and FA models are actually very different and should not be confused. One is a model of the observed variables, the other is a model of latent variables. irt.fa for Item Response Theory analyses using factor analysis, using the two parameter IRT equivalent of loadings and difficulties. VSS will produce the Very Simple Structure (VSS) and MAP criteria for the number of factors, nfactors to compare many different factor criteria. ICLUST will do a hierarchical cluster analysis alternative to factor analysis or principal components analysis.

predict.psych to find predicted scores based upon new data, fa.extension to extend the factor solution to new variables, omega for hierarchical factor analysis with one general factor. codefa.multi for hierarchical factor analysis with an arbitrary number of higher order factors. fa.sort will sort the factor loadings into echelon form. fa.organize will reorganize the factor pattern matrix into any arbitrary order of factors and items. KMO and cortest.bartlett for various tests that some people like. factor2cluster will prepare unit weighted scoring keys of the factors that can be used with scoreItems. fa.lookup will print the factor analysis loadings matrix along with the item ``content" taken from a dictionary of items. This is useful when examining the meaning of the factors. anova.psych allows for testing the difference between two (presumably nested) factor models .

Examples

#using the Harman 24 mental tests, compare a principal factor with a principal components solution
pc <- principal(Harman74.cor$cov,4,rotate="varimax")   #principal components
pa <- fa(Harman74.cor$cov,4,fm="pa" ,rotate="varimax")  #principal axis 
uls <- fa(Harman74.cor$cov,4,rotate="varimax")          #unweighted least squares is minres
wls <- fa(Harman74.cor$cov,4,fm="wls")       #weighted least squares

#to show the loadings sorted by absolute value
print(uls,sort=TRUE)

#> Factor Analysis using method =  minres
#> Call: fa(r = Harman74.cor$cov, nfactors = 4, rotate = "varimax")
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>                        item  MR1   MR3   MR2  MR4   h2   u2 com
#> SentenceCompletion        7 0.81  0.19  0.15 0.07 0.73 0.27 1.2
#> WordMeaning               9 0.81  0.20  0.05 0.22 0.74 0.26 1.3
#> PargraphComprehension     6 0.76  0.21  0.07 0.23 0.68 0.32 1.4
#> GeneralInformation        5 0.73  0.19  0.22 0.14 0.64 0.36 1.4
#> WordClassification        8 0.57  0.34  0.23 0.14 0.51 0.49 2.2
#> VisualPerception          1 0.15  0.68  0.20 0.15 0.55 0.45 1.4
#> PaperFormBoard            3 0.15  0.55 -0.01 0.11 0.34 0.66 1.2
#> Flags                     4 0.23  0.53  0.09 0.07 0.35 0.65 1.5
#> SeriesCompletion         23 0.37  0.52  0.23 0.22 0.51 0.49 2.7
#> Cubes                     2 0.11  0.45  0.08 0.08 0.23 0.77 1.3
#> Deduction                20 0.38  0.42  0.10 0.29 0.42 0.58 2.9
#> ProblemReasoning         22 0.37  0.41  0.13 0.29 0.40 0.60 3.0
#> Addition                 10 0.17 -0.11  0.82 0.16 0.74 0.26 1.2
#> CountingDots             12 0.02  0.20  0.71 0.09 0.55 0.45 1.2
#> StraightCurvedCapitals   13 0.18  0.42  0.54 0.08 0.51 0.49 2.2
#> Code                     11 0.18  0.11  0.54 0.37 0.47 0.53 2.1
#> ArithmeticProblems       24 0.36  0.19  0.49 0.29 0.49 0.51 2.9
#> NumericalPuzzles         21 0.18  0.40  0.43 0.21 0.42 0.58 2.8
#> ObjectNumber             17 0.14  0.06  0.22 0.58 0.41 0.59 1.4
#> WordRecognition          14 0.21  0.05  0.08 0.56 0.36 0.64 1.3
#> NumberRecognition        15 0.12  0.12  0.08 0.52 0.31 0.69 1.3
#> FigureRecognition        16 0.07  0.42  0.06 0.52 0.45 0.55 2.0
#> NumberFigure             18 0.02  0.31  0.34 0.45 0.41 0.59 2.7
#> FigureWord               19 0.15  0.25  0.18 0.35 0.23 0.77 2.8
#> 
#>                        MR1  MR3  MR2  MR4
#> SS loadings           3.64 2.93 2.67 2.23
#> Proportion Var        0.15 0.12 0.11 0.09
#> Cumulative Var        0.15 0.27 0.38 0.48
#> Proportion Explained  0.32 0.26 0.23 0.19
#> Cumulative Proportion 0.32 0.57 0.81 1.00
#> 
#> Mean item complexity =  1.9
#> Test of the hypothesis that 4 factors are sufficient.
#> 
#> The degrees of freedom for the null model are  276  and the objective function was  11.44
#> The degrees of freedom for the model are 186  and the objective function was  1.72 
#> 
#> The root mean square of the residuals (RMSR) is  0.04 
#> The df corrected root mean square of the residuals is  0.05 
#> 
#> Fit based upon off diagonal values = 0.98
#> Measures of factor score adequacy             
#>                                                 MR1  MR3  MR2  MR4
#> Correlation of scores with factors             0.93 0.87 0.91 0.82
#> Multiple R square of scores with factors       0.87 0.76 0.83 0.68
#> Minimum correlation of possible factor scores  0.74 0.52 0.65 0.36

#then compare with a maximum likelihood solution using factanal
mle <- factanal(covmat=Harman74.cor$cov,factors=4)
factor.congruence(list(mle,pa,pc,uls,wls))

#>         Factor1 Factor2 Factor3 Factor4  PA1  PA3  PA2  PA4  RC1  RC3  RC2  RC4
#> Factor1    1.00    0.61    0.46    0.56 1.00 0.61 0.46 0.55 1.00 0.54 0.44 0.47
#> Factor2    0.61    1.00    0.50    0.61 0.61 1.00 0.50 0.60 0.60 0.99 0.49 0.52
#> Factor3    0.46    0.50    1.00    0.57 0.46 0.50 1.00 0.56 0.45 0.44 1.00 0.48
#> Factor4    0.56    0.61    0.57    1.00 0.56 0.62 0.58 1.00 0.55 0.55 0.56 0.99
#> PA1        1.00    0.61    0.46    0.56 1.00 0.61 0.46 0.55 1.00 0.54 0.44 0.47
#> PA3        0.61    1.00    0.50    0.62 0.61 1.00 0.50 0.61 0.61 0.99 0.50 0.53
#> PA2        0.46    0.50    1.00    0.58 0.46 0.50 1.00 0.57 0.46 0.44 1.00 0.49
#> PA4        0.55    0.60    0.56    1.00 0.55 0.61 0.57 1.00 0.54 0.54 0.55 0.99
#> RC1        1.00    0.60    0.45    0.55 1.00 0.61 0.46 0.54 1.00 0.53 0.43 0.46
#> RC3        0.54    0.99    0.44    0.55 0.54 0.99 0.44 0.54 0.53 1.00 0.43 0.47
#> RC2        0.44    0.49    1.00    0.56 0.44 0.50 1.00 0.55 0.43 0.43 1.00 0.47
#> RC4        0.47    0.52    0.48    0.99 0.47 0.53 0.49 0.99 0.46 0.47 0.47 1.00
#> MR1        1.00    0.61    0.46    0.56 1.00 0.61 0.46 0.55 1.00 0.54 0.44 0.47
#> MR3        0.61    1.00    0.50    0.62 0.61 1.00 0.50 0.61 0.61 0.99 0.50 0.53
#> MR2        0.46    0.50    1.00    0.58 0.46 0.50 1.00 0.57 0.46 0.44 1.00 0.49
#> MR4        0.55    0.60    0.56    1.00 0.55 0.61 0.57 1.00 0.54 0.54 0.55 0.99
#> WLS1       0.98    0.47    0.30    0.40 0.98 0.48 0.30 0.39 0.98 0.41 0.28 0.32
#> WLS3       0.36    0.95    0.41    0.41 0.36 0.95 0.41 0.39 0.35 0.97 0.41 0.32
#> WLS2       0.23    0.22    0.95    0.36 0.23 0.22 0.95 0.35 0.22 0.16 0.95 0.28
#> WLS4       0.28    0.40    0.36    0.94 0.28 0.41 0.37 0.94 0.27 0.36 0.35 0.97
#>          MR1  MR3  MR2  MR4 WLS1 WLS3 WLS2 WLS4
#> Factor1 1.00 0.61 0.46 0.55 0.98 0.36 0.23 0.28
#> Factor2 0.61 1.00 0.50 0.60 0.47 0.95 0.22 0.40
#> Factor3 0.46 0.50 1.00 0.56 0.30 0.41 0.95 0.36
#> Factor4 0.56 0.62 0.58 1.00 0.40 0.41 0.36 0.94
#> PA1     1.00 0.61 0.46 0.55 0.98 0.36 0.23 0.28
#> PA3     0.61 1.00 0.50 0.61 0.48 0.95 0.22 0.41
#> PA2     0.46 0.50 1.00 0.57 0.30 0.41 0.95 0.37
#> PA4     0.55 0.61 0.57 1.00 0.39 0.39 0.35 0.94
#> RC1     1.00 0.61 0.46 0.54 0.98 0.35 0.22 0.27
#> RC3     0.54 0.99 0.44 0.54 0.41 0.97 0.16 0.36
#> RC2     0.44 0.50 1.00 0.55 0.28 0.41 0.95 0.35
#> RC4     0.47 0.53 0.49 0.99 0.32 0.32 0.28 0.97
#> MR1     1.00 0.61 0.46 0.55 0.98 0.36 0.23 0.28
#> MR3     0.61 1.00 0.50 0.61 0.48 0.95 0.22 0.41
#> MR2     0.46 0.50 1.00 0.57 0.30 0.41 0.95 0.37
#> MR4     0.55 0.61 0.57 1.00 0.39 0.39 0.35 0.94
#> WLS1    0.98 0.48 0.30 0.39 1.00 0.22 0.09 0.13
#> WLS3    0.36 0.95 0.41 0.39 0.22 1.00 0.17 0.23
#> WLS2    0.23 0.22 0.95 0.35 0.09 0.17 1.00 0.20
#> WLS4    0.28 0.41 0.37 0.94 0.13 0.23 0.20 1.00

#note that the order of factors and the sign of some of factors may differ 

#finally, compare the unrotated factor, ml, uls, and  wls solutions
wls <- fa(Harman74.cor$cov,4,rotate="none",fm="wls")
pa <- fa(Harman74.cor$cov,4,rotate="none",fm="pa")
minres <-  factanal(factors=4,covmat=Harman74.cor$cov,rotation="none")
mle <- fa(Harman74.cor$cov,4,rotate="none",fm="mle")
uls <- fa(Harman74.cor$cov,4,rotate="none",fm="uls")
factor.congruence(list(minres,mle,pa,wls,uls))

#>         Factor1 Factor2 Factor3 Factor4   ML1   ML2  ML3   ML4  PA1   PA2   PA3
#> Factor1    1.00    0.11    0.25    0.06  1.00  0.11 0.25  0.06 1.00 -0.04 -0.05
#> Factor2    0.11    1.00    0.06    0.07  0.11  1.00 0.06  0.07 0.14  0.98 -0.08
#> Factor3    0.25    0.06    1.00    0.01  0.25  0.06 1.00  0.01 0.30  0.10  0.95
#> Factor4    0.06    0.07    0.01    1.00  0.06  0.07 0.01  1.00 0.07  0.13 -0.04
#> ML1        1.00    0.11    0.25    0.06  1.00  0.11 0.25  0.06 1.00 -0.04 -0.05
#> ML2        0.11    1.00    0.06    0.07  0.11  1.00 0.06  0.07 0.14  0.98 -0.08
#> ML3        0.25    0.06    1.00    0.01  0.25  0.06 1.00  0.01 0.30  0.10  0.95
#> ML4        0.06    0.07    0.01    1.00  0.06  0.07 0.01  1.00 0.07  0.13 -0.04
#> PA1        1.00    0.14    0.30    0.07  1.00  0.14 0.30  0.07 1.00  0.00  0.00
#> PA2       -0.04    0.98    0.10    0.13 -0.04  0.98 0.10  0.13 0.00  1.00  0.00
#> PA3       -0.05   -0.08    0.95   -0.04 -0.05 -0.08 0.95 -0.04 0.00  0.00  1.00
#> PA4       -0.01   -0.08    0.02    0.99 -0.01 -0.08 0.02  0.99 0.00  0.00  0.00
#> WLS1       1.00    0.14    0.30    0.07  1.00  0.14 0.30  0.07 1.00  0.00  0.00
#> WLS2      -0.04    0.98    0.09    0.13 -0.04  0.98 0.09  0.13 0.00  1.00 -0.01
#> WLS3      -0.05   -0.07    0.95   -0.04 -0.05 -0.07 0.95 -0.04 0.00  0.01  1.00
#> WLS4      -0.01   -0.07    0.02    0.99 -0.01 -0.07 0.02  0.99 0.00  0.01  0.00
#> ULS1       1.00    0.14    0.30    0.07  1.00  0.14 0.30  0.07 1.00  0.00  0.00
#> ULS2      -0.04    0.98    0.09    0.13 -0.04  0.98 0.09  0.13 0.00  1.00  0.00
#> ULS3      -0.05   -0.07    0.95   -0.04 -0.05 -0.07 0.95 -0.04 0.00  0.00  1.00
#> ULS4      -0.01   -0.08    0.02    0.99 -0.01 -0.08 0.02  0.99 0.00  0.00  0.00
#>           PA4 WLS1  WLS2  WLS3  WLS4 ULS1  ULS2  ULS3  ULS4
#> Factor1 -0.01 1.00 -0.04 -0.05 -0.01 1.00 -0.04 -0.05 -0.01
#> Factor2 -0.08 0.14  0.98 -0.07 -0.07 0.14  0.98 -0.07 -0.08
#> Factor3  0.02 0.30  0.09  0.95  0.02 0.30  0.09  0.95  0.02
#> Factor4  0.99 0.07  0.13 -0.04  0.99 0.07  0.13 -0.04  0.99
#> ML1     -0.01 1.00 -0.04 -0.05 -0.01 1.00 -0.04 -0.05 -0.01
#> ML2     -0.08 0.14  0.98 -0.07 -0.07 0.14  0.98 -0.07 -0.08
#> ML3      0.02 0.30  0.09  0.95  0.02 0.30  0.09  0.95  0.02
#> ML4      0.99 0.07  0.13 -0.04  0.99 0.07  0.13 -0.04  0.99
#> PA1      0.00 1.00  0.00  0.00  0.00 1.00  0.00  0.00  0.00
#> PA2      0.00 0.00  1.00  0.01  0.01 0.00  1.00  0.00  0.00
#> PA3      0.00 0.00 -0.01  1.00  0.00 0.00  0.00  1.00  0.00
#> PA4      1.00 0.00 -0.01  0.00  1.00 0.00  0.00  0.00  1.00
#> WLS1     0.00 1.00  0.00  0.00  0.00 1.00  0.00  0.00  0.00
#> WLS2    -0.01 0.00  1.00  0.00  0.00 0.00  1.00 -0.01 -0.01
#> WLS3     0.00 0.00  0.00  1.00  0.00 0.00  0.01  1.00  0.00
#> WLS4     1.00 0.00  0.00  0.00  1.00 0.00  0.01  0.00  1.00
#> ULS1     0.00 1.00  0.00  0.00  0.00 1.00  0.00  0.00  0.00
#> ULS2     0.00 0.00  1.00  0.01  0.01 0.00  1.00  0.00  0.00
#> ULS3     0.00 0.00 -0.01  1.00  0.00 0.00  0.00  1.00  0.00
#> ULS4     1.00 0.00 -0.01  0.00  1.00 0.00  0.00  0.00  1.00

#in particular, note the similarity of the mle and min res solutions
#note that the order of factors and the sign of some of factors may differ 

#an example of where the ML and PA and MR models differ is found in Thurstone.33.
#compare the first two factors with the 3 factor solution 
Thurstone.33 <- as.matrix(Thurstone.33)
mle2 <- fa(Thurstone.33,2,rotate="none",fm="mle")
mle3 <- fa(Thurstone.33,3 ,rotate="none",fm="mle")
pa2 <- fa(Thurstone.33,2,rotate="none",fm="pa")
pa3 <- fa(Thurstone.33,3,rotate="none",fm="pa")
mr2 <- fa(Thurstone.33,2,rotate="none")
mr3 <- fa(Thurstone.33,3,rotate="none")
factor.congruence(list(mle2,mr2,pa2,mle3,pa3,mr3))

#>       ML1   ML2  MR1   MR2  PA1   PA2   ML1   ML2   ML3   PA1   PA2   PA3   MR1
#> ML1  1.00  0.16 1.00 -0.03 1.00 -0.03  0.98  0.88  0.16  1.00 -0.02  0.00  1.00
#> ML2  0.16  1.00 0.19  0.97 0.19  0.98  0.03  0.51 -0.79  0.18  0.97 -0.09  0.18
#> MR1  1.00  0.19 1.00  0.00 1.00  0.00  0.97  0.89  0.14  1.00  0.01  0.01  1.00
#> MR2 -0.03  0.97 0.00  1.00 0.00  1.00 -0.14  0.32 -0.87 -0.01  0.97 -0.17 -0.01
#> PA1  1.00  0.19 1.00  0.00 1.00  0.00  0.97  0.89  0.15  1.00  0.01  0.01  1.00
#> PA2 -0.03  0.98 0.00  1.00 0.00  1.00 -0.14  0.32 -0.87 -0.01  0.98 -0.17 -0.01
#> ML1  0.98  0.03 0.97 -0.14 0.97 -0.14  1.00  0.76  0.19  0.98 -0.17 -0.14  0.98
#> ML2  0.88  0.51 0.89  0.32 0.89  0.32  0.76  1.00 -0.01  0.88  0.39  0.26  0.88
#> ML3  0.16 -0.79 0.14 -0.87 0.15 -0.87  0.19 -0.01  1.00  0.15 -0.78  0.61  0.15
#> PA1  1.00  0.18 1.00 -0.01 1.00 -0.01  0.98  0.88  0.15  1.00  0.00  0.00  1.00
#> PA2 -0.02  0.97 0.01  0.97 0.01  0.98 -0.17  0.39 -0.78  0.00  1.00  0.00  0.00
#> PA3  0.00 -0.09 0.01 -0.17 0.01 -0.17 -0.14  0.26  0.61  0.00  0.00  1.00  0.00
#> MR1  1.00  0.18 1.00 -0.01 1.00 -0.01  0.98  0.88  0.15  1.00  0.00  0.00  1.00
#> MR2 -0.02  0.97 0.01  0.96 0.01  0.97 -0.17  0.40 -0.75  0.00  1.00  0.04  0.00
#> MR3  0.01 -0.14 0.01 -0.22 0.02 -0.22 -0.13  0.24  0.64  0.00 -0.04  1.00  0.00
#>       MR2   MR3
#> ML1 -0.02  0.01
#> ML2  0.97 -0.14
#> MR1  0.01  0.01
#> MR2  0.96 -0.22
#> PA1  0.01  0.02
#> PA2  0.97 -0.22
#> ML1 -0.17 -0.13
#> ML2  0.40  0.24
#> ML3 -0.75  0.64
#> PA1  0.00  0.00
#> PA2  1.00 -0.04
#> PA3  0.04  1.00
#> MR1  0.00  0.00
#> MR2  1.00  0.00
#> MR3  0.00  1.00

#f5 <- fa(bfi[1:25],5)
#f5  #names are not in ascending numerical order (see note)
#colnames(f5$loadings) <- paste("F",1:5,sep="")
#f5