Create dummy coded variables

Given a variable x with n distinct values, create n new dummy coded variables coded 0/1 for presence (1) or absence (0) of each variable. A typical application would be to create dummy coded college majors from a vector of college majors.

dummy.code(x)

Arguments

x
A vector to be transformed into dummy codes

Details

When coding demographic information, it is typical to create one variable with multiple categorical values (e.g., ethnicity, college major, occupation). dummy.code will convert these categories into n distinct dummy coded variables.

If using dummy coded variables as predictors, remember to use n-1 variables.

Value

A matrix of dummy coded variables

Examples

new <- dummy.code(sat.act$education)
new.sat <- data.frame(new,sat.act)
round(cor(new.sat,use="pairwise"),2)

#>              X0    X1    X2    X3    X4    X5 gender education   age   ACT
#> X0         1.00 -0.08 -0.08 -0.24 -0.15 -0.15  -0.08     -0.66 -0.27 -0.07
#> X1        -0.08  1.00 -0.07 -0.21 -0.13 -0.13  -0.05     -0.40 -0.17 -0.06
#> X2        -0.08 -0.07  1.00 -0.21 -0.13 -0.13  -0.09     -0.21  0.05 -0.08
#> X3        -0.24 -0.21 -0.21  1.00 -0.40 -0.40   0.10     -0.09 -0.39 -0.04
#> X4        -0.15 -0.13 -0.13 -0.40  1.00 -0.25  -0.02      0.29  0.24  0.07
#> X5        -0.15 -0.13 -0.13 -0.40 -0.25  1.00   0.03      0.65  0.49  0.11
#> gender    -0.08 -0.05 -0.09  0.10 -0.02  0.03   1.00      0.09 -0.02 -0.04
#> education -0.66 -0.40 -0.21 -0.09  0.29  0.65   0.09      1.00  0.55  0.15
#> age       -0.27 -0.17  0.05 -0.39  0.24  0.49  -0.02      0.55  1.00  0.11
#> ACT       -0.07 -0.06 -0.08 -0.04  0.07  0.11  -0.04      0.15  0.11  1.00
#> SATV       0.01 -0.03 -0.08  0.00  0.02  0.04  -0.02      0.05 -0.04  0.56
#> SATQ       0.03 -0.01 -0.07 -0.03  0.01  0.06  -0.17      0.03 -0.03  0.59
#>            SATV  SATQ
#> X0         0.01  0.03
#> X1        -0.03 -0.01
#> X2        -0.08 -0.07
#> X3         0.00 -0.03
#> X4         0.02  0.01
#> X5         0.04  0.06
#> gender    -0.02 -0.17
#> education  0.05  0.03
#> age       -0.04 -0.03
#> ACT        0.56  0.59
#> SATV       1.00  0.64
#> SATQ       0.64  1.00