Eysenck Personality Inventory (EPI) data for 3570 participants

The EPI is and has been a very frequently administered personality test with 57 measuring two broad dimensions, Extraversion-Introversion and Stability-Neuroticism, with an additional Lie scale. Developed by Eysenck and Eysenck, 1964. Eventually replaced with the EPQ which measures three broad dimensions. This data set represents 3570 observations collected in the early 1990s at the Personality, Motivation and Cognition lab at Northwestern. The data are included here as demonstration of scale construction.

data(epi)
data(epi.dictionary)

Format

A data frame with 3570 observations on the following 57 variables.

V1

a numeric vector

V2

a numeric vector

V3

a numeric vector

V4

a numeric vector

V5

a numeric vector

V6

a numeric vector

V7

a numeric vector

V8

a numeric vector

V9

a numeric vector

V10

a numeric vector

V11

a numeric vector

V12

a numeric vector

V13

a numeric vector

V14

a numeric vector

V15

a numeric vector

V16

a numeric vector

V17

a numeric vector

V18

a numeric vector

V19

a numeric vector

V20

a numeric vector

V21

a numeric vector

V22

a numeric vector

V23

a numeric vector

V24

a numeric vector

V25

a numeric vector

V26

a numeric vector

V27

a numeric vector

V28

a numeric vector

V29

a numeric vector

V30

a numeric vector

V31

a numeric vector

V32

a numeric vector

V33

a numeric vector

V34

a numeric vector

V35

a numeric vector

V36

a numeric vector

V37

a numeric vector

V38

a numeric vector

V39

a numeric vector

V40

a numeric vector

V41

a numeric vector

V42

a numeric vector

V43

a numeric vector

V44

a numeric vector

V45

a numeric vector

V46

a numeric vector

V47

a numeric vector

V48

a numeric vector

V49

a numeric vector

V50

a numeric vector

V51

a numeric vector

V52

a numeric vector

V53

a numeric vector

V54

a numeric vector

V55

a numeric vector

V56

a numeric vector

V57

a numeric vector

Details

The original data were collected in a group testing framework for screening participants for subsequent studies. The participants were enrolled in an introductory psychology class between Fall, 1991 and Spring, 1995.

The structure of the E scale has been shown by Rocklin and Revelle (1981) to have two subcomponents, Impulsivity and Sociability. These were subsequently used by Revelle, Humphreys, Simon and Gilliland to examine the relationship between personality, caffeine induced arousal, and cognitive performance.

Source

Data from the PMC laboratory at Northwestern.

References

Eysenck, H.J. and Eysenck, S. B.G. (1968). Manual for the Eysenck Personality Inventory.Educational and Industrial Testing Service, San Diego, CA.

Rocklin, T. and Revelle, W. (1981). The measurement of extraversion: A comparison of the Eysenck Personality Inventory and the Eysenck Personality Questionnaire. British Journal of Social Psychology, 20(4):279-284.

Examples

data(epi)
epi.keys <- make.keys(epi,list(E = c(1, 3, -5, 8, 10, 13, -15, 17, -20, 22, 25, 27,
                -29, -32, -34, -37, 39, -41, 44, 46, 49, -51, 53, 56),
   N=c(2, 4, 7, 9, 11, 14, 16, 19, 21, 23, 26, 28, 31, 33, 35, 38, 40,
     43, 45, 47, 50, 52, 55, 57),
   L = c(6, -12, -18, 24, -30, 36, -42, -48, -54),
   I =c(1, 3, -5, 8, 10, 13, 22, 39, -41),
   S = c(-11, -15, 17, -20, 25, 27, -29, -32, -37, 44, 46, -51, 53)))
scores <- scoreItems(epi.keys,epi)
  N <- epi[abs(epi.keys[,"N"]) >0]
  E <- epi[abs(epi.keys[,"E"]) >0]
  fa.lookup(epi.keys[,1:3],epi.dictionary) #show the items and keying information

#>      E N  L
#> V1   1 0  0
#> V3   1 0  0
#> V5  -1 0  0
#> V8   1 0  0
#> V10  1 0  0
#> V13  1 0  0
#> V15 -1 0  0
#> V17  1 0  0
#> V20 -1 0  0
#> V22  1 0  0
#> V25  1 0  0
#> V27  1 0  0
#> V29 -1 0  0
#> V32 -1 0  0
#> V34 -1 0  0
#> V37 -1 0  0
#> V39  1 0  0
#> V41 -1 0  0
#> V44  1 0  0
#> V46  1 0  0
#> V49  1 0  0
#> V51 -1 0  0
#> V53  1 0  0
#> V56  1 0  0
#> V2   0 1  0
#> V4   0 1  0
#> V7   0 1  0
#> V9   0 1  0
#> V11  0 1  0
#> V14  0 1  0
#> V16  0 1  0
#> V19  0 1  0
#> V21  0 1  0
#> V23  0 1  0
#> V26  0 1  0
#> V28  0 1  0
#> V31  0 1  0
#> V33  0 1  0
#> V35  0 1  0
#> V38  0 1  0
#> V40  0 1  0
#> V43  0 1  0
#> V45  0 1  0
#> V47  0 1  0
#> V50  0 1  0
#> V52  0 1  0
#> V55  0 1  0
#> V57  0 1  0
#> V6   0 0  1
#> V12  0 0 -1
#> V18  0 0 -1
#> V24  0 0  1
#> V30  0 0 -1
#> V36  0 0  1
#> V42  0 0 -1
#> V48  0 0 -1
#> V54  0 0 -1
#>                                                                                                               Content
#> V1                                                                                  Do you often long for excitement?
#> V3                                                                                          Are you usually carefree?
#> V5                                                           Do you stop and think things over before doing anything?
#> V8                                              Do you generally do and say things quickly without stopping to think?
#> V10                                                                          Would you do almost anything for a dare?
#> V13                                                                 Do you often do things on the spur of the moment?
#> V15                                                                Generally do you prefer reading to meeting people?
#> V17                                                                                      Do you like going out a lot?
#> V20                                                                    Do you prefer to have few but special friends?
#> V22                                                                       When people shout at you do you shout back?
#> V25                                       Can you usually let yourself go and enjoy yourself a lot at a lively party?
#> V27                                                                Do other people think of you as being very lively?
#> V29                                                              Are you mostly quiet when you are with other people?
#> V32 If there is something you want to know about, would you rather look it upin a book than talk to someone about it?
#> V34                                             Do you like the kind of work that you need to pay close attention to?
#> V37                                                     Do you hate being with a crowd who play jokes on one another?
#> V39                                                        Do you like doing things in which you have to act quickly?
#> V41                                                                   Are you slow and unhurried in the way you move?
#> V44                      Do you like talking to people so much that you never miss a chance of talking to a stranger?
#> V46                                   Would you be very unhappy if you could not see lots of people most of the time?
#> V49                                                                Would you say that you were fairly self-confident?
#> V51                                                   Do you find it hard to really enjoy yourself at a lively party?
#> V53                                                                   Can you easily get some life into a dull party?
#> V56                                                                             Do you like playing pranks on others?
#> V2                                                           Do you often need understanding friends to cheer you up?
#> V4                                                                 Do you find it very hard to take no for an answer?
#> V7                                                                                      Do your moods go up and down?
#> V9                                                                Do you ever feel just miserable for no good reason?
#> V11                                         Do you suddenly feel shy when you want to talk to an attractive stranger?
#> V14                                                     Do you often worry about things you should have done or said?
#> V16                                                                             Are your feelings rather easily hurt?
#> V19                                          Are you sometimes bubbling over with energy and sometimes very sluggish?
#> V21                                                                                            Do you daydream a lot?
#> V23                                                                   Are you often troubled about feelings of guilt?
#> V26                                                                   Would you call yourself tense or highly strung?
#> V28                      After you have done something important, do you come away feelingyou could have done better?
#> V31                                                          Do ideas run through your head so that you cannot sleep?
#> V33                                                                 Do you get palpitations or thumping in your hear?
#> V35                                                                       Do you get attacks of shaking or trembling?
#> V38                                                                                      Are you an irritable person?
#> V40                                                                Do you worry about awful things that might happen?
#> V43                                                                                      Do you have many nightmares?
#> V45                                                                              Are you troubled by aches and pains?
#> V47                                                                         Would you call yourself a nervous person?
#> V50                                                 Are you easily hurt when people find fault with you or your work?
#> V52                                                                      Are you troubled by feelings of inferiority?
#> V55                                                                                   Do you worry about your health?
#> V57                                                                                 Do you suffer from sleeplessness?
#> V6  If you say you will do something do you always keep your promise,no matter how inconvenient it might be to do so?
#> V12                                                            Once in a while do you lose your temper and get angry?
#> V18                    Do you occasionally have thoughts and ideas that you would not like otherpeople to know about?
#> V24                                                                      Are all your habits good and desirable ones?
#> V30                                                                                          Do you sometimes gossip?
#> V36                     Would you always declare everything at customs, even if you knewyou could never be found out?
#> V42                                                               Have you ever been late for an appointment or work?
#> V48                                       Of all the people you know, are there some whom you definitely do not like?
#> V54                                                        Do you sometimes talk about things you know nothing about?