Sort factor analysis or principal components analysis loadings

Although the print.psych function will sort factor analysis loadings, sometimes it is useful to do this outside of the print function. fa.sort takes the output from the fa or principal functions and sorts the loadings for each factor. Items are located in terms of their greatest loading. The new order is returned as an element in the fa list.

fa.sort(fa.results,polar=FALSE)
fa.organize(fa.results,o=NULL,i=NULL,cn=NULL)

Arguments

fa.results
The output from a factor analysis or principal components analysis using fa or principal.

polar
Sort by polar coordinates of first two factors (FALSE)

o
The order in which to order the factors

i
The order in which to order the items

cn
new factor names

Details

The fa.results$loadings are replaced with sorted loadings.

fa.organize takes a factor analysis or components output and reorganizes the factors in the o order. Items are organized in the i order. This is useful when comparing alternative factor solutions.

Value

A sorted factor analysis, principal components analysis, or omega loadings matrix.

These sorted values are used internally by the various diagram functions.

The values returned are the same as fa, except in sorted order. In addition, the order is returned as an additional element in the fa list.

See also

See Also as fa,print.psych, fa.diagram,

Examples

test.simple <- fa(sim.item(16),2)
fa.sort(test.simple)

#> Factor Analysis using method =  minres
#> Call: fa(r = sim.item(16), nfactors = 2)
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>       MR1   MR2   h2   u2 com
#> V4   0.67 -0.01 0.46 0.54 1.0
#> V11 -0.65 -0.04 0.42 0.58 1.0
#> V3   0.63 -0.05 0.41 0.59 1.0
#> V12 -0.63 -0.01 0.39 0.61 1.0
#> V10 -0.62 -0.03 0.38 0.62 1.0
#> V2   0.62  0.01 0.38 0.62 1.0
#> V1   0.60 -0.02 0.36 0.64 1.0
#> V9  -0.55  0.03 0.31 0.69 1.0
#> V5  -0.03  0.63 0.40 0.60 1.0
#> V6   0.10  0.62 0.38 0.62 1.1
#> V14  0.03 -0.61 0.37 0.63 1.0
#> V8   0.00  0.60 0.36 0.64 1.0
#> V15  0.00 -0.56 0.31 0.69 1.0
#> V16  0.03 -0.54 0.30 0.70 1.0
#> V7  -0.01  0.53 0.28 0.72 1.0
#> V13  0.05 -0.51 0.26 0.74 1.0
#> 
#>                        MR1  MR2
#> SS loadings           3.12 2.67
#> Proportion Var        0.19 0.17
#> Cumulative Var        0.19 0.36
#> Proportion Explained  0.54 0.46
#> Cumulative Proportion 0.54 1.00
#> 
#>  With factor correlations of 
#>      MR1  MR2
#> MR1  1.0 -0.1
#> MR2 -0.1  1.0
#> 
#> Mean item complexity =  1
#> Test of the hypothesis that 2 factors are sufficient.
#> 
#> The degrees of freedom for the null model are  120  and the objective function was  3.98 with Chi Square of  1961.97
#> The degrees of freedom for the model are 89  and the objective function was  0.18 
#> 
#> The root mean square of the residuals (RMSR) is  0.02 
#> The df corrected root mean square of the residuals is  0.03 
#> 
#> The harmonic number of observations is  500 with the empirical chi square  72.16  with prob <  0.9 
#> The total number of observations was  500  with Likelihood Chi Square =  86.12  with prob <  0.57 
#> 
#> Tucker Lewis Index of factoring reliability =  1.002
#> RMSEA index =  0  and the 90 % confidence intervals are  0 0.023
#> BIC =  -466.98
#> Fit based upon off diagonal values = 0.99
#> Measures of factor score adequacy             
#>                                                 MR1  MR2
#> Correlation of scores with factors             0.92 0.90
#> Multiple R square of scores with factors       0.84 0.80
#> Minimum correlation of possible factor scores  0.68 0.61

fa.organize(test.simple,c(2,1))  #the factors but not the items have been rearranged

#> Factor Analysis using method =  minres
#> Call: fa(r = sim.item(16), nfactors = 2)
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>       MR2   MR1   h2   u2 com
#> V1  -0.02  0.60 0.36 0.64 1.0
#> V2   0.01  0.62 0.38 0.62 1.0
#> V3  -0.05  0.63 0.41 0.59 1.0
#> V4  -0.01  0.67 0.46 0.54 1.0
#> V5   0.63 -0.03 0.40 0.60 1.0
#> V6   0.62  0.10 0.38 0.62 1.1
#> V7   0.53 -0.01 0.28 0.72 1.0
#> V8   0.60  0.00 0.36 0.64 1.0
#> V9   0.03 -0.55 0.31 0.69 1.0
#> V10 -0.03 -0.62 0.38 0.62 1.0
#> V11 -0.04 -0.65 0.42 0.58 1.0
#> V12 -0.01 -0.63 0.39 0.61 1.0
#> V13 -0.51  0.05 0.26 0.74 1.0
#> V14 -0.61  0.03 0.37 0.63 1.0
#> V15 -0.56  0.00 0.31 0.69 1.0
#> V16 -0.54  0.03 0.30 0.70 1.0
#> 
#>                        MR2  MR1
#> SS loadings           2.67 3.12
#> Proportion Var        0.17 0.19
#> Cumulative Var        0.17 0.36
#> Proportion Explained  0.46 0.54
#> Cumulative Proportion 0.46 1.00
#> 
#>  With factor correlations of 
#>      MR2  MR1
#> MR2  1.0 -0.1
#> MR1 -0.1  1.0
#> 
#> Mean item complexity =  1
#> Test of the hypothesis that 2 factors are sufficient.
#> 
#> The degrees of freedom for the null model are  120  and the objective function was  3.98 with Chi Square of  1961.97
#> The degrees of freedom for the model are 89  and the objective function was  0.18 
#> 
#> The root mean square of the residuals (RMSR) is  0.02 
#> The df corrected root mean square of the residuals is  0.03 
#> 
#> The harmonic number of observations is  500 with the empirical chi square  72.16  with prob <  0.9 
#> The total number of observations was  500  with Likelihood Chi Square =  86.12  with prob <  0.57 
#> 
#> Tucker Lewis Index of factoring reliability =  1.002
#> RMSEA index =  0  and the 90 % confidence intervals are  0 0.023
#> BIC =  -466.98
#> Fit based upon off diagonal values = 0.99
#> Measures of factor score adequacy             
#>                                                 MR2  MR1
#> Correlation of scores with factors             0.90 0.92
#> Multiple R square of scores with factors       0.80 0.84
#> Minimum correlation of possible factor scores  0.61 0.68