Find R = F F' + U2 is the basic factor model

The basic factor or principal components model is that a correlation or covariance matrix may be reproduced by the product of a factor loading matrix times its transpose. Find this reproduced matrix. Used by factor.fit, VSS, ICLUST, etc.

factor.model(f,Phi=NULL,U2=TRUE)

Arguments

f
A matrix of loadings.

Phi
A matrix of factor correlations

U2
Should the diagonal be model by ff' (U2 = TRUE) or replaced with 1's (U2 = FALSE)

Value

A correlation or covariance matrix.

References

Gorsuch, Richard, (1983) Factor Analysis. Lawrence Erlebaum Associates. Revelle, W. In preparation) An Introduction to Psychometric Theory with applications in R (http://personality-project.org/r/book/)

See also

ICLUST.graph,ICLUST.cluster, cluster.fit , VSS, omega

Examples

f2 <- matrix(c(.9,.8,.7,rep(0,6),.6,.7,.8),ncol=2)
mod <- factor.model(f2)
round(mod,2)

#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,] 0.81 0.72 0.63 0.00 0.00 0.00
#> [2,] 0.72 0.64 0.56 0.00 0.00 0.00
#> [3,] 0.63 0.56 0.49 0.00 0.00 0.00
#> [4,] 0.00 0.00 0.00 0.36 0.42 0.48
#> [5,] 0.00 0.00 0.00 0.42 0.49 0.56
#> [6,] 0.00 0.00 0.00 0.48 0.56 0.64