Various ways to estimate factor scores for the factor analysis model

A fundamental problem with factor analysis is that although the model is defined at the structural level, it is indeterminate at the data level. This problem of factor indeterminancy leads to alternative ways of estimating factor scores, none of which is ideal. Following Grice (2001) four different methods are available here.

factor.scores(x, f, Phi = NULL, method = c("Thurstone", "tenBerge", "Anderson",
       "Bartlett", "Harman","components"),rho=NULL,impute="none")

Arguments

x
Either a matrix of data if scores are to be found, or a correlation matrix if just the factor weights are to be found.

f
The output from the fa or irt.fa functions, or a factor loading matrix.

Phi
If a pattern matrix is provided, then what were the factor intercorrelations. Does not need to be specified if f is the output from the fa or irt.fa functions.

method
Which of four factor score estimation procedures should be used. Defaults to "Thurstone" or regression based weights. See details below for the other four methods.

rho
If x is a set of data and rho is specified, then find scores based upon the fa results and the correlations reported in rho. Used when scoring fa.poly results.

impute
By default, only complete cases are scored. But, missing data can be imputed using "median" or "mean". The number of missing by subject is reported.

Details

Although the factor analysis model is defined at the structural level, it is undefined at the data level. This is a well known but little discussed problem with factor analysis.

Factor scores represent estimates of common part of the variables and should not be thought of as identical to the factors themselves. If a factor is thought of as a chop stick stuck into the center of an ice cream cone and factor score estimates are represented by straws anywhere along the edge of the cone the problem of factor indeterminacy becomes clear, for depending on the shape of the cone, two straws can be negatively correlated with each other. (The imagery is taken from Niels Waller, adapted from Stanley Mulaik). In a very clear discussion of the problem of factor score indeterminacy, Grice (2001) reviews several alternative ways of estimating factor scores and considers weighting schemes that will produce uncorrelated factor score estimates as well as the effect of using course coded (unit weighted) factor weights. factor.scores uses four different ways of estimate factor scores. In all cases, the factor score estimates are based upon the data matrix, X, times a weighting matrix, W, which weights the observed variables.

For polytomous or dichotmous data, factor scores can be estimated using Item Response Theory techniques (e.g., using link{irt.fa} and then link{scoreIrt}. Such scores are still just factor score estimates, for the IRT model is a latent variable model equivalent to factor analysis.

method="Thurstone" finds the regression based weights: \(W = R^{-1} F\) where R is the correlation matrix and F is the factor loading matrix.

method="tenBerge" finds weights such that the correlation between factors for an oblique solution is preserved. Note that formula 8 in Grice has a typo in the formula for C and should be: \(L = F \Phi^(1/2) \) \(C = R^(-1/2) L (L' R^(-1) L)^(-1/2) \) \(W = R ^(-1/2) C \Phi^(1/2) \)

method="Anderson" finds weights such that the factor scores will be uncorrelated: \(W = U^{-2}F (F' U^{-2} R U^{-2} F)^{-1/2}\) where U is the diagonal matrix of uniquenesses. The Anderson method works for orthogonal factors only, while the tenBerge method works for orthogonal or oblique solutions.

method = "Bartlett" finds weights given \(W = U^{-2}F (F' U^{-2}F)^{-1}\)

method="Harman" finds weights based upon socalled "idealized" variables: \(W = F (t(F) F)^{-1}\).

method="components" uses weights that are just component loadings.

Value

scores (the factor scores if the raw data is given)

weights (the factor weights)

r.scores (The correlations of the factor score estimates.)

missing A vector of the number of missing observations per subject

References

Grice, James W.,2001, Computing and evaluating factor scores, Psychological Methods, 6,4, 430-450. (note the typo in equation 8)

ten Berge, Jos M.F., Wim P. Krijnen, Tom Wansbeek and Alexander Shapiro (1999) Some new results on correlation-preserving factor scores prediction methods. Linear Algebra and its Applications, 289, 311-318.

Revelle, William. (in prep) An introduction to psychometric theory with applications in R. Springer. Working draft available at http://personality-project.org/r/book/

See also

fa, factor.stats

Examples

f3 <- fa(Thurstone)
f3$weights  #just the scoring weights

#>                          MR1
#> Sentences         0.18680979
#> Vocabulary        0.23056715
#> Sent.Completion   0.15762958
#> First.Letters     0.13887845
#> Four.Letter.Words 0.13244727
#> Suffixes          0.09233885
#> Letter.Series     0.10932454
#> Pedigrees         0.11176173
#> Letter.Group      0.10401145

f5 <- fa(bfi,5)
round(cor(f5$scores,use="pairwise"),2)

#>       MR2   MR1   MR3   MR5   MR4
#> MR2  1.00 -0.23 -0.23 -0.05 -0.02
#> MR1 -0.23  1.00  0.33  0.44  0.25
#> MR3 -0.23  0.33  1.00  0.24  0.25
#> MR5 -0.05  0.44  0.24  1.00  0.19
#> MR4 -0.02  0.25  0.25  0.19  1.00

#compare to the f5 solution