factor2cluster {psych}	R Documentation

Extract cluster definitions from factor loadings

Description

Given a factor or principal components loading matrix, assign each item to a cluster corresponding to the largest (signed) factor loading for that item. Essentially, this is a Very Simple Structure approach to cluster definition that corresponds to what most people actually do: highlight the largest loading for each item and ignore the rest.

Usage

factor2cluster(loads, cut = 0,aslist=FALSE)

Arguments

loads
either a matrix of loadings, or the result of a factor analysis/principal components analyis with a loading component

cut
Extract items with absolute loadings > cut

aslist
if TRUE, Return a keys list, else return a keys matrix (old style)

Details

A factor/principal components analysis loading matrix is converted to a cluster (-1,0,1) definition matrix where each item is assigned to one and only one cluster. This is a fast way to extract items that will be unit weighted to form cluster composites. Use this function in combination with cluster.cor to find the corrleations of these composite scores.

A typical use in the SAPA project is to form item composites by clustering or factoring (see ICLUST, principal), extract the clusters from these results (factor2cluster), and then form the composite correlation matrix using cluster.cor. The variables in this reduced matrix may then be used in multiple R procedures using mat.regress.

The input may be a matrix of item loadings, or the output from a factor analysis which includes a loadings matrix.

Value

a keys list (new style or a matrix of -1,0,1 cluster definitions for each item.

Author(s)

https://personality-project.org/revelle.html

Maintainer: William Revelle revelle@northwestern.edu

References

https://personality-project.org/r/r.vss.html

See Also

cluster.cor, factor2cluster, fa, principal, ICLUST make.keys, keys2list

Examples

#matches the factanal example
f4  <- fa(Harman74.cor$cov,4,rotate="varimax")
factor2cluster(f4) 

[Package psych version 1.9.12 Index]