Transformations of r including Fisher r to z and z to r and confidence intervals

Convert a correlation to a z or t, or d, or chi or covariance matrix or z to r using the Fisher transformation or find the confidence intervals for a specified correlation. r2d converts a correlation to an effect size (Cohen's d) and d2r converts a d into an r. g2r converts Hedge's g to a correlation. t2r converts a t test to r, r2t converts a correlation to a t-test value. chi2r converts a chi square to r, r2chi converts it back. r2c and cor2cov convert a correlation matrix to a covariance matrix.

fisherz(rho)
fisherz2r(z)
r.con(rho,n,p=.95,twotailed=TRUE)
r2t(rho,n)
r2d(rho)
d2r(d)
t2r(t,df)
g2r(g,df,n)
chi2r(chi2,n)
r2chi(rho,n)
r2c(rho,sigma)
cor2cov(rho,sigma)

Arguments

rho
a Pearson r

z
A Fisher z

n
Sample size for confidence intervals

df
degrees of freedom for t, or g

p
Confidence interval

twotailed
Treat p as twotailed p

d
an effect size (Cohen's d)

g
An effect size (Hedge's g)

t
A student's t value

chi2
A chi square

sigma
a vector of standard deviations to be used to convert a correlation matrix to a covariance matrix

Value

z

value corresponding to r (fisherz)

r

r corresponding to z (fisherz2r)

r.con

lower and upper p confidence intervals (r.con)

t

t with n-2 df (r2t)

r

r corresponding to effect size d or d corresponding to r.

r2c

r2c is the reverse of the cor2con function of base R. It just converts a correlation matrix to the corresponding covariance matrix given a vector of standard deviations.

Examples

 n <- 30
 r <- seq(0,.9,.1)
 d <- r2d(r)
 rc <- matrix(r.con(r,n),ncol=2)
 t <- r*sqrt(n-2)/sqrt(1-r^2)
 p <- (1-pt(t,n-2))*2
 r1 <- t2r(t,(n-2))
 r2 <- d2r(d)
 chi <- r2chi(r,n)
 r3 <- chi2r(chi,n)
 r.rc <- data.frame(r=r,z=fisherz(r),lower=rc[,1],upper=rc[,2],t=t,p=p,d=d,
     chi2=chi,d2r=r2,t2r=r1,chi2r=r3)
 round(r.rc,2)

#>      r    z lower upper     t    p    d chi2 d2r t2r chi2r
#> 1  0.0 0.00 -0.36  0.36  0.00 1.00 0.00  0.0 0.0 0.0   0.0
#> 2  0.1 0.10 -0.27  0.44  0.53 0.60 0.20  0.3 0.1 0.1   0.1
#> 3  0.2 0.20 -0.17  0.52  1.08 0.29 0.41  1.2 0.2 0.2   0.2
#> 4  0.3 0.31 -0.07  0.60  1.66 0.11 0.63  2.7 0.3 0.3   0.3
#> 5  0.4 0.42  0.05  0.66  2.31 0.03 0.87  4.8 0.4 0.4   0.4
#> 6  0.5 0.55  0.17  0.73  3.06 0.00 1.15  7.5 0.5 0.5   0.5
#> 7  0.6 0.69  0.31  0.79  3.97 0.00 1.50 10.8 0.6 0.6   0.6
#> 8  0.7 0.87  0.45  0.85  5.19 0.00 1.96 14.7 0.7 0.7   0.7
#> 9  0.8 1.10  0.62  0.90  7.06 0.00 2.67 19.2 0.8 0.8   0.8
#> 10 0.9 1.47  0.80  0.95 10.93 0.00 4.13 24.3 0.9 0.9   0.9