phi2tetra {psych}	R Documentation

Convert a phi coefficient to a tetrachoric correlation

Description

Given a phi coefficient (a Pearson r calculated on two dichotomous variables), and the marginal frequencies (in percentages), what is the corresponding estimate of the tetrachoric correlation?

Given a two x two table of counts

a 	 b

c 	 d

The phi coefficient is (a - (a+b)*(a+c))/sqrt((a+b)(a+c)(b+d)(c+c)).

This function reproduces the cell entries for specified marginals and then calls the tetrachoric function. (Which was originally based upon John Fox's polychor function.) The phi2poly name will become deprecated in the future.

Usage

phi2tetra(ph,m,n=NULL,correct=TRUE)
phi2poly(ph,cp,cc,n=NULL,correct=TRUE) #deprecated

Arguments

ph
phi

m
a vector of the selection ratio and probability of criterion. In the case where ph is a matrix, m is a vector of the frequencies of the selected cases

correct
When finding tetrachoric correlations, should we correct for continuity for small marginals. See tetrachoric for a discussion.

n
If the marginals are given as frequencies, what was the total number of cases?

cp
probability of the predictor – the so called selection ratio

cc
probability of the criterion – the so called success rate.

Details

used to require the mvtnorm package but this has been replaced with mnormt

Value

a tetrachoric correlation

Author(s)

William Revelle

See Also

tetrachoric, Yule2phi.matrix, phi2poly.matrix

Examples

 phi2tetra(.3,c(.5,.5))
#phi2poly(.3,.3,.7)

[Package psych version 1.9.12 Index]