Tests of significance for correlations

Tests the significance of a single correlation, the difference between two independent correlations, the difference between two dependent correlations sharing one variable (Williams's Test), or the difference between two dependent correlations with different variables (Steiger Tests).

r.test(n, r12, r34 = NULL, r23 = NULL, r13 = NULL, r14 = NULL, r24 = NULL,
          n2 = NULL,pooled=TRUE, twotailed = TRUE)

Arguments

n
Sample size of first group

r12
Correlation to be tested

r34
Test if this correlation is different from r12, if r23 is specified, but r13 is not, then r34 becomes r13

r23
if ra = r(12) and rb = r(13) then test for differences of dependent correlations given r23

r13
implies ra =r(12) and rb =r(34) test for difference of dependent correlations

r14
implies ra =r(12) and rb =r(34)

r24
ra =r(12) and rb =r(34)

n2
n2 is specified in the case of two independent correlations. n2 defaults to n if if not specified

pooled
use pooled estimates of correlations

twotailed
should a twotailed or one tailed test be used

Details

Depending upon the input, one of four different tests of correlations is done. 1) For a sample size n, find the t value for a single correlation where $$t = \frac{r * \sqrt(n-2)}{\sqrt(1-r^2)} $$ and $$se = \sqrt{\frac{1-r^2}{n-2}}) $$.

2) For sample sizes of n and n2 (n2 = n if not specified) find the z of the difference between the z transformed correlations divided by the standard error of the difference of two z scores: $$z = \frac{z_1 - z_2}{\sqrt{\frac{1}{(n_1 - 3) + (n_2 - 3)}}}$$.

3) For sample size n, and correlations r12, r13 and r23 test for the difference of two dependent correlations (r12 vs r13).

4) For sample size n, test for the difference between two dependent correlations involving different variables.

For clarity, correlations may be specified by value. If specified by location and if doing the test of dependent correlations, if three correlations are specified, they are assumed to be in the order r12, r13, r23. Consider the example the example from Steiger: where Masculinity at time 1 (M1) correlates with Verbal Ability .5 (r12), femininity at time 1 (F1) correlates with Verbal ability r13 =.4, and M1 correlates with F1 (r23= .1). Then, given the correlations: r12 = .4, r13 = .5, and r23 = .1, t = -.89 for n =103, i.e., r.test(n=103, r12=.4, r13=.5,r23=.1)

Value

test

Label of test done

z

z value for tests 2 or 4

t

t value for tests 1 and 3

p

probability value of z or t

References

Cohen, J. and Cohen, P. and West, S.G. and Aiken, L.S. (2003) Applied multiple regression/correlation analysis for the behavioral sciences, L.Erlbaum Associates, Mahwah, N.J.

Olkin, I. and Finn, J. D. (1995). Correlations redux. Psychological Bulletin, 118(1):155-164.

Steiger, J.H. (1980), Tests for comparing elements of a correlation matrix, Psychological Bulletin, 87, 245-251.

Williams, E.J. (1959) Regression analysis. Wiley, New York, 1959.

Note

Steiger specifically rejects using the Hotelling T test to test the difference between correlated correlations. Instead, he recommends Williams' test. (See also Dunn and Clark, 1971). These tests follow Steiger's advice. The test of two independent correlations is just a z test of the difference of the Fisher's z transformed correlations divided by the standard error of the difference. (See Cohen et al, p 49).

See also

See also corr.test which tests all the elements of a correlation matrix, and cortest.mat to compare two matrices of correlations. r.test extends the tests in paired.r,r.con

Examples

n <- 30
r <- seq(0,.9,.1)
rc <- matrix(r.con(r,n),ncol=2)
test <- r.test(n,r)
r.rc <- data.frame(r=r,z=fisherz(r),lower=rc[,1],upper=rc[,2],t=test$t,p=test$p)
round(r.rc,2)

#>      r    z lower upper     t    p
#> 1  0.0 0.00 -0.36  0.36  0.00 1.00
#> 2  0.1 0.10 -0.27  0.44  0.53 0.60
#> 3  0.2 0.20 -0.17  0.52  1.08 0.29
#> 4  0.3 0.31 -0.07  0.60  1.66 0.11
#> 5  0.4 0.42  0.05  0.66  2.31 0.03
#> 6  0.5 0.55  0.17  0.73  3.06 0.00
#> 7  0.6 0.69  0.31  0.79  3.97 0.00
#> 8  0.7 0.87  0.45  0.85  5.19 0.00
#> 9  0.8 1.10  0.62  0.90  7.06 0.00
#> 10 0.9 1.47  0.80  0.95 10.93 0.00

r.test(50,r)

#> Correlation tests 
#> Call:r.test(n = 50, r12 = r)
#> Test of significance of a  correlation 
#>  t value 0 0.7 1.41 2.18 3.02 4 5.2 6.79 9.24 14.3    with probability < 1 0.49 0.16 0.034 0.004 0.00022 4.1e-06 1.5e-08 3.2e-12 0
#>  and confidence interval  -0.28 -0.18 -0.08 0.02 0.14 0.26 0.39 0.52 0.67 0.83 0.28 0.37 0.45 0.53 0.61 0.68 0.75 0.82 0.88 0.94

r.test(30,.4,.6)       #test the difference between two independent correlations

#> Correlation tests 
#> Call:r.test(n = 30, r12 = 0.4, r34 = 0.6)
#> Test of difference between two independent correlations 
#>  z value 0.99    with probability  0.32

r.test(103,.4,.5,.1)   #Steiger case A of dependent correlations 

#> Correlation tests 
#> Call:[1] "r.test(n =  103 ,  r12 =  0.4 ,  r23 =  0.1 ,  r13 =  0.5 )"
#> Test of difference between two correlated  correlations 
#>  t value -0.89    with probability < 0.37

r.test(n=103, r12=.4, r13=.5,r23=.1)

#> Correlation tests 
#> Call:[1] "r.test(n =  103 ,  r12 =  0.4 ,  r23 =  0.1 ,  r13 =  0.5 )"
#> Test of difference between two correlated  correlations 
#>  t value -0.89    with probability < 0.37

#for complicated tests, it is probably better to specify correlations by name
r.test(n=103,r12=.5,r34=.6,r13=.7,r23=.5,r14=.5,r24=.8)   #steiger Case B 

#> Correlation tests 
#> Call:r.test(n = 103, r12 = 0.5, r34 = 0.6, r23 = 0.5, r13 = 0.7, r14 = 0.5, 
#>     r24 = 0.8)
#> Test of difference between two dependent correlations 
#>  z value -1.4    with probability  0.16