rescale {psych}	R Documentation

Function to convert scores to “conventional " metrics

Description

Psychologists frequently report data in terms of transformed scales such as “IQ" (mean=100, sd=15, “SAT/GRE" (mean=500, sd=100), “ACT" (mean=18, sd=6), “T-scores" (mean=50, sd=10), or “Stanines" (mean=5, sd=2). The rescale function converts the data to standard scores and then rescales to the specified mean(s) and standard deviation(s).

Usage

rescale(x, mean = 100, sd = 15,df=TRUE)

Arguments

x
A matrix or data frame

mean
Desired mean of the rescaled scores- may be a vector

sd
Desired standard deviation of the rescaled scores

df
if TRUE, returns a data frame, otherwise a matrix

Value

A data.frame (default) or matrix of rescaled scores.

Author(s)

William Revelle

See Also

See Also scale

Examples

T <- rescale(attitude,50,10) #all put on same scale
describe(T)
T1 <- rescale(attitude,seq(0,300,50),seq(10,70,10)) #different means and sigmas
describe(T1)

[Package psych version 1.9.12 Index]