Function to convert scores to ``conventional " metrics

Psychologists frequently report data in terms of transformed scales such as ``IQ" (mean=100, sd=15, ``SAT/GRE" (mean=500, sd=100), ``ACT" (mean=18, sd=6), ``T-scores" (mean=50, sd=10), or ``Stanines" (mean=5, sd=2). The rescale function converts the data to standard scores and then rescales to the specified mean(s) and standard deviation(s).

rescale(x, mean = 100, sd = 15,df=TRUE)

Arguments

x
A matrix or data frame

mean
Desired mean of the rescaled scores- may be a vector

sd
Desired standard deviation of the rescaled scores

df
if TRUE, returns a data frame, otherwise a matrix

Value

A data.frame (default) or matrix of rescaled scores.

See also

See Also scale

Examples

T <- rescale(attitude,50,10) #all put on same scale
describe(T)

#>            vars  n mean sd median trimmed   mad   min   max range  skew
#> rating        1 30   50 10  50.71   50.47  8.53 29.76 66.73 36.97 -0.36
#> complaints    2 30   50 10  48.80   50.36 11.14 27.77 67.57 39.81 -0.22
#> privileges    3 30   50 10  48.67   49.69  8.48 31.09 74.41 43.32  0.38
#> learning      4 30   50 10  50.11   50.18 12.63 30.94 65.88 34.93 -0.05
#> raises        5 30   50 10  48.91   49.87 10.69 29.19 72.47 43.28  0.20
#> critical      6 30   50 10  52.76   51.08  7.49 23.96 67.42 43.46 -0.87
#> advance       7 30   50 10  48.12   48.93  8.65 32.57 78.25 45.68  0.85
#>            kurtosis   se
#> rating        -0.77 1.83
#> complaints    -0.68 1.83
#> privileges    -0.41 1.83
#> learning      -1.22 1.83
#> raises        -0.60 1.83
#> critical       0.17 1.83
#> advance        0.47 1.83

T1 <- rescale(attitude,seq(0,300,50),seq(10,70,10)) #different means and sigmas
describe(T1)

#>            vars  n mean sd median trimmed   mad    min    max  range  skew
#> rating        1 30    0 10   0.71    0.47  8.53 -20.24  16.73  36.97 -0.36
#> complaints    2 30   50 20  47.60   50.73 22.27   5.54  85.15  79.61 -0.22
#> privileges    3 30  100 30  96.00   99.06 25.45  43.28 173.23 129.95  0.38
#> learning      4 30  150 40 150.45  150.74 50.53  73.77 213.50 139.73 -0.05
#> raises        5 30  200 50 194.55  199.36 53.47  95.97 312.37 216.40  0.20
#> critical      6 30  250 60 266.57  256.47 44.95  93.76 354.50 260.74 -0.87
#> advance       7 30  300 70 286.85  292.52 60.52 177.99 497.76 319.77  0.85
#>            kurtosis    se
#> rating        -0.77  1.83
#> complaints    -0.68  3.65
#> privileges    -0.41  5.48
#> learning      -1.22  7.30
#> raises        -0.60  9.13
#> critical       0.17 10.95
#> advance        0.47 12.78