sim.VSS {psych}	R Documentation

create VSS like data

Description

Simulation is one of most useful techniques in statistics and psychometrics. Here we simulate a correlation matrix with a simple structure composed of a specified number of factors. Each item is assumed to have complexity one. See circ.sim and item.sim for alternative simulations.

Usage

sim.VSS(ncases=1000, nvariables=16, nfactors=4, meanloading=.5,dichot=FALSE,cut=0)

Arguments

ncases
number of simulated subjects

nvariables
Number of variables

nfactors
Number of factors to generate

meanloading
with a mean loading

dichot
dichot=FALSE give continuous variables, dichot=TRUE gives dichotomous variables

cut
if dichotomous = TRUE, then items with values > cut are assigned 1, otherwise 0.

Value

a ncases x nvariables matrix

Author(s)

William Revelle

See Also

VSS, ICLUST

Examples

## Not run: 
simulated <- sim.VSS(1000,20,4,.6)
vss <- VSS(simulated,rotate="varimax")
VSS.plot(vss)

## End(Not run)

[Package psych version 1.9.12 Index]