Find the Squared Multiple Correlation (SMC) of each variable with the remaining variables in a matrix

The squared multiple correlation of a variable with the remaining variables in a matrix is sometimes used as initial estimates of the communality of a variable.

SMCs are also used when estimating reliability using Guttman's lambda 6 guttman coefficient.

The SMC is just 1 - 1/diag(R.inv) where R.inv is the inverse of R.

smc(R,covar=FALSE)

Arguments

R
A correlation matrix or a dataframe. In the latter case, correlations are found.

covar
if covar = TRUE and R is either a covariance matrix or data frame, then return the smc * variance for each item

Value

a vector of squared multiple correlations. Or, if covar=TRUE, a vector of squared multiple correlations * the item variances

If the matrix is not invertible, then a vector of 1s is returned.

In the case of correlation or covariance matrices with some NAs, those variables with NAs are dropped and the SMC for the remaining variables are found. The missing SMCs are then estimated by finding the maximum correlation for that column (with a warning).

See also

mat.regress, fa

Examples

R <- make.hierarchical()
round(smc(R),2)

#>   V1   V2   V3   V4   V5   V6   V7   V8   V9 
#> 0.44 0.37 0.28 0.30 0.24 0.17 0.18 0.14 0.09