tr {psych}	R Documentation

Find the trace of a square matrix

Description

Hardly worth coding, if it didn't appear in so many formulae in psychometrics, the trace of a (square) matrix is just the sum of the diagonal elements.

Usage

tr(m)

Arguments

m
A square matrix

Details

The tr function is used in various matrix operations and is the sum of the diagonal elements of a matrix.

Value

The sum of the diagonal elements of a square matrix.
i.e. tr(m) <- sum(diag(m)).

Examples

 m <- matrix(1:16,ncol=4)
 m
 tr(m)
 

[Package psych version 1.9.12 Index]